% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{add_axis_variance}
\alias{add_axis_variance}
\title{Add component variance to the axis labels}
\usage{
add_axis_variance(dim_object, dim_plot, x, y)
}
\arguments{
\item{dim_object}{A PCA object containing the PCA variance. Created by \code{hd_run_pca()}.}

\item{dim_plot}{A ggplot object with the data points on the 2D plane.}

\item{x}{The name of the column in \code{dim_res} that contains the x-axis values.}

\item{y}{The name of the column in \code{dim_res} that contains the y-axis values.}
}
\value{
A ggplot object
}
\description{
\code{add_axis_variance()} adds the explained variance of the components to the axis labels.
}
\keyword{internal}
