% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{apply_palette}
\alias{apply_palette}
\title{Apply palette to plot}
\usage{
apply_palette(plot, palette, type = "color")
}
\arguments{
\item{plot}{The plot to apply the palette.}

\item{palette}{The color palette to apply. It can be either a character with the name of the
palette from the HPA or a custom palette (for example \code{c("M" = "red", "F" = "blue")}).}

\item{type}{The type of palette to apply. Default is "color". Other option is "fill".}
}
\value{
The plot with the selected palette.
}
\description{
\code{apply_palette} applies the color palette to the plot. It checks if the palette is a valid
palette from the Human Protein Atlas (HPA) or a custom palette.
}
\keyword{internal}
