% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_background_lists.R
\name{background_lists}
\alias{background_lists}
\title{Background lists for Over Representation Analysis (ORA)}
\usage{
background_lists()
}
\value{
A list of background gene sets for Over Representation Analysis (ORA).
}
\description{
\code{background_list()} returns a list of background gene sets for Over Representation Analysis (ORA).
}
\keyword{internal}
