% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{balance_groups}
\alias{balance_groups}
\title{Balance groups}
\usage{
balance_groups(dat, variable, case = 1, seed = 123)
}
\arguments{
\item{dat}{A dataset containing the case and control groups.}

\item{variable}{The name of the column containing the case and control groups. Default is "Disease".}

\item{case}{The case class. Default is 1.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A balanced dataset.
}
\description{
\code{balance_groups()} balances the groups based on the case variable provided.
It downsamples the control group to match the number of samples in the case group.
}
\keyword{internal}
