% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_na_percentage}
\alias{calc_na_percentage}
\title{Calculate the percentage of NAs in each column}
\usage{
calc_na_percentage(df)
}
\arguments{
\item{df}{(tibble). The input dataframe}
}
\value{
na_percentage (tibble). A tibble with the column names and the percentage of NAs in each column
}
\description{
The function calculates the percentage of NAs in each column of the input dataframe.
It filters out the columns with 0\% missing data and returns the rest in descending order.
}
\examples{
na_percentages <- calc_na_percentage(example_metadata)
print(na_percentages)
}
