% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{calculate_na_percentage}
\alias{calculate_na_percentage}
\title{Calculate NA percentage}
\usage{
calculate_na_percentage(dat)
}
\arguments{
\item{dat}{A dataset.}
}
\value{
A tibble containing the column names and their respective NA percentages.
}
\description{
\code{calculate_na_percentage()} calculates the percentage of missing values in each column of a dataset.
}
\keyword{internal}
