% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{check_col_types}
\alias{check_col_types}
\title{Check the column types of the dataframe}
\usage{
check_col_types(df)
}
\arguments{
\item{df}{(tibble). The input dataframe.}
}
\value{
class_summary (table). A table with the counts of each class in the dataframe.
}
\description{
The function checks the column types of the input dataframe and returns the counts of each class.
}
\keyword{internal}
