% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{check_data}
\alias{check_data}
\title{Check data structure of object}
\usage{
check_data(dat, variable = "Disease")
}
\arguments{
\item{dat}{An \code{hd_model} object or a list containing the train and test data.}

\item{variable}{The name of the column containing the case and control groups. Default is "Disease".}
}
\value{
A model object containing the train and test data.
}
\description{
\code{check_data()} checks the structure of the object provided. It checks if the
object is an \code{hd_model} object or a list containing the train and test data.
It also checks if the variable is present in the train and test data.
}
\keyword{internal}
