% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_numeric_columns}
\alias{check_numeric_columns}
\title{Check for non-numeric columns}
\usage{
check_numeric_columns(dat)
}
\arguments{
\item{dat}{The dataframe to check.}
}
\value{
A warning with the names of non-numeric columns if any.
}
\description{
\code{check_numeric_columns()} checks if all columns except the first (Sample ID)
in a dataframe are numeric.
}
\keyword{internal}
