% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{clean_data}
\alias{clean_data}
\title{Preprocess data}
\usage{
clean_data(
  df_in,
  keep_cols = c("DAid", "Assay", "NPX"),
  filter_plates = NULL,
  filter_assays = NULL,
  filter_assay_warning = FALSE,
  remove_na_cols = c("DAid", "NPX"),
  replace_w_na = c(0, "0", "", "Unknown", "unknown", "none", NA, "na")
)
}
\arguments{
\item{df_in}{The input dataframe.}

\item{keep_cols}{The columns to keep in the output dataframe.}

\item{filter_plates}{The plates to exclude.}

\item{filter_assays}{The assays to filter out.}

\item{filter_assay_warning}{If TRUE, only the rows with Assay_Warning == "PASS" are kept. Default is FALSE.}

\item{remove_na_cols}{The columns to check for NAs and remove respective rows. Defaults is c("DAid", "NPX").}

\item{replace_w_na}{The values to replace with NA. Default is c(0, "0", "", "Unknown", "unknown", "none", NA, "na").}
}
\value{
The preprocessed dataframe.
}
\description{
\code{clean_data()} preprocesses the data by filtering out rows based on the specified criteria.
\itemize{
\item It keeps only the specified columns.
\item It keeps only the data of the specified plates and assays.
\item It can remove all samples with Assay_Warning != "PASS".
\item It removes rows with NAs in the DAid and NPX columns.
\item It replaces the specified values with NA.
}
}
\examples{
# Unprocessed data
example_data

# Preprocessed data
clean_data(example_data, filter_plates = c("Plate1", "Plate2"), filter_assay_warning = TRUE)
}
