% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_data}
\alias{cluster_data}
\title{Cluster data}
\usage{
cluster_data(
  df,
  distance_method = "euclidean",
  clustering_method = "ward.D2",
  cluster_rows = T,
  cluster_cols = T,
  wide = T
)
}
\arguments{
\item{df}{(tibble). A data frame with the data to cluster.}

\item{distance_method}{(character). The distance method to use. Default is "euclidean".}

\item{clustering_method}{(character). The clustering method to use. Default is "ward.D2".}

\item{cluster_rows}{(logical). Whether to cluster rows. Default is TRUE.}

\item{cluster_cols}{(logical). Whether to cluster columns. Default is TRUE.}

\item{wide}{(logical). Whether the data is wide or long. Default is TRUE.}
}
\value{
clustering_results (data.frame). A data frame with the clustered data.
}
\description{
This function clusters data and returns a data frame with the clustered data.
This data can be used to plot a heatmap with ggplot2.
}
\examples{
clean_df <- example_data |> dplyr::select(DAid, Assay, NPX)
clustered_df <- cluster_data(clean_df, wide = FALSE)
}
