% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_help}
\alias{cluster_help}
\title{Cluster data helper function}
\usage{
cluster_help(x, k = NULL, distance, method)
}
\arguments{
\item{x}{A numeric matrix or data frame.}

\item{k}{The number of clusters to return. If NULL, the function simply returns the hierarchical clustering object.}

\item{distance}{The distance method to use.}

\item{method}{The clustering method to use.}
}
\value{
A list with the cluster assignment.
}
\description{
\code{cluster_help()} is a helper function that acts as a wrapper for the \code{hclust()} function.
It takes a dataset and returns the clustering of the rows or columns.
}
\keyword{internal}
