% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\title{Correlate data}
\usage{
correlate(x, y = NULL, use = "pairwise.complete.obs", method = "pearson")
}
\arguments{
\item{x}{A numeric vector, matrix or tibble.}

\item{y}{A numeric vector, matrix or tibble with compatible dimensions with \code{x}. Default is NULL.}

\item{use}{A character string. The method to use for computing correlations. Default is "pairwise.complete.obs".}

\item{method}{A character string. The correlation method to use. Default is "pearson".}
}
\value{
A matrix of protein-protein correlations.
}
\description{
\code{correlate()} calculates the correlation matrix of the input dataset.
}
\keyword{internal}
