% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{create_corr_heatmap}
\alias{create_corr_heatmap}
\title{Create a correlation heatmap and report the protein-protein correlations above a threshold}
\usage{
create_corr_heatmap(df, threshold)
}
\arguments{
\item{df}{(tibble). The input dataframe}

\item{threshold}{(numeric). The reporting protein-protein correlation threshold. Default is 0.8}
}
\value{
A list containing the following elements:
\itemize{
\item cor_matrix (matrix). A matrix of protein-protein correlations
\item cor_results (tibble). A tibble with the filtered protein pairs and their correlation values
\item p (plot). A heatmap of protein-protein correlations
}
}
\description{
The function calculates the correlation matrix of the input dataframe.
It filters out the protein pairs with correlation values above or below the threshold.
It then creates a heatmap of the correlation matrix.
}
\examples{
df <- example_data |>
  dplyr::select(DAid, Assay, NPX) |>
  tidyr::pivot_wider(names_from = "Assay", values_from = "NPX")
cor_results <- create_corr_heatmap(df, threshold = 0.7)
}
