% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{create_dendrograms}
\alias{create_dendrograms}
\title{Create dendrograms (under development - it should be inside a ggplot heatmap function)}
\usage{
create_dendrograms(hc_rows, hc_cols)
}
\arguments{
\item{hc_rows}{(hclust). The hierarchical clustering object for rows.}

\item{hc_cols}{(hclust). The hierarchical clustering object for columns.}
}
\value{
(list). A list with the following elements:
\itemize{
\item dend_rows (ggplot). A ggplot object with the dendrogram for rows.
\item dend_cols (ggplot). A ggplot object with the dendrogram for columns.
}
}
\description{
This function creates dendrograms for the rows and columns of the clustered data.
}
\keyword{internal}
