% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_directory}
\alias{create_directory}
\title{Create directory w/o system date}
\usage{
create_directory(dir_name, date = FALSE)
}
\arguments{
\item{dir_name}{The name of the directory to create}

\item{date}{Logical. If TRUE, the current date and time will be appended to the directory name}
}
\description{
The function creates a directory with the specified name. If the directory already exists, a message is printed.
}
\examples{
create_directory("my_directory", date = FALSE)
create_directory("my_directory/inner_dir", date = TRUE)
}
