% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{create_volcano}
\alias{create_volcano}
\title{Create volcano plots}
\usage{
create_volcano(disease, de_result)
}
\arguments{
\item{disease}{(character). The disease of interest}

\item{de_result}{(tibble). The differential expression results}
}
\value{
p (plot). A ggplot object with the volcano plot
}
\description{
This function creates volcano plots for the differential expression results.
}
\examples{
test_data <- tibble::tibble(
  Assay = c("AARSD1", "ABL1", "ACAA1", "ACAN", "ACE2"),
  logFC = c(0.5, -1.2, 0.3, -0.8, 0.6),
  adj.P.Val = c(0.01, 0.03, 0.2, 0.05, 0.001),
  sig = c("significant up",
          "significant down",
          "not significant",
          "significant down",
          "significant up")
)
create_volcano("AML", test_data)
}
