% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{do_elnet}
\alias{do_elnet}
\title{Elastic net classification model pipeline}
\usage{
do_elnet(
  olink_data,
  metadata,
  case,
  wide = TRUE,
  balance_groups = TRUE,
  only_female = NULL,
  only_male = NULL,
  exclude_cols = "Sex",
  ratio = 0.75,
  type = "lasso",
  cv_sets = 5,
  grid_size = 10,
  ncores = 4,
  hypopt_vis = TRUE,
  palette = NULL,
  vline = TRUE,
  subtitle = c("accuracy", "sensitivity", "specificity", "auc", "features",
    "top-features", "mixture"),
  nfeatures = 9,
  points = T,
  seed = 123
)
}
\arguments{
\item{olink_data}{Olink data.}

\item{metadata}{Metadata.}

\item{case}{Disease to predict.}

\item{wide}{Whether the data is wide format. Default is TRUE.}

\item{balance_groups}{Whether to balance the groups. Default is TRUE.}

\item{only_female}{Vector of diseases that are female specific. Default is NULL.}

\item{only_male}{Vector of diseases that are male specific. Default is NULL.}

\item{exclude_cols}{Columns to exclude from the data before the model is tuned. Default is "Sex".}

\item{ratio}{Ratio of training data to test data. Default is 0.75.}

\item{type}{Type of regularization. Default is "lasso". Other options are "ridge" and "elnet".}

\item{cv_sets}{Number of cross-validation sets. Default is 5.}

\item{grid_size}{Size of the hyperparameter optimization grid. Default is 10.}

\item{ncores}{Number of cores to use for parallel processing. Default is 4.}

\item{hypopt_vis}{Whether to visualize hyperparameter optimization results. Default is TRUE.}

\item{palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is NULL.}

\item{vline}{Whether to add a vertical line at 50\% importance. Default is TRUE.}

\item{subtitle}{Vector of subtitle elements to include in the plot. Default is a list with all.}

\item{nfeatures}{Number of top features to include in the boxplot. Default is 9.}

\item{points}{Whether to add points to the boxplot. Default is TRUE.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A list with results for each disease. The list contains:
\itemize{
\item hypopt_res: Hyperparameter optimization results.
\item finalfit_res: Final model fitting results.
\item testfit_res: Test model fitting results.
\item var_imp_res: Variable importance results.
}
}
\description{
\code{do_elnet()} runs the elastic net classification model pipeline. It splits the
data into training and test sets, creates class-balanced case-control groups,
and fits the model. It also performs hyperparameter optimization, fits the best
model, tests it, and plots useful the feature variable importance.
}
\examples{
do_elnet(example_data,
         example_metadata,
         "AML",
         balance_groups = TRUE,
         wide = FALSE,
         type = "elnet",
         palette = "cancers12",
         cv_sets = 5,
         grid_size = 20,
         ncores = 1)
}
