% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{do_elnet_multi}
\alias{do_elnet_multi}
\title{Elastic net multiclassification model pipeline}
\usage{
do_elnet_multi(
  olink_data,
  metadata,
  wide = TRUE,
  exclude_cols = "Sex",
  ratio = 0.75,
  type = "lasso",
  cv_sets = 5,
  grid_size = 10,
  ncores = 4,
  hypopt_vis = TRUE,
  palette = NULL,
  seed = 123
)
}
\arguments{
\item{olink_data}{Olink data.}

\item{metadata}{Metadata.}

\item{wide}{Whether the data is wide format. Default is TRUE.}

\item{exclude_cols}{Columns to exclude from the data before the model is tuned.}

\item{ratio}{Ratio of training data to test data. Default is 0.75.}

\item{type}{Type of regularization. Default is "lasso". Other options are "ridge" and "elnet".}

\item{cv_sets}{Number of cross-validation sets. Default is 5.}

\item{grid_size}{Size of the hyperparameter optimization grid. Default is 10.}

\item{ncores}{Number of cores to use for parallel processing. Default is 4.}

\item{hypopt_vis}{Whether to visualize hyperparameter optimization results. Default is TRUE.}

\item{palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is NULL.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A list with the following elements:
\itemize{
\item hypopt_res: Hyperparameter optimization results.
\item finalfit_res: Final model fitting results.
\item roc_curve: ROC curve plot.
\item auc: AUC values for each class.
\item auc_barplot: AUC barplot.
}
}
\description{
\code{do_elnet_multi()} runs the elastic net multiclassification model pipeline. It splits the
data into training and test sets, creates class-balanced case-control groups,
and fits the model. It performs hyperparameter optimization and fits the best
model. It also plots the ROC curve and the AUC barplot for each class.
}
\details{
If the data contain missing values, KNN imputation will be applied.
}
\examples{
do_elnet_multi(example_data,
               example_metadata,
               wide = FALSE,
               palette = "cancers12",
               cv_sets = 5,
               grid_size = 5,
               ncores = 1)
}
