% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{do_gsea}
\alias{do_gsea}
\title{Perform gene set enrichment analysis}
\usage{
do_gsea(de_results, database = c("KEGG", "GO"))
}
\arguments{
\item{de_results}{(list). A list containing the results of the differential expression analysis.}

\item{database}{(character). The database to perform the GSEA. It can be either "KEGG" or "GO".}
}
\value{
enrichment_results (list). A list containing the results of the GSEA.
}
\description{
This function performs gene set enrichment analysis (GSEA) using the clusterProfiler package.
}
\examples{
#enrichment_results <- do_gsea(de_results, database = "KEGG")
}
