% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{do_limma_continuous}
\alias{do_limma_continuous}
\title{Run differential expression analysis with limma for continuous variable}
\usage{
do_limma_continuous(
  olink_data,
  metadata,
  variable,
  correct = c("Sex"),
  correct_type = c("factor"),
  wide = TRUE,
  volcano = TRUE,
  pval_lim = 0.05,
  logfc_lim = 0,
  top_up_prot = 40,
  top_down_prot = 10,
  palette = "diff_exp",
  report_nproteins = TRUE,
  subtitle = NULL,
  save = FALSE
)
}
\arguments{
\item{olink_data}{A tibble with the Olink data in wide format.}

\item{metadata}{A tibble with the metadata.}

\item{variable}{The variable of interest.}

\item{correct}{The variables to correct the results with. Default is c("Sex").}

\item{correct_type}{The type of the variables to correct the results with. Default is c("factor").}

\item{wide}{If the data is in wide format. Default is TRUE.}

\item{volcano}{Generate volcano plots. Default is TRUE.}

\item{pval_lim}{The p-value limit for significance. Default is 0.05.}

\item{logfc_lim}{The logFC limit for significance. Default is 0.}

\item{top_up_prot}{The number of top up regulated proteins to label on the plot. Default is 40.}

\item{top_down_prot}{The number of top down regulated proteins to label on the plot. Default is 10.}

\item{palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is "diff_exp".}

\item{report_nproteins}{If the number of significant proteins should be reported in the subtitle. Default is TRUE.}

\item{subtitle}{The subtitle of the plot or NULL for no subtitle.}

\item{save}{Save the volcano plots. Default is FALSE.}
}
\value{
A list with the differential expression results and volcano plots.
\itemize{
\item de_results: A list with the differential expression results.
\item volcano_plot: A list with the volcano plots.
}
}
\description{
This function runs differential expression analysis using limma for a continuous variable.
It can generate and save volcano plots.
}
\details{
It will filter out rows with NA values in any of the columns that are used for
correction, either the \code{variable} or in \code{correct}.
}
\examples{
do_limma_continuous(example_data, example_metadata, "Age", wide = FALSE)
}
