% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{do_limma_continuous_de}
\alias{do_limma_continuous_de}
\title{Differential expression analysis with limma for continuous variable}
\usage{
do_limma_continuous_de(
  join_data,
  variable,
  correct = c("Sex"),
  correct_type = c("factor"),
  pval_lim = 0.05,
  logfc_lim = 0
)
}
\arguments{
\item{join_data}{A tibble with the Olink data in wide format joined with metadata.}

\item{variable}{The variable of interest.}

\item{correct}{The variables to correct the results with. Default is c("Sex").}

\item{correct_type}{The type of the variables to correct the results with. Default is c("factor").}

\item{pval_lim}{The p-value limit for significance. Default is 0.05.}

\item{logfc_lim}{The logFC limit for significance. Default is 0.}
}
\value{
A tibble with the differential expression results.
}
\description{
This function performs differential expression analysis using limma for a continuous variable.
The output dataframe includes the logFC, the p-values, as well as the adjusted p-values with FDR.
The function removes the NAs from the columns that are used to correct for.
}
\keyword{internal}
