% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{do_limma_de}
\alias{do_limma_de}
\title{Differential expression analysis with limma}
\usage{
do_limma_de(
  join_data,
  disease,
  correct = c("Sex", "Age", "BMI"),
  correct_type = c("factor", "numeric", "numeric"),
  only_female = NULL,
  only_male = NULL,
  pval_lim = 0.05,
  logfc_lim = 0
)
}
\arguments{
\item{join_data}{(tibble). A tibble with the Olink data in wide format joined with metadata.}

\item{disease}{(character). The disease of interest.}

\item{correct}{(character or vector). The variables to correct the results with. Default c("Sex", "Age", "BMI").}

\item{correct_type}{(character or vector). The type of the variables to correct the results with. Default c("factor", "numeric", "numeric").}

\item{only_female}{(character or vector). The female specific diseases. Default is NULL.}

\item{only_male}{(character or vector). The male specific diseases. Default is NULL.}

\item{pval_lim}{(numeric). The p-value limit for significance. Default is 0.05.}

\item{logfc_lim}{(numeric). The logFC limit for significance. Default is 0.}
}
\value{
de_res (tibble). A tibble with the differential expression results.
}
\description{
This function performs differential expression analysis using limma.
It can correct the results with sex, age, and BMI.
The output dataframe includes the logFC, the p-values, as well as the adjusted p-values with FDR.
}
\keyword{internal}
