% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{do_lreg}
\alias{do_lreg}
\title{Fit logistic regression model for single predictors}
\usage{
do_lreg(
  olink_data,
  metadata,
  variable = "Disease",
  case,
  control,
  wide = TRUE,
  strata = TRUE,
  balance_groups = TRUE,
  only_female = NULL,
  only_male = NULL,
  exclude_cols = "Sex",
  ratio = 0.75,
  cor_threshold = 0.9,
  normalize = TRUE,
  cv_sets = 5,
  ncores = 4,
  palette = NULL,
  points = TRUE,
  boxplot_xaxis_names = FALSE,
  seed = 123
)
}
\arguments{
\item{olink_data}{Olink data.}

\item{metadata}{Metadata.}

\item{variable}{The variable to predict. Default is "Disease".}

\item{case}{The case group.}

\item{control}{The control groups.}

\item{wide}{Whether the data is wide format. Default is TRUE.}

\item{strata}{Whether to stratify the data. Default is TRUE.}

\item{balance_groups}{Whether to balance the groups. Default is TRUE.}

\item{only_female}{Vector of diseases.}

\item{only_male}{Vector of diseases.}

\item{exclude_cols}{Columns to exclude from the data before the model is tuned.}

\item{ratio}{Ratio of training data to test data. Default is 0.75.}

\item{cor_threshold}{Threshold of absolute correlation values. This will be used to remove the minimum number of features so that all their resulting absolute correlations are less than this value.}

\item{normalize}{Whether to normalize numeric data to have a standard deviation of one and a mean of zero. Default is TRUE.}

\item{cv_sets}{Number of cross-validation sets. Default is 5.}

\item{ncores}{Number of cores to use for parallel processing. Default is 4.}

\item{palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is NULL.}

\item{points}{Whether to add points to the boxplot. Default is TRUE.}

\item{boxplot_xaxis_names}{Whether to add x-axis names to the boxplot. Default is FALSE.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A list with two elements:
\itemize{
\item fit_res: A list with 4 elements:
\item lreg_wf: Workflow object.
\item train_set: Training set.
\item test_set: Testing set.
\item final: Fitted model.
\item metrics: A list with the model metrics:
\item accuracy: Accuracy of the model.
\item sensitivity: Sensitivity of the model.
\item specificity: Specificity of the model.
\item auc: AUC of the model.
\item conf_matrix: Confusion matrix of the model.
\item roc_curve: ROC curve of the model.
}
}
\description{
\code{lreg_fit()} fits a logistic regression model for a single predictor and calculates
the ROC AUC, accuracy, sensitivity, and specificity. It also performs cross-validation
and plots the ROC curve.
}
\examples{
# Data with single predictor
test_data <- example_data |> dplyr::filter(Assay == "ADA")

# Run model
do_lreg(test_data,
        example_metadata,
        variable = "Disease",
        case = "AML",
        control = "CLL",
        wide = FALSE,
        ncores = 1,
        palette = "cancers12")
}
