% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{do_pca}
\alias{do_pca}
\title{Run PCA analysis}
\usage{
do_pca(
  olink_data,
  metadata = NULL,
  color = "Disease",
  palette = NULL,
  wide = TRUE,
  impute = TRUE,
  plots = TRUE,
  pcs = 4,
  nproteins = 8,
  save = FALSE
)
}
\arguments{
\item{olink_data}{A tibble with the data to be used in the PCA analysis.}

\item{metadata}{A tibble with metadata information to be used in the PCA plots. Default is NULL.}

\item{color}{The name of the column in \code{metadata} that contains the variable.
to be used to plot the points color. Default is "Disease".}

\item{palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}.}

\item{wide}{If TRUE, the data is assumed to be in wide format. Default is TRUE.}

\item{impute}{If TRUE, the data is imputed before the PCA analysis. Default is TRUE.}

\item{plots}{If TRUE, the function creates plots of the PCA results. Default is TRUE.}

\item{pcs}{The number of PCs to be plotted. Default is 4.}

\item{nproteins}{The number of proteins to be plotted. Default is 8.}

\item{save}{If TRUE, the plots are saved in the results directory. Default is FALSE.}
}
\value{
A list with the PCA results and, if requested, the PCA plots.
\itemize{
\item pca_res: A tibble with the PCA results.
\item loadings: A tibble with the PCA loadings.
\item pca_plot: A ggplot object with the data points on the 1st and 2nd PCs plane.
\item loadings_plot: A PCA loadings ggplot object.
}
}
\description{
\code{do_pca()} runs a PCA analysis on the provided data. The function can visualize
the sample points on the first and second PC plane as well as the PCA loadings.
It can also save the plots in the results directory.
}
\examples{
test_data <- example_data |> dplyr::select(DAid, Assay, NPX)

# Run PCA analysis
do_pca(test_data,
       metadata = example_metadata,
       wide = FALSE,
       color = "Disease",
       palette = "cancers12")
}
