% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{do_rf}
\alias{do_rf}
\title{Random forest classification model pipeline}
\usage{
do_rf(
  olink_data,
  metadata,
  wide = F,
  only_female = NULL,
  only_male = NULL,
  exclude_cols = "Sex",
  ratio = 0.75,
  cv_sets = 5,
  grid_size = 10,
  ncores = 4,
  hypopt_vis = TRUE,
  palette = NULL,
  vline = T,
  subtitle = c("accuracy", "sensitivity", "specificity", "auc", "features",
    "top-features"),
  nfeatures = 9,
  points = T,
  seed = 123
)
}
\arguments{
\item{olink_data}{(tibble). Olink data.}

\item{metadata}{(tibble). Metadata.}

\item{wide}{(logical). Whether the data is wide format. Default is FALSE.}

\item{only_female}{(vector). Vector of diseases that are female specific. Default is NULL.}

\item{only_male}{(vector). Vector of diseases that are male specific. Default is NULL.}

\item{exclude_cols}{(vector). Columns to exclude from the model. Default is "Sex".}

\item{ratio}{(numeric). Ratio of training data to test data. Default is 0.75.}

\item{cv_sets}{(numeric). Number of cross-validation sets. Default is 5.}

\item{grid_size}{(numeric). Size of the grid for hyperparameter optimization. Default is 10.}

\item{ncores}{(numeric). Number of cores to use for parallel processing. Default is 4.}

\item{hypopt_vis}{(logical). Whether to visualize hyperparameter optimization results. Default is TRUE.}

\item{palette}{(character or vector). The color palette for the plot. If it is a character, it should be one of the palettes from get_hpa_palettes(). Default is NULL.}

\item{vline}{(logical). Whether to add a vertical line at 50\% importance. Default is TRUE.}

\item{subtitle}{(vector). Vector of subtitles to include in the plot. Default is a list with all except mixture.}

\item{nfeatures}{(numeric). Number of top features to include in the boxplot. Default is 9.}

\item{points}{(logical). Whether to add points to the boxplot. Default is TRUE.}

\item{seed}{(numeric). Seed for reproducibility. Default is 123.}
}
\value{
A list with results for each disease. The list contains:
\itemize{
\item hypopt_res (list). Hyperparameter optimization results.
\item finalfit_res (list). Final model fitting results.
\item testfit_res (list). Test model fitting results.
\item var_imp_res (list). Variable importance results.
}
}
\description{
This function runs the random forest classification model pipeline.
It splits the data into training and test sets, creates class-balanced groups, and fits the model.
It also performs hyperparameter optimization, fits the final model, tests it, and plots useful visualizations.
}
\examples{
unique_samples <- unique(example_data$Sample)
filtered_data <- example_data |>
 dplyr::filter(Sample \%in\% unique_samples[1:148])

res <- do_rf(filtered_data,
             example_metadata,
             palette = "cancers12",
             cv_sets = 2,
             grid_size = 1,
             ncores = 1)
}
