% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{do_rf_multi}
\alias{do_rf_multi}
\title{Random forest multiclassification model pipeline}
\usage{
do_rf_multi(
  olink_data,
  metadata,
  variable = "Disease",
  wide = TRUE,
  strata = TRUE,
  exclude_cols = "Sex",
  ratio = 0.75,
  cor_threshold = 0.9,
  normalize = TRUE,
  cv_sets = 5,
  grid_size = 10,
  ncores = 4,
  hypopt_vis = TRUE,
  palette = NULL,
  seed = 123
)
}
\arguments{
\item{olink_data}{Olink data.}

\item{metadata}{Metadata.}

\item{variable}{The variable to predict. Default is "Disease".}

\item{wide}{Whether the data is wide format. Default is TRUE.}

\item{strata}{Whether to stratify the data. Default is TRUE.}

\item{exclude_cols}{Columns to exclude from the data before the model is tuned.}

\item{ratio}{Ratio of training data to test data. Default is 0.75.}

\item{cor_threshold}{Threshold of absolute correlation values. This will be used to remove the minimum number of features so that all their resulting absolute correlations are less than this value.}

\item{normalize}{Whether to normalize numeric data to have a standard deviation of one and a mean of zero. Default is TRUE.}

\item{cv_sets}{Number of cross-validation sets. Default is 5.}

\item{grid_size}{Size of the hyperparameter optimization grid. Default is 10.}

\item{ncores}{Number of cores to use for parallel processing. Default is 4.}

\item{hypopt_vis}{Whether to visualize hyperparameter optimization results. Default is TRUE.}

\item{palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is NULL.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A list with the following elements:
\itemize{
\item hypopt_res: Hyperparameter optimization results.
\item finalfit_res: Final model fitting results.
\item roc_curve: ROC curve plot.
\item auc: AUC values for each class.
\item auc_barplot: AUC barplot.
}
}
\description{
\code{do_rf_multi()} runs the random forest multiclassification model pipeline. It splits the
data into training and test sets, creates class-balanced case-control groups,
and fits the model. It performs hyperparameter optimization and fits the best
model. It also plots the ROC curve and the AUC barplot for each class.
}
\details{
If the data contain missing values, KNN imputation will be applied.
If no check for feature correlation is preferred, set \code{cor_threshold} to 1.
It will filter out rows that contain NAs in Disease.
}
\examples{
do_rf_multi(example_data,
            example_metadata,
            wide = FALSE,
            palette = "cancers12",
            cv_sets = 5,
            grid_size = 5,
            ncores = 1)
}
