% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{do_rreg}
\alias{do_rreg}
\title{Regularized classification model pipeline}
\usage{
do_rreg(
  olink_data,
  metadata,
  variable = "Disease",
  case,
  control,
  wide = TRUE,
  strata = TRUE,
  balance_groups = TRUE,
  only_female = NULL,
  only_male = NULL,
  exclude_cols = "Sex",
  ratio = 0.75,
  type = "lasso",
  cor_threshold = 0.9,
  cv_sets = 5,
  grid_size = 10,
  ncores = 4,
  hypopt_vis = TRUE,
  palette = NULL,
  vline = TRUE,
  subtitle = c("accuracy", "sensitivity", "specificity", "auc", "features",
    "top-features", "mixture"),
  varimp_yaxis_names = FALSE,
  nfeatures = 9,
  points = TRUE,
  boxplot_xaxis_names = FALSE,
  seed = 123
)
}
\arguments{
\item{olink_data}{Olink data.}

\item{metadata}{Metadata.}

\item{variable}{The variable to predict. Default is "Disease".}

\item{case}{The case group.}

\item{control}{The control groups.}

\item{wide}{Whether the data is wide format. Default is TRUE.}

\item{strata}{Whether to stratify the data. Default is TRUE.}

\item{balance_groups}{Whether to balance the groups. Default is TRUE.}

\item{only_female}{Vector of diseases that are female specific. Default is NULL.}

\item{only_male}{Vector of diseases that are male specific. Default is NULL.}

\item{exclude_cols}{Columns to exclude from the data before the model is tuned. Default is "Sex".}

\item{ratio}{Ratio of training data to test data. Default is 0.75.}

\item{type}{Type of regularization. Default is "lasso". Other options are "ridge" and "elnet".}

\item{cor_threshold}{Threshold of absolute correlation values. This will be used to remove the minimum number of features so that all their resulting absolute correlations are less than this value.}

\item{cv_sets}{Number of cross-validation sets. Default is 5.}

\item{grid_size}{Size of the hyperparameter optimization grid. Default is 10.}

\item{ncores}{Number of cores to use for parallel processing. Default is 4.}

\item{hypopt_vis}{Whether to visualize hyperparameter optimization results. Default is TRUE.}

\item{palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is NULL.}

\item{vline}{Whether to add a vertical line at 50\% importance. Default is TRUE.}

\item{subtitle}{Vector of subtitle elements to include in the plot. Default is a list with all.}

\item{varimp_yaxis_names}{Whether to add y-axis names to the variable importance plot. Default is FALSE.}

\item{nfeatures}{Number of top features to include in the boxplot. Default is 9.}

\item{points}{Whether to add points to the boxplot. Default is TRUE.}

\item{boxplot_xaxis_names}{Whether to add x-axis names to the boxplot. Default is FALSE.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A list with results for each disease. The list contains:
\itemize{
\item hypopt_res: Hyperparameter optimization results.
\item finalfit_res: Final model fitting results.
\item testfit_res: Test model fitting results.
\item var_imp_res: Variable importance results.
}
}
\description{
\code{do_rreg()} runs the regularized classification model pipeline. It splits the
data into training and test sets, creates class-balanced case-control groups,
and fits the model. It also performs hyperparameter optimization, fits the best
model, tests it, and plots useful the feature variable importance.
}
\details{
If the data contain missing values, KNN imputation will be applied.
If no check for feature correlation is preferred, set \code{cor_threshold} to 1.
}
\examples{
do_rreg(example_data,
        example_metadata,
        case = "AML",
        control = c("CLL", "MYEL"),
        balance_groups = TRUE,
        wide = FALSE,
        type = "elnet",
        palette = "cancers12",
        cv_sets = 5,
        grid_size = 10,
        ncores = 1)
}
