% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{do_ttest}
\alias{do_ttest}
\title{Run differential expression analysis with t-test}
\usage{
do_ttest(
  olink_data,
  metadata,
  wide = T,
  only_female = NULL,
  only_male = NULL,
  volcano = T,
  pval_lim = 0.05,
  logfc_lim = 0,
  top_up_prot = 40,
  top_down_prot = 10,
  palette = c(`not significant` = "grey", `significant down` = "blue", `significant up` =
    "red"),
  subtitle = T,
  save = F
)
}
\arguments{
\item{olink_data}{(tibble). A tibble with the Olink data in wide format.}

\item{metadata}{(tibble). A tibble with the metadata.}

\item{wide}{(logical). If the data is in wide format. Default is TRUE.}

\item{only_female}{(character or vector). The female specific diseases. Default is NULL.}

\item{only_male}{(character or vector). The male specific diseases. Default is NULL.}

\item{volcano}{(logical). Generate volcano plots. Default is TRUE.}

\item{pval_lim}{(numeric). The p-value limit for significance. Default is 0.05.}

\item{logfc_lim}{(numeric). The logFC limit for significance. Default is 0.}

\item{top_up_prot}{(numeric). The number of top up regulated proteins to label on the plot. Default is 40.}

\item{top_down_prot}{(numeric). The number of top down regulated proteins to label on the plot. Default is 10.}

\item{palette}{(vector). The color palette for the plot. Default is c("not significant" = "grey", "significant down" = "blue", "significant up" = "red").}

\item{subtitle}{(logical). If the subtitle should be displayed. Default is TRUE.}

\item{save}{(logical). Save the volcano plots. Default is FALSE.}
}
\value{
de_results (list). A list with the differential expression results and volcano plots.
\itemize{
\item de_results: A list with the differential expression results.
\item volcano_plots: A list with the volcano plots.
}
}
\description{
This function runs differential expression analysis using t-test.
It can generate and save volcano plots.
}
\examples{
test_data <- example_data |> dplyr::select(DAid, Assay, NPX)
do_ttest(test_data, example_metadata, wide = FALSE)
}
