% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{do_ttest_de}
\alias{do_ttest_de}
\title{Differential expression analysis with t-test}
\usage{
do_ttest_de(
  long_data,
  disease,
  assays,
  normality_res,
  only_female = NULL,
  only_male = NULL,
  pval_lim = 0.05,
  logfc_lim = 0
)
}
\arguments{
\item{long_data}{(tibble). A tibble with the Olink data in long format and the Sex column from metadata.}

\item{disease}{(character). The disease of interest.}

\item{assays}{(character or vector). The assays to run the differential expression analysis on.}

\item{only_female}{(character or vector). The female specific diseases. Default is NULL.}

\item{only_male}{(character or vector). The male specific diseases. Default is NULL.}

\item{pval_lim}{(numeric). The p-value limit for significance. Default is 0.05.}

\item{logfc_lim}{(numeric). The logFC limit for significance. Default is 0.}
}
\value{
de_res (tibble). A tibble with the differential expression results.
}
\description{
This function performs differential expression analysis using t-test.
}
\keyword{internal}
