% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{elnet_finalfit}
\alias{elnet_finalfit}
\title{Fit the final elastic net model}
\usage{
elnet_finalfit(train_set, tune_res, wf, seed = 123)
}
\arguments{
\item{train_set}{(tibble). Training set.}

\item{tune_res}{(tibble). Hyperparameter optimization results.}

\item{wf}{(workflow). Workflow object.}

\item{seed}{(numeric). Seed for reproducibility. Default is 123.}
}
\value{
A list with three elements:
\itemize{
\item final_elnet (parsnip model). Final elastic net model.
\item best_elnet (tibble). Best hyperparameters from hyperparameter optimization.
\item final_wf (workflow). Final workflow object.
}
}
\description{
This function fits the final elastic net model using the best hyperparameters from hyperparameter optimization.
}
\keyword{internal}
