% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{elnet_hypopt_multi}
\alias{elnet_hypopt_multi}
\title{Hyperparameter optimization for elastic net multiclassification model}
\usage{
elnet_hypopt_multi(
  train_data,
  test_data,
  variable = "Disease",
  type = "lasso",
  cor_threshold = 0.9,
  cv_sets = 5,
  grid_size = 10,
  ncores = 4,
  hypopt_vis = TRUE,
  exclude_cols = NULL,
  seed = 123
)
}
\arguments{
\item{train_data}{Training data set from \code{make_groups()}.}

\item{test_data}{Testing data set from \code{make_groups()}.}

\item{type}{Type of regularization. Default is "lasso". Other options are "ridge" and "elnet".}

\item{cor_threshold}{Threshold of absolute correlation values. This will be used to remove the minimum number of features so that all their resulting absolute correlations are less than this value.}

\item{cv_sets}{Number of cross-validation sets. Default is 5.}

\item{grid_size}{Size of the hyperparameter optimization grid. Default is 10.}

\item{ncores}{Number of cores to use for parallel processing. Default is 4.}

\item{hypopt_vis}{Whether to visualize hyperparameter optimization results. Default is TRUE.}

\item{exclude_cols}{Columns to exclude from the data before the model is tuned. Default is NULL.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A list with five elements:
\itemize{
\item elnet_tune: Hyperparameter optimization results.
\item elnet_wf: Workflow object.
\item train_set: Training set.
\item test_set: Testing set.
\item hyperopt_vis: Hyperparameter optimization plot.
}
}
\description{
\code{elnet_hypopt_multi()} tunes an elastic net model and performs hyperparameter optimization.
It uses the glmnet engine for multinomial regression and tunes either only penalty (Lasso or Ridge) or
both penalty and mixture (Elastic Regression). For the hyperparameter optimization, it
uses the \code{grid_space_filling()} function from the dials package.
}
\keyword{internal}
