% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{evaluate_multiclass_model}
\alias{evaluate_multiclass_model}
\title{Finalize and evaluate the multiclass model}
\usage{
evaluate_multiclass_model(
  dat,
  variable = "Disease",
  mixture = NULL,
  verbose = TRUE,
  seed = 123
)
}
\arguments{
\item{dat}{An \code{hd_model} object coming from a tuning function.}

\item{variable}{The variable to predict. Default is "Disease".}

\item{mixture}{The mixture parameter for the elastic net. If NULL it will be tuned. Default is NULL.}

\item{verbose}{Whether to print progress messages. Default is TRUE.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A model object containing the train and test data, the final model, the metrics, the ROC curve, and the mixture parameter.
}
\description{
\code{evaluate_multiclass_model()} finalizes the model using the best hyperparameters
and evaluates the model using the test set. It calculates the accuracy, sensitivity,
specificity, AUC, and confusion matrix. It also plots the ROC curve.
}
\keyword{internal}
