% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{example_metadata}
\alias{example_metadata}
\title{Cancer cohort metadata}
\format{
A tibble with 586 rows and 9 columns:
\describe{
\item{DAid}{The Disease Atlas sample ID}
\item{Sample}{The Sample ID}
\item{Disease}{The cancer type}
\item{Stage}{The cancer stage}
\item{Grade}{The cancer grade}
\item{Sex}{The patient sex}
\item{Age}{The patient age}
\item{BMI}{The patient BMI}
\item{Cohort}{The cohort the patient belongs to}
}
}
\source{
\url{https://github.com/buenoalvezm/Pan-cancer-profiling/blob/main/data/cancer_metadata_synthetic.rds}
}
\usage{
example_metadata
}
\description{
A data subset from synthetic cancer metadata. DAid, Age, BMI, and Cohort have
been added as extra columns. The original dataset was processed with the
\code{process_example_metadata} script.
}
\examples{
example_metadata
}
\keyword{datasets}
