% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{filter_sex_specific_disease}
\alias{filter_sex_specific_disease}
\title{Create control groups for sex specific diseases}
\usage{
filter_sex_specific_disease(
  control_data,
  disease,
  diseases,
  only_female = NULL,
  only_male = NULL
)
}
\arguments{
\item{control_data}{Control data to filter.}

\item{disease}{Disease to create control group.}

\item{diseases}{Diseases.}

\item{only_female}{Diseases that are female specific.}

\item{only_male}{Diseases that are male specific.}
}
\value{
A list with two elements:
\itemize{
\item control_data: Control data to filter.
\item diseases_subset: Filtered diseases vector.
}
}
\description{
\code{filter_sex_specific_disease()} creates control groups for sex-specific diseases
by filtering samples to include only those from the relevant sex and by
updating the diseases vector to include only the diseases that will be sampled.
}
\keyword{internal}
