% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{fix_components_names}
\alias{fix_components_names}
\title{Fix PC names}
\usage{
fix_components_names(
  pca_res,
  components,
  by_sample,
  sample_id,
  var_name,
  type = "pca"
)
}
\arguments{
\item{pca_res}{A tibble with the PCA results. Created by \code{hd_run_pca()}.}

\item{components}{The number of PCs to be calculated.}

\item{by_sample}{If TRUE, points represent samples. If FALSE, points represent features.}

\item{sample_id}{The name of the column in the data that contains the sample IDs.}

\item{var_name}{The name of the column in the data that contains the feature IDs.}

\item{type}{The type of analysis. Default is "pca".}
}
\value{
A tibble with the PCA results with fixed PC names.
}
\description{
\code{fix_pca_names()} fixes the names of the PCs in the PCA results. If number of PCs is higher than 10,
the function will remove the zeros in front of the PC names (for example PC01 -> PC1).
}
\keyword{internal}
