% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{gene_to_entrezid}
\alias{gene_to_entrezid}
\title{Convert gene names to ENTREZID}
\usage{
gene_to_entrezid(gene_list, background = NULL)
}
\arguments{
\item{gene_list}{A character vector containing the gene names.}

\item{background}{A character vector containing the background genes if any.}
}
\value{
A list containing the gene list and the background with ENTREZID.
}
\description{
\code{gene_to_entrezid()} converts gene names to ENTREZID using the \code{org.Hs.eg.db} database.
}
\keyword{internal}
