% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{get_hpa_palettes}
\alias{get_hpa_palettes}
\title{HPA color palettes}
\usage{
get_hpa_palettes()
}
\value{
A list of color palettes.
}
\description{
This function returns a list of color palettes used by the Human Protein Atlas (HPA) project.
}
\keyword{internal}
