% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{get_optimal_k}
\alias{get_optimal_k}
\title{Get optimal number of clusters}
\usage{
get_optimal_k(x, k_max = 15, b = 50, distance, method, verbose = FALSE)
}
\arguments{
\item{x}{A numeric matrix or data frame.}

\item{k_max}{The maximum number of clusters to test. Default is 15.}

\item{b}{The number of times to repeat the clustering in the gap analysis. Default is 50.}

\item{distance}{The distance method to use.}

\item{method}{The clustering method to use.}

\item{verbose}{A logical value indicating whether to print the progress of the gap analysis. Default is FALSE.}
}
\value{
A list with the gap statistic results and the optimal number of clusters.
}
\description{
\code{get_optimal_k()} takes a dataset and returns the optimal number of clusters
based on the gap statistic.
}
\keyword{internal}
