% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{hd_assess_clusters}
\alias{hd_assess_clusters}
\title{Assess clusters}
\usage{
hd_assess_clusters(
  cluster_object,
  nrep = 100,
  ji_lim = 0.5,
  nsample_lim = 10,
  seed = 123,
  verbose = FALSE
)
}
\arguments{
\item{cluster_object}{The clustering results obtained from \code{hd_cluster_samples()}.}

\item{nrep}{The number of bootstrap replicates. Default is 100.}

\item{ji_lim}{The minimum limit for the mean Jaccard index. Default is 0.5.}

\item{nsample_lim}{The minimum limit for the sample size. Default is 10.}

\item{seed}{The seed to use for reproducibility. Default is 123.}

\item{verbose}{A logical value indicating whether to print the progress of the bootstrapping. Default is FALSE.}
}
\value{
A list with the updated cluster assignment and the stability assessment.
}
\description{
\code{hd_assess_clusters()} assesses the stability of the clusters by bootstrapping the data.
It calculates the mean Jaccard index for each cluster and removes clusters with a mean
Jaccard index below 0.5 or a sample size below 10. The remaining clusters are renamed
in order of size.
}
\examples{
\dontshow{if (requireNamespace("cluster", quietly = TRUE) && requireNamespace("fpc", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Create the HDAnalyzeR object providing the data and metadata
hd_object <- hd_initialize(example_data, example_metadata)

# Clustered data
clustering <- hd_cluster_samples(hd_object, gap_b = 10)

# Assess clusters
clustering <- hd_assess_clusters(clustering, nrep = 20)

# Access the results
clustering$cluster_assessment
\dontshow{\}) # examplesIf}
}
