% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{hd_auto_umap}
\alias{hd_auto_umap}
\title{Run UMAP analysis and plot the results}
\usage{
hd_auto_umap(
  dat,
  metadata = NULL,
  by_sample = TRUE,
  plot_x = "UMAP1",
  plot_y = "UMAP2",
  plot_color = NULL,
  plot_palette = NULL
)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format and sample_id as its first column.}

\item{metadata}{A dataset containing the metadata information with the sample ID as the first column. If a HDAnalyzeR object is provided, this parameter is not needed.}

\item{by_sample}{If TRUE, points represent samples. If FALSE, points represent features. Default is TRUE.}

\item{plot_x}{The name of the column in \code{dim_res} that contains the x-axis values. Default is "PC1".}

\item{plot_y}{The name of the column in \code{dim_res} that contains the y-axis values. Default is "PC2".}

\item{plot_color}{The name of the column in \code{dim_res} that contains the variable to be used to plot the points color. Default is NULL.}

\item{plot_palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is NULL.}
}
\value{
A list with the UMAP results and UMAP plots.
}
\description{
\code{hd_auto_umap()} runs a UMAP analysis on the provided data and plots the UMAP results.
}
\examples{
# Create the HDAnalyzeR object providing the data and metadata
hd_object <- hd_initialize(example_data, example_metadata)

# Run the UMAP analysis
hd_auto_umap(hd_object, plot_color = "Disease", plot_palette = "cancers12")
}
