% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{hd_cluster}
\alias{hd_cluster}
\title{Cluster data}
\usage{
hd_cluster(
  dat,
  distance_method = "euclidean",
  clustering_method = "ward.D2",
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  normalize = TRUE
)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format and sample_id as its first column.}

\item{distance_method}{The distance method to use. Default is "euclidean".
Other options are "maximum", "manhattan", "canberra", "binary" or "minkowski".}

\item{clustering_method}{The clustering method to use. Default is "ward.D2".
Other options are "ward.D", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA),
"median" (= WPGMC) or "centroid" (= UPGMC)}

\item{cluster_rows}{Whether to cluster rows. Default is TRUE.}

\item{cluster_cols}{Whether to cluster columns. Default is TRUE.}

\item{normalize}{A logical value indicating whether to normalize the data. Z-score normalization is applied using the \code{hd_normalize()} function. Default is TRUE.}
}
\value{
A list with the dataset ordered according to the clustering of the rows and columns and the hierarchical clustering object for rows and columns.
}
\description{
\code{hd_cluster()} takes a dataset and returns the same dataset ordered
according to the hierarchical clustering of the rows and columns. This data
can be used to plot a heatmap with ggplot2 that is not having clustering functionality.
}
\details{
You can read more about the distance and clustering methods in the documentation
of the \code{dist()} and \code{hclust()} functions in the \code{stats} package.
}
\examples{
# Create the HDAnalyzeR object providing the data and metadata
hd_object <- hd_initialize(example_data, example_metadata)

# Clustered data
hd_cluster(hd_object)
}
