% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{hd_de_ttest}
\alias{hd_de_ttest}
\title{Differential expression analysis with t-test}
\usage{
hd_de_ttest(dat, metadata = NULL, variable = "Disease", case, control = NULL)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format and sample_id as its first column.}

\item{metadata}{A dataset containing the metadata information with the sample ID as the first column. If a HDAnalyzeR object is provided, this parameter is not needed.}

\item{variable}{The name of the column containing the case and control groups.}

\item{case}{The case group.}

\item{control}{The control groups. If NULL, it will be set to all other unique values of the variable that are not the case.}
}
\value{
An object with the DE results.
}
\description{
\code{hd_de_ttest()} performs differential expression analysis using t-test.
It separates the data in case-control groups, checks for data normality and
perform a t-test or Wilcoxon test respectively. It also performs p value FDR adjustment.
}
\examples{
# Initialize an HDAnalyzeR object
hd_object <- hd_initialize(example_data, example_metadata)

# Run differential expression analysis for AML vs all others
hd_de_ttest(hd_object, case = "AML")

# Run differential expression analysis for AML vs CLL
hd_de_ttest(hd_object, case = "AML", control = "CLL")
}
