% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hd_detect_vartype}
\alias{hd_detect_vartype}
\title{Detect variable type}
\usage{
hd_detect_vartype(var, unique_threshold = 5)
}
\arguments{
\item{var}{The variable (vector or dataframe column)to detect the type of.}

\item{unique_threshold}{The threshold to consider a numeric variable as categorical. Default is 5.}
}
\value{
The type of the variable as a string: "categorical", "continuous", or "unknown".
}
\description{
\code{hd_detect_vartype()} detects the type of a variable based on its content.
If a variable is a factor or character, it is considered categorical.
If a variable is numeric and has less than or equal to \code{unique_threshold} unique values,
it is considered categorical. Otherwise, it is considered continuous.
}
\details{
If you want to apply this function to each column of a dataframe, you can use the
\code{sapply()} function. See examples. For more information check \code{sapply} documentation.
}
\examples{
# Check categorical data
category <- c("A", "B", "A", "C")
hd_detect_vartype(category)

# Check continuous data
continuous <- c(1, 2, 3, 4, 5, 6)
hd_detect_vartype(continuous)

# Apply the function to each column of a dataframe
example <- data.frame(Category = c("A", "B", "A", "C", "B", "A"),
                      Continuous = c(1.1, 2.5, 3.8, 4.0, 5.8, 9),
                      Mixed = c(1, "1", 2, 2, "3", 3))

sapply(example, hd_detect_vartype)
}
