% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hd_filter}
\alias{hd_filter}
\title{Filter an HD object based on a variable}
\usage{
hd_filter(hd_obj, variable, values, flag, verbose = TRUE)
}
\arguments{
\item{hd_obj}{An HD object to be filtered.}

\item{variable}{The name of the variable to filter on.}

\item{values}{The values to filter on (for categorical variables) or the value to compare to (for continuous variables).}

\item{flag}{The type of filter to apply (see details).}

\item{verbose}{A logical indicating whether to print messages during filtering. Default is TRUE.}
}
\value{
The filtered HD object
}
\description{
Filter an HD object based on a variable, either in the data or metadata component.
}
\details{
The \code{flag} argument can take the following values:
\itemize{
\item "k" : Keep only rows where the variable matches the values (categorical)
\item "r" : Remove rows where the variable matches the values (categorical)
\item "=" : Keep only rows where the variable equals the value (continuous)
\item "<" : Keep only rows where the variable is less than the value (continuous)
\item "<=" : Keep only rows where the variable is less than or equal to the value (continuous)
\item ">" : Keep only rows where the variable is greater than the value (continuous)
\item ">=" : Keep only rows where the variable is greater than or equal to the value (continuous)
\item "!=" : Keep only rows where the variable is not equal to the value (continuous)
}
}
\examples{
# Create the HDAnalyzeR object providing the data and metadata
hd_obj <- hd_initialize(example_data, example_metadata)
hd_obj

# Filter by categorical variable
hd_filter(hd_obj, variable = "Sex", values = "F", flag = "k")

# Filter by continuous variable
hd_filter(hd_obj, variable = "Age", values = 80, flag = ">")
}
