% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hd_import_data}
\alias{hd_import_data}
\title{Import data from file}
\usage{
hd_import_data(path_name)
}
\arguments{
\item{path_name}{The path to the file to import.}
}
\value{
The imported data as a tibble or an R object.
}
\description{
\code{hd_import_data()} imports data from a file. The file format can be CSV,
TSV, TXT, RDA, RDS, XLSX, or Parquet format. It recognizes the file format,
reads it and returns it as a tibble or an R object.
}
\examples{
# Save a dataframe as an RDS file
hd_save_data(example_metadata, "my_data/metadata.rds")

# Import the saved RDS file again as a tibble
hd_import_data("my_data/metadata.rds")

unlink("my_data", recursive = TRUE)  # Clean up the created directory
}
