% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{hd_impute_knn}
\alias{hd_impute_knn}
\title{Impute via k-nearest neighbors}
\usage{
hd_impute_knn(dat, k = 5, verbose = TRUE)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format and sample_id as its first column.}

\item{k}{The number of neighbors to consider for imputation.}

\item{verbose}{If TRUE, the percentage of missing values in each column is displayed.}
}
\value{
The imputed dataset.
}
\description{
\code{hd_impute_knn()} imputes missing values in a dataset using the k-nearest neighbors method.
It allows the user to exclude certain columns from imputation and can also display the
percentage of missing values in each column before imputation. The user can also specify
the number of neighbors to consider for imputation.
}
\examples{
# Create the HDAnalyzeR object providing the data and metadata
hd_object <- hd_initialize(example_data, example_metadata)
hd_object$data

# Data after imputation
res <- hd_impute_knn(hd_object, k = 3)
res$data
}
