% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{hd_impute_median}
\alias{hd_impute_median}
\title{Impute via Median}
\usage{
hd_impute_median(dat, verbose = TRUE)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format and sample_id as its first column.}

\item{verbose}{If TRUE, the percentage of missing values in each column is displayed.}
}
\value{
The imputed dataset.
}
\description{
\code{hd_impute_median()} imputes missing values in a dataset using the median of each column.
It allows the user to exclude certain columns from imputation and can also display the
percentage of missing values in each column before imputation.
}
\details{
This is the fastest but usually least accurate imputation method.
}
\examples{
# Create the HDAnalyzeR object providing the data and metadata
hd_object <- hd_initialize(example_data, example_metadata)

# Data after imputation
res <- hd_impute_median(hd_object)
res$data
}
