% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hd_initialize}
\alias{hd_initialize}
\title{Initialize HDAnalyzeR object}
\usage{
hd_initialize(
  dat,
  metadata = NULL,
  is_wide = FALSE,
  sample_id = "DAid",
  var_name = "Assay",
  value_name = "NPX"
)
}
\arguments{
\item{dat}{A tibble containing the omics data.}

\item{metadata}{A tibble containing the metadata. Default is NULL.}

\item{is_wide}{A logical indicating if the data is in wide format. Default is FALSE.}

\item{sample_id}{The name of the column containing the sample IDs. Default is "DAid".}

\item{var_name}{The name of the column containing the variable names. Default is "Assay".}

\item{value_name}{The name of the column containing the values. Default is "NPX".}
}
\value{
An HDAnalyzeR object.
}
\description{
\code{hd_initialize()} initializes an HDAnalyzeR object with the data, metadata, and other parameters. This object can be used to run various analyses in the package.
}
\details{
It is strongly recommended to use this function to initialize an HDAnalyzeR
object before starting any analysis. This initialization step ensures that
your input data and metadata are correctly formatted and compatible with the
package. Specifically, the function validates your data and automatically
converts it to wide format if provided in long format (\code{is_wide}).

While it is possible to use HDAnalyzeR functions on data outside of an HDAnalyzeR object,
users must ensure that the data adheres to the required structure: the first column
must contain sample IDs, followed by numeric columns (protein expression) in wide format.
Additionally, the metadata must include the same set of sample IDs.
Proper initialization or careful adherence to these requirements is crucial
for accurate and efficient analysis. Defaults are provided according to the
Human Disease Blood Atlas Olink data format.
}
\examples{
# Initialize an HDAnalyzeR object
hd_initialize(example_data, example_metadata)
}
