% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hd_log_transform}
\alias{hd_log_transform}
\title{Log transform data with base 2}
\usage{
hd_log_transform(dat)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format.}
}
\value{
The log-transformed data.
}
\description{
\code{hd_log_transform()} log transforms the data in a dataset.
It replaces non-positive values (<= 0) with NA values and informs the user.
}
\examples{
# Create the HDAnalyzeR object providing the data and metadata
hd_object <- hd_initialize(example_data, example_metadata)
hd_object$data

# Log transform the data
hd_object_transformed <- hd_log_transform(hd_object)
# Normally you should not transform Olink data as they are already log-transformed
hd_object_transformed$data
}
