% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hd_long_data}
\alias{hd_long_data}
\title{Convert omics data to long format}
\usage{
hd_long_data(dat, exclude = "DAid", names_to = "Assay", values_to = "NPX")
}
\arguments{
\item{dat}{A tibble containing data in wide format.}

\item{exclude}{The name of the columns to exclude from the transformation.}

\item{names_to}{The name of the column to create for the variable names.}

\item{values_to}{The name of the column to create for the values.}
}
\value{
A tibble containing the data in long format.
}
\description{
\code{hd_long_data()} transforms omics data from wide to long format.
}
\examples{
# Olink data in wide format
example_data_wide <- hd_widen_data(example_data)
example_data_wide

# Transform Olink data in long format
hd_long_data(example_data_wide)
}
