% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{hd_model_test}
\alias{hd_model_test}
\title{Validate model on new data}
\usage{
hd_model_test(
  model_object,
  train_set,
  test_set,
  variable = "Disease",
  metadata_cols = NULL,
  case,
  control = NULL,
  balance_groups = TRUE,
  palette = NULL,
  seed = 123
)
}
\arguments{
\item{model_object}{An \code{hd_model} object coming from \code{hd_model_rreg()} and \code{hd_model_rf()} binary or multiclass classification models.}

\item{train_set}{The training set as an HDAnalyzeR object or a dataset in wide format with sample_id as its first column and class column as its second column.}

\item{test_set}{The validation/test set as an HDAnalyzeR object or a dataset in wide format with sample_id as its first column and class column as its second column.}

\item{variable}{The name of the column containing the case and control groups. Default is "Disease".}

\item{metadata_cols}{The metadata columns to include in the analysis. Default is NULL.}

\item{case}{The case class.}

\item{control}{The control groups. If NULL, it will be set to all other unique values of the variable that are not the case. Default is NULL.}

\item{balance_groups}{Whether to balance the groups in the train set. Default is TRUE.}

\item{palette}{The color palette for the classes. If it is a character, it should be one of the palettes from \code{hd_palettes()}. Default is NULL.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
The model object containing the validation set, the metrics, the ROC curve, the probability plot, and the confusion matrix for the new data.
}
\description{
\code{hd_model_test()} validates the model on new data. It evaluates the model on
the validation (new test) set, calculates the metrics and plots the probability
and ROC curve based on the new data.
}
\examples{
# Initialize an HDAnalyzeR object
hd_object <- hd_initialize(example_data, example_metadata)

# Split the data for training and validation sets
dat <- hd_object$data
train_indices <- sample(1:nrow(dat), size = floor(0.8 * nrow(dat)))
train_data <- dat[train_indices, ]
validation_data <- dat[-train_indices, ]

hd_object_train <- hd_initialize(train_data, example_metadata, is_wide = TRUE)
hd_object_val <- hd_initialize(validation_data, example_metadata, is_wide = TRUE)

# Split the training set into training and inner test sets
hd_split <- hd_split_data(hd_object_train, variable = "Disease")

# Run the regularized regression model pipeline
model_object <- hd_model_rreg(hd_split,
                              variable = "Disease",
                              case = "AML",
                              grid_size = 5,
                              palette = "cancers12")

# Run the model evaluation pipeline
hd_model_test(model_object, hd_object_train, hd_object_val, case = "AML", palette = "cancers12")
}
