% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{hd_pca}
\alias{hd_pca}
\title{Run PCA analysis}
\usage{
hd_pca(dat, components = 10, by_sample = TRUE, seed = 123)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format and sample_id as its first column.}

\item{components}{The number of PCs to be calculated. Default is 10.}

\item{by_sample}{If TRUE, points represent samples. If FALSE, points represent features. Default is TRUE.}

\item{seed}{The seed to be used in the PCA analysis. Default is 123.}
}
\value{
A list with the PCA results, PCA loadings and explained variance.
}
\description{
\code{hd_pca()} runs a PCA analysis on the provided data. If data contain missing values,
the function imputes them using the k-nearest neighbors algorithm (k = 5). The number of PCs
to be calculated is defined by the user. The function returns a tibble with the PCA results.
}
\examples{
# Create the HDAnalyzeR object providing the data and metadata
hd_object <- hd_initialize(example_data, example_metadata)

# Run the PCA analysis
hd_pca(hd_object, components = 5, by_sample = TRUE, seed = 123)

# Run the PCA analysis by feature
hd_pca(hd_object, components = 5, by_sample = FALSE, seed = 123)
}
