% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{hd_plot_feature_boxplot}
\alias{hd_plot_feature_boxplot}
\title{Plot feature boxplots}
\usage{
hd_plot_feature_boxplot(
  dat,
  metadata = NULL,
  variable = "Disease",
  features,
  case,
  type = "case_vs_all",
  points = TRUE,
  x_labels = TRUE,
  yaxis_title = "NPX",
  palette = NULL
)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format and sample_id as its first column.}

\item{metadata}{A dataset containing the metadata information with the sample ID as the first column. If a HDAnalyzeR object is provided, this parameter is not needed.}

\item{variable}{The name of the column containing the case and control groups. Default is "Disease".}

\item{features}{The features to include in the boxplot. They should be columns of the data.}

\item{case}{The case class.}

\item{type}{The type of boxplot to plot. Default is "case_vs_all". Other option is "case_vs_control".}

\item{points}{Whether to add points to the boxplot.}

\item{x_labels}{Whether to show the x-axis labels.}

\item{yaxis_title}{The title of the y-axis. Default is "NPX".}

\item{palette}{The color palette for the classes. If it is a character, it should be one of the palettes from \code{hd_palettes()}. Default is NULL.}
}
\value{
The boxplot panel of the selected features
}
\description{
\code{hd_plot_feature_boxplot()} plots boxplots for the specified features in the dataset.
It annotates the boxplot with color for the selected case
It is also possible to add points to the boxplot.
}
\examples{
# Initialize an HDAnalyzeR object
hd_object <- hd_initialize(example_data, example_metadata)

# Boxplots for AARSD1 and ABL1 in AML vs all other classes
hd_plot_feature_boxplot(hd_object,
                        variable = "Disease",
                        features = c("AARSD1", "ABL1"),
                        case = "AML",
                        palette = "cancers12")

# Boxplots for AARSD1 and ABL1 in AML vs joint control group
hd_plot_feature_boxplot(hd_object,
                        variable = "Disease",
                        features = c("AARSD1", "ABL1"),
                        case = "AML",
                        type = "case_vs_control",
                        palette = "cancers12")
}
