% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{hd_plot_gsea}
\alias{hd_plot_gsea}
\title{Plot gene set enrichment analysis results}
\usage{
hd_plot_gsea(enrichment, seed = 123)
}
\arguments{
\item{enrichment}{The enrichment results obtained from \code{hd_gsea()}.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
The input object enriched with the plots.
}
\description{
\code{hd_plot_gsea()} produces useful plots to visualize the results of the
gene set enrichment analysis.
}
\details{
When KEGG database is used, a cnetplot is generated with ENTREZIDs instead of gene names.
For GO and Reactome databases the ENTREZIDs are converted to gene names.
If you get the "grid.Call(C_convert, x, as.integer(whatfrom), as.integer(whatto),  :
Viewport has zero dimension(s)" warning or error, try to increase the RStudio's viewer window size.
}
\examples{
\dontshow{if (requireNamespace("clusterProfiler", quietly = TRUE) && requireNamespace("enrichplot", quietly = TRUE) && requireNamespace("org.Hs.eg.db", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Initialize an HDAnalyzeR object
hd_object <- hd_initialize(example_data, example_metadata)

# Run differential expression analysis for AML vs all others
de_results <- hd_de_limma(hd_object, case = "AML")

# Run GSEA with Reactome database
enrichment <- hd_gsea(de_results,
                      database = "GO",
                      ontology = "BP",
                      ranked_by = "logFC",
                      pval_lim = 0.9)
# Remember that the data is artificial, this is why we use an absurdly high p-value cutoff

# Plot the results
enrichment <- hd_plot_gsea(enrichment)

# Access the plots
enrichment$dotplot
enrichment$gseaplot
enrichment$cnetplot
enrichment$ridgeplot
\dontshow{\}) # examplesIf}
}
