% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{hd_plot_pca_loadings}
\alias{hd_plot_pca_loadings}
\title{Plot PCA loadings}
\usage{
hd_plot_pca_loadings(pca_object, displayed_pcs = 6, displayed_features = 15)
}
\arguments{
\item{pca_object}{A PCA object containing the PCA loadings. Created by \code{hd_pca()}.}

\item{displayed_pcs}{The number of PCs to be displayed. Default is 6.}

\item{displayed_features}{The number of features to be displayed. Default is 15.}
}
\value{
A ggplot object
}
\description{
\code{hd_plot_pca_loadings()} plots the PCA loadings for the top n features and first m PCs.
n and m are defined by the user. The contribution direction of the features
is indicated by the color of the bars.
}
\examples{
# Create the HDAnalyzeR object providing the data and metadata
hd_object <- hd_initialize(example_data, example_metadata)

# Run the PCA analysis and create the loadings plot
pca_object <- hd_pca(hd_object, components = 5, by_sample = TRUE, seed = 123) |>
  hd_plot_pca_loadings()

pca_object$pca_loadings_plot
}
