% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{hd_plot_pca_variance}
\alias{hd_plot_pca_variance}
\title{Plot PCA variance}
\usage{
hd_plot_pca_variance(pca_object)
}
\arguments{
\item{pca_object}{A PCA object containing the PCA variance. Created by \code{hd_run_pca()}.}
}
\value{
A ggplot object
}
\description{
\code{plot_explained_variance()} plots the explained variance and cumulative explained variance of the PCs.
}
\examples{
# Create the HDAnalyzeR object providing the data and metadata
hd_object <- hd_initialize(example_data, example_metadata)

# Run the PCA analysis and create the variance plot
pca_object <- hd_run_pca(hd_object, components = 5, by_sample = TRUE, seed = 123) |>
  hd_plot_pca_variance()
pca_object$pca_variance_plot
}
