% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{hd_plot_volcano}
\alias{hd_plot_volcano}
\title{Visualize differential expression results}
\usage{
hd_plot_volcano(
  de_object,
  pval_lim = 0.05,
  logfc_lim = 0,
  top_up_prot = 10,
  top_down_prot = 5,
  palette = "diff_exp",
  title = NULL,
  report_nproteins = TRUE,
  user_defined_proteins = NULL
)
}
\arguments{
\item{de_object}{The differential expression object. Created by \code{hd_de_limma()} or \code{hd_de_ttest()}.}

\item{pval_lim}{The p-value limit for significance. Default is 0.05.}

\item{logfc_lim}{The logFC limit for significance. Default is 0.}

\item{top_up_prot}{The number of top up regulated proteins to label on the plot. Default is 10.}

\item{top_down_prot}{The number of top down regulated proteins to label on the plot. Default is 5.}

\item{palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{hd_palettes()}. Default is "diff_exp".}

\item{title}{The title of the plot or NULL for no title.}

\item{report_nproteins}{If the number of significant proteins should be reported in the title. Default is TRUE.}

\item{user_defined_proteins}{A vector with the protein names to label on the plot if customization is required. Default is NULL.}
}
\value{
A DE object with the volcano plot.
}
\description{
\code{hd_plot_volcano()} creates volcano plots for the differential expression results.
It colors and labels the top-n up and down regulated proteins.
}
\examples{
# Initialize an HDAnalyzeR object
hd_object <- hd_initialize(example_data, example_metadata)

# Run differential expression analysis for AML vs all others
de_results <- hd_de_limma(hd_object, case = "AML")

# Create a volcano plot
hd_plot_volcano(de_results)
}
