% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna.R
\name{hd_plot_wgcna}
\alias{hd_plot_wgcna}
\title{Plot WGCNA results}
\usage{
hd_plot_wgcna(dat, metadata = NULL, wgcna, clinical_vars = NULL)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format and sample ID as its first column.}

\item{metadata}{A dataset containing the metadata information with the sample ID as the first column. If a HDAnalyzeR object is provided, this parameter is not needed.}

\item{wgcna}{The WGCNA analysis results obtained from \code{hd_wgcna()}.}

\item{clinical_vars}{A character vector containing the names of the clinical variables to be used in the predictive power score analysis.}
}
\value{
The input object enriched with the plots.
}
\description{
\code{hd_plot_wgcna()} generates useful visualizations for the results of the WGCNA analysis.
The function generates a heatmap of proteins and their adjacency, a heatmap of module
eigengene (MEs) adjacency, heatmaps of predictive power score (PPS) between MEs and
metadata and the dendrogram of all genes.
}
\examples{
\dontshow{if (requireNamespace("WGCNA", quietly = TRUE) && requireNamespace("ppsr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Initialize an HDAnalyzeR object
hd_object <- hd_initialize(example_data, example_metadata)

# Perform WGCNA analysis
wgcna_res <- hd_wgcna(hd_object)

# Plot WGCNA results
wgcna_res <- hd_plot_wgcna(hd_object,
                           wgcna = wgcna_res,
                           clinical_vars = c("Disease", "Sex", "Age", "BMI"))

# Access the plots
wgcna_res$tom_heatmap
wgcna_res$me_adjacency
wgcna_res$pps
wgcna_res$me_pps_heatmap
wgcna_res$var_pps_heatmap
wgcna_res$me_cor_heatmap
wgcna_res$dendrogram
\dontshow{\}) # examplesIf}
}
