% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{hd_run_data_split}
\alias{hd_run_data_split}
\title{Split data into training and test sets}
\usage{
hd_run_data_split(
  dat,
  metadata = NULL,
  variable = "Disease",
  metadata_cols = NULL,
  ratio = 0.75,
  seed = 123
)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format and sample_id as its first column.}

\item{metadata}{A dataset containing the metadata information with the sample ID as the first column. If a HDAnalyzeR object is provided, this parameter is not needed.}

\item{variable}{The name of the column containing the case and control groups. Default is "Disease".}

\item{metadata_cols}{The columns to be selected from the metadata as predictors. Default is NULL.}

\item{ratio}{The ratio of training data to test data. Default is 0.75.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A split object containing train and test data.
}
\description{
\code{hd_run_data_split()} splits the data into training and test sets based on the ratio
provided. It also stratifies the data based on the variable provided.
}
\examples{
# Initialize an HDAnalyzeR object
hd_object <- hd_initialize(example_data, example_metadata)

# Split the data into training and test sets
hd_split <- hd_run_data_split(hd_object, variable = "Disease")
}
