% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{hd_run_gsea}
\alias{hd_run_gsea}
\title{Perform gene set enrichment analysis}
\usage{
hd_run_gsea(
  de_results,
  database = c("GO", "Reactome", "KEGG"),
  ontology = c("BP", "CC", "MF", "ALL"),
  expression = c("both", "up", "down"),
  ranked_by = c("logFC", "adj.P.Val"),
  pval_lim = 0.05,
  logfc_lim = 0,
  pval_lim_enrichment = 0.05
)
}
\arguments{
\item{de_results}{A tibble containing the results of a differential expression analysis.}

\item{database}{The database to perform the ORA. It can be either "GO", "KEGG", or "Reactome".}

\item{ontology}{The ontology to use when database = "GO". It can be "BP" (Biological Process), "CC" (Cellular Component), "MF" (Molecular Function), or "ALL". In the case of KEGG and Reactome, this parameter is ignored.}

\item{expression}{The type of differentially expressed proteins to consider in the analysis. It can be "both", "up", or "down".}

\item{ranked_by}{The metric to rank the proteins. It can be "logFC" or "adj.P.Val".}

\item{pval_lim}{The p-value threshold to consider a term as significant in the differential expression analysis. Default is 0.05.}

\item{logfc_lim}{The log fold change threshold to consider a term as significant in the differential expression analysis. Default is 0.}

\item{pval_lim_enrichment}{The p-value threshold to consider a term as significant in the enrichment analysis. Default is 0.05.}
}
\value{
A list containing the results of the GSEA.
}
\description{
This function performs gene set enrichment analysis (GSEA) using the clusterProfiler package.
}
\examples{
# Initialize an HDAnalyzeR object
hd_object <- hd_initialize(example_data, example_metadata)

# Run differential expression analysis for AML vs all others
de_results <- hd_run_de_limma(hd_object, case = "AML")

# Run GSEA with Reactome database
hd_run_gsea(de_results,
            database = "GO",
            ontology = "BP",
            expression = "both",
            ranked_by = "logFC",
            pval_lim_enrichment = 0.9)
# Remember that the data is artificial, this is why we use an absurdly high p-value cutoff
}
