% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_search.R
\name{hd_run_literature_search}
\alias{hd_run_literature_search}
\title{Run literature search for gene-disease pairs in PubMed}
\usage{
hd_run_literature_search(
  feature_class_list,
  max_articles = 10,
  keywords = NULL,
  fields = "All Fields",
  api_key = NULL,
  verbose = TRUE
)
}
\arguments{
\item{feature_class_list}{A list of features (gene names) and classes (diseases). The names of the list are the classes and the elements are vectors of features. See examples.}

\item{max_articles}{The maximum number of articles to retrieve for each gene-disease pair. Default is 10.}

\item{keywords}{Additional keywords to include in the search. They are added to the query as "AND keywords". Default is NULL.}

\item{fields}{The fields to search for the keywords. Default is "All Fields". Other options are "Title", "Abstract", "Author", "Journal", "Affiliation", "MeSH Terms", "Other Terms".}

\item{api_key}{user-specific API key to increase the limit of queries per second. You can obtain your key from NCBI but not required. Default is NULL.}

\item{verbose}{Whether to print progress messages. Default is TRUE.}
}
\value{
A list of tibbles. Each tibble contains the articles found for a gene-disease pair.
}
\description{
\code{hd_run_literature_search()} searches for articles for gene-disease pairs in PubMed.
A list of proteins and diseases is provided as input. The function retrieves the
articles for each gene-disease pair. The input should be in the correct format,
a list with diseases as names and gene vectors associated with each disease as
elements (see examples).
}
\details{
The disease and gene names should be correct in order for the query to
be successful. For example AML should be written as "acute myeloid leukemia".
The query is constructed as "gene\link{field} AND disease\link{field} AND keywords".
For more details check the \code{easyPubMed} package documentation.
}
\examples{
# Prepare the list of gene-disease pairs
feature_class_list <- list("acute myeloid leukemia" = c("FLT3", "EPO"),
                           "chronic lymphocytic leukemia" = c("PARP1"))

# Run the literature search
lit_search_results <- hd_run_literature_search(feature_class_list, max_articles = 1)

# Results for FLT3 in acute myeloid leukemia
lit_search_results$`acute myeloid leukemia`$FLT3
}
