% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hd_save_path}
\alias{hd_save_path}
\title{Create directory to save results}
\usage{
hd_save_path(path_name, date = FALSE, verbose = TRUE)
}
\arguments{
\item{path_name}{The name of the directory to create.}

\item{date}{If TRUE, a directory with the current date as name will be created inside the directory with \code{dir_name}.}

\item{verbose}{A logical for logging results (other than warnings and errors, which are always shown).}
}
\value{
The relative directory path as a string.
}
\description{
\code{hd_save_path()} creates a directory with in a specified or the current path to save results.
The user can optionally choose to create another inner directory named with the current date.
If the directory already exists, a message is printed.
}
\examples{
# Create a directory
hd_save_path("my_directory", date = FALSE)
unlink("my_directory", recursive = TRUE)  # Clean up the created directory

# Create a directory and an inner directory with the current date as name
hd_save_path("my_directory", date = TRUE)
unlink("my_directory", recursive = TRUE)

# Create a directory inside another directory
hd_save_path("outer_directory/inner_directory", date = FALSE)
unlink("outer_directory", recursive = TRUE)

# Create a directory inside a pre existing one
hd_save_path("outer_directory", date = FALSE)
hd_save_path("outer_directory/inner_directory", date = FALSE)
unlink("outer_directory", recursive = TRUE)

# Create a directory with the current date as name inside a pre existing one
hd_save_path("outer_directory", date = FALSE)
hd_save_path("outer_directory", date = TRUE)
unlink("outer_directory", recursive = TRUE)
}
