% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_background_lists.R
\name{hd_show_backgrounds}
\alias{hd_show_backgrounds}
\title{Show available background lists}
\usage{
hd_show_backgrounds()
}
\value{
A character vector of available background lists.
}
\description{
\code{hd_show_backgrounds()} is used to show the available background lists for Over Representation Analysis (ORA).
}
\examples{
hd_show_backgrounds()
}
