% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna.R
\name{hd_wgcna}
\alias{hd_wgcna}
\title{Weighted gene co-expression network analysis}
\usage{
hd_wgcna(dat, power = NULL)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format and sample ID as its first column.}

\item{power}{The power parameter for the WGCNA analysis as an integer between 1 and 30. If NULL, the function will select an optimal power value. Default is NULL.}
}
\value{
A list containing the results of the WGCNA.
}
\description{
\code{hd_wgcna()} performs a weighted gene co-expression network analysis (WGCNA) on the provided data.
The user can specify the power parameter for the analysis or the function will select an optimal
power value based on the data. The function returns a list containing the WGCNA object, the power
and the power plots (in case optimization is performed). If the data contain missing values, the function
imputes them using the k-nearest neighbors algorithm (k = 5).
}
\details{
If you want to learn more about WGCNA, please refer to the following tutorial:
\itemize{
\item https://edo98811.github.io/WGCNA_official_documentation/
}
}
\examples{
\dontshow{if (requireNamespace("WGCNA", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Initialize an HDAnalyzeR object
hd_object <- hd_initialize(example_data, example_metadata)

# Perform WGCNA analysis
wgcna_res <- hd_wgcna(hd_object)

# Access the WGCNA results
wgcna_res$wgcna$colors
head(wgcna_res$wgcna$MEs)
wgcna_res$power
wgcna_res$power_plots
\dontshow{\}) # examplesIf}
}
