% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hd_widen_data}
\alias{hd_widen_data}
\title{Widen omics data}
\usage{
hd_widen_data(dat, id = "DAid", names_from = "Assay", values_from = "NPX")
}
\arguments{
\item{dat}{A tibble containing data in long format.}

\item{id}{The name of the column containing the sample IDs. If not specified (NULL),
there will be no sample identifiers as a column.}

\item{names_from}{The name of the column containing the variable names.}

\item{values_from}{The name of the column containing the values.}
}
\value{
A tibble containing the data in wide format.
}
\description{
\code{hd_widen_data()} transforms the data from long to wide format. It should be used to
transform omics data from long to wide format with variables like different \code{Assays}
as column names and expression values like \code{NPX} as values.
The first column will contain the samples IDs, unless specified otherwise.
}
\examples{
# Olink data in long format
example_data

# Transform Olink data in wide format
hd_widen_data(example_data)

# Use Sample name instead of Sample ID and Olink IDs instead of Assay names
hd_widen_data(example_data, id = "Sample", names_from = "OlinkID")
}
