% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_df}
\alias{import_df}
\title{Import dataframe}
\usage{
import_df(file_path)
}
\arguments{
\item{file_path}{The path to the file to import}
}
\value{
df The imported dataframe
}
\description{
The function imports a dataframe from a file in CSV, TSV, RDA, RDS, XLSX, XLS or TXT format.
}
\examples{
# df <- import_df("data/example_data.csv")
}
