% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_knn}
\alias{impute_knn}
\title{Impute via k-nearest neighbors}
\usage{
impute_knn(
  olink_data,
  wide = TRUE,
  k = 5,
  exclude_cols = c("DAid", "Disease"),
  show_na_percentage = TRUE
)
}
\arguments{
\item{olink_data}{The input dataset.}

\item{wide}{If TRUE, the data is in wide format.}

\item{k}{The number of neighbors to consider for imputation.}

\item{exclude_cols}{The columns to exclude from imputation.}

\item{show_na_percentage}{If TRUE, the percentage of missing values in each column is displayed.}
}
\value{
The imputed dataset.
}
\description{
\code{impute_knn()} imputes missing values in a dataset using the k-nearest neighbors method.
It allows the user to exclude certain columns from imputation and can also display the
percentage of missing values in each column before imputation. The user can also specify
the number of neighbors to consider for imputation.
}
\examples{
# Data before imputation
test_data <- example_data |>
  dplyr::select(DAid, Assay, NPX) |>
  tidyr::pivot_wider(names_from = "Assay", values_from = "NPX") |>
  dplyr::slice_head(n = 100)
test_data

# Data after imputation
impute_knn(test_data, k = 3)
}
