% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_median}
\alias{impute_median}
\title{Impute via Median}
\usage{
impute_median(
  wide_data,
  exclude_cols = c("DAid", "Disease"),
  show_na_percentage = TRUE
)
}
\arguments{
\item{wide_data}{(tibble). The input dataframe}

\item{exclude_cols}{(string or vector of strings). The columns to exclude from imputation}

\item{show_na_percentage}{(TRUE or NULL). If TRUE, the percentage of missing values in each column is displayed}
}
\value{
imputed_data (tibble). The dataframe with missing values imputed
}
\description{
This function imputes missing values in a dataset using the median of each column.
It allows the user to exclude certain columns from imputation and can also display the
percentage of missing values in each column before imputation.
}
\examples{
test_data <- tibble::tibble(
  A = c(80, 44, NA, 50, 29),
  B = c(30, NA, 85, 70, 54),
  C = c(7, 10, 25, 74, 49)
)
imputed_data <- impute_median(test_data)
}
