% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_missForest}
\alias{impute_missForest}
\title{Impute via missForest}
\usage{
impute_missForest(
  olink_data,
  wide = TRUE,
  maxiter = 10,
  ntree = 100,
  parallelize = "variables",
  ncores = 4,
  exclude_cols = c("DAid", "Disease"),
  show_na_percentage = TRUE
)
}
\arguments{
\item{olink_data}{The input dataset.}

\item{wide}{If TRUE, the data is in wide format.}

\item{maxiter}{The maximum number of iterations.}

\item{ntree}{The number of trees to grow.}

\item{parallelize}{The type of parallelization to use. Options are "no", "variables", or "forests".}

\item{ncores}{The number of cores to use for parallelization.}

\item{exclude_cols}{The columns to exclude from imputation.}

\item{show_na_percentage}{If TRUE, the percentage of missing values in each column is displayed.}
}
\value{
The imputed dataset.
}
\description{
\code{impute_missForest()} imputes missing values in a dataset using the \code{missForest} method.
It allows the user to exclude certain columns from imputation and can also display the
percentage of missing values in each column before imputation. The user can also specify
the maximum number of iterations, the number of trees to grow, the type of parallelization
("no", "variables", or "forests"), as well as the number of cores to use for parallelization.
}
\examples{
# Data before imputation
test_data <- example_data |>
  dplyr::select(DAid, Assay, NPX) |>
  tidyr::pivot_wider(names_from = "Assay", values_from = "NPX") |>
  dplyr::slice_head(n = 100)
test_data

# Data after imputation
impute_missForest(test_data, maxiter = 1, ntree = 50, parallelize = "no")
}
