% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_search.R
\name{literature_search}
\alias{literature_search}
\title{Automated PubMed literature search}
\usage{
literature_search(prot_dis_list, max_articles = 10)
}
\arguments{
\item{prot_dis_list}{(list). A list of proteins and diseases. The names of the list are the diseases and the elements are vectors of proteins.}

\item{max_articles}{(numeric). The maximum number of articles to retrieve for each protein-disease pair. Default is 10.}
}
\value{
articles_database (list). A list of dataframes. Each dataframe contains the articles found for a protein-disease pair.
}
\description{
This function searches for articles on proteins and diseases in PubMed.
A list of proteins and diseases is provided as input. The function retrieves the articles for each protein-disease pair.
The input should be in the correct format: a list with diseases as names and protein vectors associated with each disease as elements.
The disease and gene names should be correct in order for the query to be successful.
}
\examples{
prot_dis_list <- list("acute myeloid leukemia" = c("FLT3", "EPO"),
                      "chronic lymphocytic leukemia" = c("PARP1", "FCER2"))
articles_database <- literature_search(prot_dis_list, max_articles = 2)
}
