% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{lreg_fit}
\alias{lreg_fit}
\title{Fit logistic regression model for single predictors}
\usage{
lreg_fit(
  train_data,
  test_data,
  variable = "Disease",
  case,
  cor_threshold = 0.9,
  cv_sets = 4,
  ncores = 1,
  exclude_cols = NULL,
  palette = NULL,
  seed = 123
)
}
\arguments{
\item{train_data}{Training data set from \code{make_groups()}.}

\item{test_data}{Testing data set from \code{make_groups()}.}

\item{variable}{The variable to predict. Default is "Disease".}

\item{case}{Case to predict.}

\item{cor_threshold}{Threshold of absolute correlation values. This will be used to remove the minimum number of features so that all their resulting absolute correlations are less than this value.}

\item{cv_sets}{Number of cross-validation sets. Default is 5.}

\item{ncores}{Number of cores to use for parallel processing. Default is 4.}

\item{exclude_cols}{Columns to exclude from the data before the model is tuned. Default is NULL.}

\item{palette}{Color palette for the ROC curve. Default is NULL.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A list with two elements:
\itemize{
\item fit_res: A list with 4 elements:
\item lreg_wf: Workflow object.
\item train_set: Training set.
\item test_set: Testing set.
\item final: Fitted model.
\item metrics: A list with the model metrics:
\item accuracy: Accuracy of the model.
\item sensitivity: Sensitivity of the model.
\item specificity: Specificity of the model.
\item auc: AUC of the model.
\item conf_matrix: Confusion matrix of the model.
\item roc_curve: ROC curve of the model.
}
}
\description{
\code{lreg_fit()} fits a logistic regression model for single predictors.
It uses the glm engine for logistic regression and fits the model using the
\code{logistic_reg()} function from the parsnip package. It also calculates the
accuracy, sensitivity, specificity, AUC, and confusion matrix of the model.
}
\keyword{internal}
