% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{make_groups}
\alias{make_groups}
\title{Make groups for classification models}
\usage{
make_groups(
  join_data,
  diseases,
  only_female = NULL,
  only_male = NULL,
  seed = 123
)
}
\arguments{
\item{join_data}{(tibble). Wide data joined with metadata.}

\item{diseases}{(vector). Vector of diseases.}

\item{only_female}{(vector). Vector of diseases that are female specific.}

\item{only_male}{(vector). Vector of diseases that are male specific.}

\item{seed}{(numeric). Seed for reproducibility. Default is 123.}
}
\value{
A list with combined, balanced control-case groups for each disease
}
\description{
This function creates class-balanced groups for classification models.
It separates the data into control and case groups.
It also filters the control data based on the disease in case of sex specific diseases.
}
\keyword{internal}
