% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{na_search}
\alias{na_search}
\title{Summary of missing values}
\usage{
na_search(
  olink_data,
  metadata,
  wide = TRUE,
  metadata_cols = NULL,
  palette = NULL
)
}
\arguments{
\item{olink_data}{The Olink dataset.}

\item{metadata}{The metadata dataset.}

\item{wide}{If TRUE, the data is in wide format.}

\item{metadata_cols}{The metadata columns to include in the summary.}

\item{palette}{The color palettes to use for the heatmap annotations (check examples bellow).}
}
\value{
A list containing the summary of missing values and a heatmap.
}
\description{
\code{na_search()} provides a summary of missing values in a dataset. It allows the user to
specify the metadata columns to include in the summary and the color palette to use for
the heatmap annotations.
}
\details{
When coloring annotations, the user can use custom palettes or the
Human Protein Atlas (HPA) palettes. It is not required to provide a palette
for all annotations, but when a palette is provided, it must be in correct
format (check examples bellow).
}
\examples{
# Use custom palettes for coloring annotations
palette = list(Sex = c(M = "blue", F = "pink"))
na_search(example_data,
          example_metadata,
          wide = FALSE,
          metadata_cols = c("Disease", "Sex"),
          palette = palette)

# Use HPA palettes for coloring annotations
palette = list(Disease = get_hpa_palettes()$cancers12, Sex = get_hpa_palettes()$sex_hpa)
na_search(example_data,
          example_metadata,
          wide = FALSE,
          metadata_cols = c("Disease", "Sex"),
          palette = palette)
}
