% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{optional_packages}
\alias{optional_packages}
\title{Packages that are only needed by some of the functionality}
\usage{
optional_packages
}
\description{
Named vector mapping an optional package to the repository it is installed
from, so that \code{check_installed()} can give an actionable hint.
}
\keyword{internal}
