% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{plot_de_summary}
\alias{plot_de_summary}
\title{Create a summary plot for the differential expression results}
\usage{
plot_de_summary(
  de_results,
  disease_palette = NULL,
  diff_exp_palette = "diff_exp"
)
}
\arguments{
\item{de_results}{(list). A list with the differential expression results.}

\item{disease_palette}{(character or vector). The color palette for the disease. If it is a character, it should be one of the palettes from get_hpa_palettes(). Default is NULL.}

\item{diff_exp_palette}{(character or vector). The color palette for the differential expression. If it is a character, it should be one of the palettes from get_hpa_palettes(). Default is "diff_exp".}
}
\value{
A list containing the following plots:
\itemize{
\item de_barplot: A barplot with the number of significant proteins for each disease.
\item upset_plot_up: An upset plot with the significant up regulated proteins for each disease.
\item upset_plot_down: An upset plot with the significant down regulated proteins for each disease.
}
}
\description{
This function creates summary plots for the differential expression results.
It creates a barplot with the number of significant proteins for each disease.
It also creates upset plots both for the significant up and down regulated proteins for each disease.
}
\examples{
de_results <- do_limma(example_data, example_metadata, wide = FALSE)
plots <- plot_de_summary(de_results)
}
