% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot the results of the ORA}
\usage{
plot_enrichment(enrichment)
}
\arguments{
\item{enrichment}{(list). The results of the ORA.}
}
\value{
dotplot (plot). A dotplot showing the results of the ORA.
}
\description{
Plot the results of the ORA
}
\examples{
enrichment <- do_ora(c("TP53", "BRCA1", "BRCA2"), database = "KEGG")
plots <- plot_enrichment(enrichment)
}
