% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{plot_explained_variance}
\alias{plot_explained_variance}
\title{Plot explained variance and cumulative explained variance}
\usage{
plot_explained_variance(explained_variance)
}
\arguments{
\item{explained_variance}{A tibble with the explained variance and cumulative explained variance.}
}
\value{
A ggplot object
}
\description{
\code{plot_explained_variance()} plots the explained variance and cumulative explained variance.
}
\keyword{internal}
