% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{plot_gsea}
\alias{plot_gsea}
\title{Plot the results of the gene set enrichment analysis}
\usage{
plot_gsea(enrichment, de_results, pval_lim = 0.05, ncateg = 10, fontsize = 10)
}
\arguments{
\item{enrichment}{The results of the gene set enrichment analysis.}

\item{de_results}{A tibble containing the results of a differential expression analysis. It should be the same as the one used in \code{do_gsea()}.}

\item{pval_lim}{The p-value threshold to consider a term as significant.}

\item{ncateg}{The number of categories to show in the plots.}

\item{fontsize}{The font size for the plots.}
}
\value{
A list containing the plots.
}
\description{
\code{plot_gsea()} produces useful plots to visualize the results of the
gene set enrichment analysis.
}
\examples{
# Perform Differential Expression Analysis
control = c("BRC", "CLL", "CRC", "CVX", "ENDC", "GLIOM", "LUNGC", "LYMPH", "MYEL", "OVC", "PRC")
de_res <- do_limma(example_data,
                   example_metadata,
                   case = "AML",
                   control = control,
                   wide = FALSE)
de_results <- de_res$de_results

# Run GSEA with Reactome database
enrichment <- do_gsea(de_results, database = "GO", pval_lim = 0.9)

# Plot the results
plot_gsea(enrichment, de_results, pval_lim = 0.9, ncateg = 7, fontsize = 7)
# Remember that the data is artificial, this is why we use an absurdly high p-value cutoff
}
