% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{plot_missing_values}
\alias{plot_missing_values}
\title{Create the missing value distribution}
\usage{
plot_missing_values(missing_values, yaxis_name)
}
\arguments{
\item{missing_values}{A tibble with the column/row names and the percentage of NAs in each column/row.}

\item{yaxis_name}{The name of the y-axis.}
}
\value{
A histogram of the missing value distribution.
}
\description{
\code{plot_missing_values()} creates a histogram of the missing value distribution.
}
\keyword{internal}
