% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{plot_ora}
\alias{plot_ora}
\title{Plot the results of the over-representation analysis}
\usage{
plot_ora(enrichment, protein_list, pval_lim = 0.05, ncateg = 10, fontsize = 10)
}
\arguments{
\item{enrichment}{The results of the over-representation analysis.}

\item{protein_list}{A character vector containing the protein names. It should be the same as the one used in \code{do_ora()}.}

\item{pval_lim}{The p-value threshold to consider a term as significant.}

\item{ncateg}{The number of categories to show in the plots.}

\item{fontsize}{The font size for the plots.}
}
\value{
A list containing the plots.
}
\description{
\code{plot_ora()} produces useful plots to visualize the results of the
over-representation analysis.
}
\examples{
# Perform Differential Expression Analysis
control = c("BRC", "CLL", "CRC", "CVX", "ENDC", "GLIOM", "LUNGC", "LYMPH", "MYEL", "OVC", "PRC")
de_res <- do_limma(example_data,
                   example_metadata,
                   case = "AML",
                   control = control,
                   wide = FALSE)

# Extract the up-regulated proteins for AML
sig_up_proteins_aml <- de_res$de_results |>
  dplyr::filter(sig == "significant up") |>
  dplyr::pull(Assay)

# Perform ORA with GO database
enrichment <- do_ora(sig_up_proteins_aml, database = "GO")

# Plot the results
plot_ora(enrichment, sig_up_proteins_aml, pval_lim = 0.05, ncateg = 5)
}
