% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{plot_points}
\alias{plot_points}
\title{Plot sample data points in a two-dimensional plane}
\usage{
plot_points(dim_res, x, y, color = NULL)
}
\arguments{
\item{dim_res}{A tibble with the results of the dimensionality reduction analysis.}

\item{x}{The name of the column in \code{dim_res} that contains the x-axis values.}

\item{y}{The name of the column in \code{dim_res} that contains the y-axis values.}

\item{color}{The name of the column in \code{dim_res} that contains the variable to be used to plot the points color.}
}
\value{
A ggplot object
}
\description{
\code{plot_points()} plots the sample data points in a two-dimensional plane.
The points can be plotted in the PCx/PCy or UMAP1/UMAP2 space.
}
\keyword{internal}
