% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_protein_boxplot}
\alias{plot_protein_boxplot}
\title{Create boxplots for proteins}
\usage{
plot_protein_boxplot(join_data, proteins, disease, points = T, palette = NULL)
}
\arguments{
\item{join_data}{(tibble). The dataset with the wide Olink data joined with the metadata.}

\item{proteins}{(vector). The proteins to include in the boxplot.}

\item{disease}{(character). The disease to annotate.}

\item{points}{(logical). Whether to add points to the boxplot.}

\item{palette}{(character). The color palette to use. Default is red3.}
}
\value{
boxplot_panel (plot). The boxplot panel with the selected proteins.
}
\description{
This function creates a boxplot for the top n proteins in the dataset.
It annotates the boxplot with color for the selected disease.
It is also possible to add points to the boxplot.
}
\examples{
wide_data <- widen_data(example_data)
join_data <- wide_data |>
  dplyr::left_join(example_metadata |> dplyr::select(DAid, Disease, Sex))
plot_protein_boxplot(join_data, c("AARSD1", "ABL1"), "AML", palette = "cancers12")
}
