% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_scatter_with_regression}
\alias{plot_scatter_with_regression}
\title{Create a scatter plot with regression line}
\usage{
plot_scatter_with_regression(
  plot_data,
  x,
  y,
  se = F,
  line_color = "black",
  r_2 = T
)
}
\arguments{
\item{plot_data}{(tibble). The wide dataset containing the data to plot as cols.}

\item{x}{(character). The column name of the x-axis variable.}

\item{y}{(character). The column name of the y-axis variable.}

\item{se}{(logical). Whether to add the standard error of the regression line.}

\item{line_color}{(character). The color of the regression line.}

\item{r_2}{(logical). Whether to add the R-squared and p-value to the plot.}
}
\value{
scatter (plot). The scatter plot with the regression line.
}
\description{
This function creates a scatter plot with a linear regression line.
It is possible to add the standard error of the regression line, as well as the R-squared and p-value.
}
\examples{
plot_scatter_with_regression(example_metadata, "Age", "BMI")
}
