% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for model fitting}
\usage{
prepare_data(
  dat,
  variable = "Disease",
  case,
  control = NULL,
  balance_groups = TRUE,
  cv_sets = 5,
  seed = 123
)
}
\arguments{
\item{dat}{An \code{hd_model} object coming from \code{check_data()}.}

\item{variable}{The variable to predict. Default is "Disease".}

\item{case}{The case class.}

\item{control}{The control class. Default is NULL.}

\item{balance_groups}{Whether to balance the groups. Default is TRUE.}

\item{cv_sets}{Number of cross-validation sets. Default is 5.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A model object containing the train and test data and cross-validation sets.
}
\description{
\code{prepare_data()} prepares the data for model fitting. It filters out rows with
NAs in the variable column, converts the variable to a factor, and creates
cross-validation sets.
}
\keyword{internal}
