% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{prepare_set}
\alias{prepare_set}
\title{Prepare the data for the \code{hd_model_test()} function}
\usage{
prepare_set(dat, variable, metadata_cols = NULL)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format with sample_id as its first column and class column as its second column.}

\item{variable}{The name of the column containing the case and control groups. Default is "Disease".}

\item{metadata_cols}{The metadata columns to include in the analysis. Default is NULL.}
}
\value{
The prepared data.
}
\description{
\code{prepare_set()} prepares the data for the \code{hd_model_test()} function.
}
\keyword{internal}
