% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{qc_summary_metadata}
\alias{qc_summary_metadata}
\title{Summarize the quality control results of metadata}
\usage{
qc_summary_metadata(metadata, report = T)
}
\arguments{
\item{metadata}{(tibble). The metadata dataframe.}

\item{report}{(logical). Whether to print the summary. Default is TRUE.}
}
\value{
A list containing the following elements:
\itemize{
\item na_percentage_col (tibble). A tibble with the column names and the percentage of NAs in each column.
\item na_percentage_row (tibble). A tibble with the DAids and the percentage of NAs in each row.
}
}
\description{
The function summarizes the quality control results of the input dataframe.
}
\examples{
qc_summary_metadata(example_metadata)
}
