% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_na}
\alias{remove_na}
\title{Check for NAs in a column and remove these rows}
\usage{
remove_na(df_in, cols)
}
\arguments{
\item{df_in}{(tibble). The input dataframe}

\item{cols}{(string or vector of strings). The column to check for NAs}
}
\value{
df_out (tibble). The dataframe with NAs removed
}
\description{
The function checks for NAs in the specified column and removes the rows with NAs.
It returns a warning message if any rows are removed stating the number and the indexes of removed rows.
}
\examples{
df <- data.frame(x = c(1, 2, NA, 4), y = c(NA, 2, 3, 4))
df_out <- remove_na(df, "x")
}
