% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{replace_with_na}
\alias{replace_with_na}
\title{Replace specific values with NA}
\usage{
replace_with_na(
  df_in,
  replace_w_na = c(0, "0", "", "Unknown", "unknown", "none", NA, "na")
)
}
\arguments{
\item{df_in}{The input dataframe.}

\item{replace_w_na}{The values to replace with NA. Default is c(0, "0", "", "Unknown", "unknown", "none", NA, "na").}
}
\value{
The tibble with the specified values replaced with NA.
}
\description{
\code{replace_with_na()} replaces the specified values in the input vector with NAs.
}
\keyword{internal}
