% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{rf_hypopt}
\alias{rf_hypopt}
\title{Hyperparameter optimization for random forest}
\usage{
rf_hypopt(
  train_data,
  test_data,
  disease,
  cv_sets = 5,
  grid_size = 10,
  ncores = 4,
  hypopt_vis = TRUE,
  exclude_cols = NULL,
  seed = 123
)
}
\arguments{
\item{train_data}{(list). List of training data sets from make_groups().}

\item{test_data}{(list). List of testing data sets from make_groups().}

\item{disease}{(character). Disease to predict.}

\item{cv_sets}{(numeric). Number of cross-validation sets. Default is 5.}

\item{grid_size}{(numeric). Size of the grid for hyperparameter optimization. Default is 10.}

\item{ncores}{(numeric). Number of cores to use for parallel processing. Default is 4.}

\item{hypopt_vis}{(logical). Whether to visualize hyperparameter optimization results. Default is TRUE.}

\item{exclude_cols}{(vector). Columns to exclude from the model. Default is NULL.}

\item{seed}{(numeric). Seed for reproducibility. Default is 123.}
}
\value{
A list with four elements:
\itemize{
\item rf_tune (tibble). Hyperparameter optimization results.
\item rf_wf (workflow). Workflow object.
\item train_set (tibble). Training set.
\item test_set (tibble). Testing set.
\item hyperopt_vis (plot). Hyperparameter optimization plot.
}
}
\description{
This function performs hyperparameter optimization for random forest models.
It uses the ranger engine for logistic regression and tunes the number of trees and
the number of variables randomly sampled at each split.
}
\keyword{internal}
