% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{rf_hypopt_multi}
\alias{rf_hypopt_multi}
\title{Hyperparameter optimization for random forest multiclassification model}
\usage{
rf_hypopt_multi(
  train_data,
  test_data,
  cor_threshold = 0.9,
  normalize = TRUE,
  cv_sets = 5,
  grid_size = 10,
  ncores = 4,
  hypopt_vis = TRUE,
  exclude_cols = NULL,
  seed = 123
)
}
\arguments{
\item{train_data}{Training data set from \code{make_groups()}.}

\item{test_data}{Testing data set from \code{make_groups()}.}

\item{cor_threshold}{Threshold of absolute correlation values. This will be used to remove the minimum number of features so that all their resulting absolute correlations are less than this value.}

\item{normalize}{Whether to normalize numeric data to have a standard deviation of one and a mean of zero. Default is TRUE.}

\item{cv_sets}{Number of cross-validation sets. Default is 5.}

\item{grid_size}{Size of the grid for hyperparameter optimization. Default is 10.}

\item{ncores}{Number of cores to use for parallel processing. Default is 4.}

\item{hypopt_vis}{Whether to visualize hyperparameter optimization results. Default is TRUE.}

\item{exclude_cols}{Columns to exclude from the data before the model is tuned. Default is NULL.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A list with five elements:
\itemize{
\item rf_tune: Hyperparameter optimization results.
\item rf_wf: Workflow object.
\item train_set: Training set.
\item test_set: Testing set.
\item hyperopt_vis: Hyperparameter optimization plot.
}
}
\description{
\code{rf_hypopt_multi()} performs hyperparameter optimization for random forest models.
It uses the ranger engine for multinomial regression and tunes the number of
predictors that will be randomly sampled at each split when creating the
tree models, as well as the minimum number of data points in a node that are
required for the node to be split further. For the hyperparameter optimization,
it uses the \code{grid_space_filling()} function from the dials package.
}
\keyword{internal}
