% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{run_de}
\alias{run_de}
\title{Run differential expression analysis}
\usage{
run_de(
  olink_data,
  metadata,
  correct_sex = F,
  correct_age = F,
  correct_bmi = F,
  wide = T,
  only_female = NULL,
  only_male = NULL,
  stat = "limma",
  volcano = T,
  save = F
)
}
\arguments{
\item{olink_data}{(tibble). A tibble with the Olink data in wide format.}

\item{metadata}{(tibble). A tibble with the metadata}

\item{correct_sex}{(logical). Correct the results with Sex. Default is FALSE}

\item{correct_age}{(logical). Correct the results with Age. Default is FALSE}

\item{correct_bmi}{(logical). Correct the results with BMI. Default is FALSE}

\item{wide}{(logical). If the data is in wide format. Default is TRUE}

\item{only_female}{(character or vector). The female specific diseases. Default is NULL}

\item{only_male}{(character or vector). The male specific diseases. Default is NULL}

\item{stat}{(character). The statistical test to use (limma or ttest). Default is "limma"}

\item{volcano}{(logical). Generate volcano plots. Default is TRUE}

\item{save}{(logical). Save the volcano plots. Default is FALSE}
}
\value{
de_results (list). A list with the differential expression results and volcano plots
\itemize{
\item de_results: A list with the differential expression results
\item volcano_plots: A list with the volcano plots
}
}
\description{
This function runs differential expression analysis using limma or t-test.
It can generate and save volcano plots.
}
\examples{
test_data <- example_data |> dplyr::select(DAid, Assay, NPX)
run_de(test_data, example_metadata, wide = FALSE)
}
