% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_color_hpa}
\alias{scale_color_hpa}
\title{HPA color scales}
\usage{
scale_color_hpa(palette = "palette1")
}
\arguments{
\item{palette}{(character). The name of the palette to use. Available palettes are: "sex", "diff_exp", "cancers12", "cancers15".}
}
\value{
A ggplot2 scale for color aesthetics.
}
\description{
This function creates a ggplot2 scale for color aesthetics using the color palettes from the Human Protein Atlas (HPA) project.
}
\keyword{internal}
