% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_background_lists.R
\name{select_background}
\alias{select_background}
\title{Select background list for Over Representation Analysis (ORA)}
\usage{
select_background(background)
}
\arguments{
\item{background}{A character string specifying the background list to be used.}
}
\value{
A character vector of gene symbols.
}
\description{
\code{select_background()} is used to select a background list for Over Representation Analysis (ORA).
}
\keyword{internal}
