% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{testfit}
\alias{testfit}
\title{Test the final model}
\usage{
testfit(
  train_set,
  test_set,
  disease,
  finalfit_res,
  exclude_cols = NULL,
  type = "lasso",
  seed = 123,
  palette = NULL
)
}
\arguments{
\item{train_set}{(tibble). Training set.}

\item{test_set}{(tibble). Testing set.}

\item{disease}{(character). Disease to predict.}

\item{finalfit_res}{(list). Results from elnet_finalfit().}

\item{exclude_cols}{(vector). Columns to exclude from the model. Default is NULL.}

\item{type}{(character). Type of regularization. Default is "lasso". Other options are "ridge" and "elnet".}

\item{seed}{(numeric). Seed for reproducibility. Default is 123.}

\item{palette}{(character or vector). The color palette for the plot. If it is a character, it should be one of the palettes from get_hpa_palettes(). Default is NULL.}
}
\value{
A list with two elements:
\itemize{
\item metrics (list). A list with 5 metrics:
\item accuracy (numeric). Accuracy of the model.
\item sensitivity (numeric). Sensitivity of the model.
\item specificity (numeric). Specificity of the model.
\item auc (numeric). AUC of the model.
\item conf_matrix (tibble). Confusion matrix of the model.
\item roc_curve (tibble). ROC curve of the model.
\item mixture (numeric). Mixture of lasso and ridge regularization.
}
}
\description{
This function tests the final model on the test set.
It calculates the accuracy, sensitivity, specificity, AUC, and confusion matrix.
}
\keyword{internal}
