% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_hpa}
\alias{theme_hpa}
\title{HPA theme}
\usage{
theme_hpa(angled = F, axis_x = T, axis_y = T, facet_title = T)
}
\arguments{
\item{angled}{(logical). If TRUE, the x-axis text is angled at 90 degrees. Default is FALSE.}

\item{axis_x}{(logical). If FALSE, the x-axis is removed. Default is TRUE.}

\item{axis_y}{(logical). If FALSE, the y-axis is removed. Default is TRUE.}

\item{facet_title}{(logical). If FALSE, the facet title is removed. Default is TRUE.}
}
\value{
A ggplot2 theme object.
}
\description{
This function creates a theme for ggplot2 plots that is used by the Human Protein Atlas (HPA) project.
}
\examples{
plot <- example_metadata |>
  ggplot2::ggplot(ggplot2::aes(x = Sex)) +
  ggplot2::geom_bar() +
  theme_hpa()
}
