% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{tune_lr_model}
\alias{tune_lr_model}
\title{Hyperparameter optimization for logistic regression models}
\usage{
tune_lr_model(
  dat,
  variable = "Disease",
  cor_threshold = 0.9,
  verbose = TRUE,
  seed = 123
)
}
\arguments{
\item{dat}{An \code{hd_model} object coming from \code{prepare_data()}.}

\item{variable}{The variable to predict. Default is "Disease".}

\item{cor_threshold}{Threshold of absolute correlation values. This will be used
to remove the minimum number of features so that all their resulting absolute
correlations are less than this value.}

\item{verbose}{Whether to print progress messages. Default is TRUE.}

\item{seed}{Seed for reproducibility. Default is 123.}

\item{grid_size}{Size of the hyperparameter optimization grid. Default is 10.}
}
\value{
A model object containing the train and test data, the tuned model, and the workflow.
}
\description{
\code{tune_rf_model()} performs hyperparameter optimization for logistic regression
models. It tunes the model using the provided grid size and
cross-validation sets. It returns the best model and hyperparameters.
}
\keyword{internal}
