% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{variable_imp}
\alias{variable_imp}
\title{Extract model features and plot variable importance}
\usage{
variable_imp(
  dat,
  variable = "Disease",
  case,
  mixture = NULL,
  palette = NULL,
  title = c("accuracy", "sensitivity", "specificity", "auc", "features", "top-features"),
  verbose = TRUE,
  seed = 123
)
}
\arguments{
\item{dat}{An \code{hd_model} object coming from \code{evaluate_model()}.}

\item{variable}{The variable to predict. Default is "Disease".}

\item{case}{The case class.}

\item{mixture}{The mixture parameter for the elastic net. If NULL it will be tuned. Default is NULL.}

\item{palette}{The color palette for the classes. If it is a character, it should be one of the palettes from \code{hd_palettes()}. Default is NULL.}

\item{title}{Vector of title elements to include in the plot.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A model object containing the features and the variable importance plot.
}
\description{
\code{variable_imp()} calculates the variable importance of the model and plots the top features.
It also generates a title for the plot based on the model metrics and the mixture parameter.
}
\keyword{internal}
