% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{vis_hypopt}
\alias{vis_hypopt}
\title{Visualize hyperparameter optimization results}
\usage{
vis_hypopt(tune_res, x, color, disease)
}
\arguments{
\item{tune_res}{(tibble). Hyperparameter optimization results.}

\item{x}{(character). X-axis variable of the plot.}

\item{color}{(character). Color variable of the plot. Default is NULL.}

\item{disease}{(character). Disease to predict.}
}
\value{
hypopt_plot (plot). Hyperparameter optimization plot.
}
\description{
This function visualizes the hyperparameter optimization results.
}
\keyword{internal}
