% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{xgboost_hypopt_multi}
\alias{xgboost_hypopt_multi}
\title{Hyperparameter optimization for XGBoost multiclassification model}
\usage{
xgboost_hypopt_multi(
  train_data,
  test_data,
  variable = "Disease",
  cor_threshold = 0.9,
  normalize = TRUE,
  cv_sets = 5,
  grid_size = 10,
  ncores = 4,
  hypopt_vis = TRUE,
  exclude_cols = NULL,
  seed = 123
)
}
\arguments{
\item{train_data}{Training data set from \code{make_groups()}.}

\item{test_data}{Testing data set from \code{make_groups()}.}

\item{variable}{The variable to predict. Default is "Disease".}

\item{cor_threshold}{Threshold of absolute correlation values. This will be used to remove the minimum number of features so that all their resulting absolute correlations are less than this value.}

\item{normalize}{Whether to normalize numeric data to have a standard deviation of one and a mean of zero. Default is TRUE.}

\item{cv_sets}{Number of cross-validation sets. Default is 5.}

\item{grid_size}{Size of the grid for hyperparameter optimization. Default is 10.}

\item{ncores}{Number of cores to use for parallel processing. Default is 4.}

\item{hypopt_vis}{Whether to visualize hyperparameter optimization results. Default is TRUE.}

\item{exclude_cols}{Columns to exclude from the data before the model is tuned. Default is NULL.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A list with five elements:
\itemize{
\item xgboost_tune: Hyperparameter optimization results.
\item xgboost_wf: Workflow object.
\item train_set: Training set.
\item test_set: Testing set.
\item hypopt_vis: Hyperparameter optimization plot.
}
}
\description{
\code{xgboost_hypopt_multi()} tunes an XGBoost model and performs hyperparameter optimization.
It uses the xgboost engine for multinomial regression and tunes the number of trees,
tree depth, minimum number of data points in a node, loss reduction, sample size,
and number of predictors randomly sampled at each split. For the hyperparameter optimization,
it uses the \code{grid_space_filling()} function from the dials package.
}
\keyword{internal}
