% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{calc_na_percentage_row}
\alias{calc_na_percentage_row}
\title{Calculate the percentage of NAs in each row}
\usage{
calc_na_percentage_row(df)
}
\arguments{
\item{df}{(tibble). The input dataframe.}
}
\value{
na_percentage (tibble). A tibble with the DAids and the percentage of NAs in each row.
}
\description{
The function calculates the percentage of NAs in each row of the input dataframe.
It filters out the rows with 0\% missing data and returns the rest in descending order.
}
\keyword{internal}
