% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{calc_na_percentage_row}
\alias{calc_na_percentage_row}
\title{Calculate the percentage of NAs in each row of the dataset}
\usage{
calc_na_percentage_row(dat, sample_id)
}
\arguments{
\item{dat}{The input dataset.}

\item{sample_id}{The name of the column containing the sample IDs.}
}
\value{
A tibble with the DAids and the percentage of NAs in each row.
}
\description{
\code{calc_na_percentage_row()} calculates the percentage of NAs in each row of the input dataset.
It filters out the rows with 0\% missing data and returns the rest in descending order.
}
\keyword{internal}
