% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{check_col_types}
\alias{check_col_types}
\title{Check the column types of the dataset}
\usage{
check_col_types(df)
}
\arguments{
\item{df}{The input dataset.}
}
\value{
A table with the counts of each class in the dataset.
}
\description{
\code{check_col_types()} checks the column types of the input dataset and
returns the counts of each class.
}
\keyword{internal}
