% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{check_normality}
\alias{check_normality}
\title{Check normality of the data}
\usage{
check_normality(df)
}
\arguments{
\item{df}{The input dataset.}
}
\value{
A tibble with the protein names, p-values, adjusted p-values, and normality status.
}
\description{
\code{check_normality()} checks the normality of the input dataset using the Shapiro-Wilk test.
It performs the test in parallel and returns the p-values, adjusted p-values, and
normality status for each protein.
}
\keyword{internal}
