% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{check_normality}
\alias{check_normality}
\title{Check normality of the data}
\usage{
check_normality(df)
}
\arguments{
\item{df}{(tibble). The input dataframe}
}
\value{
normality_results (tibble). A tibble with the protein names, p-values, adjusted p-values, and normality status
}
\description{
Check normality of the data
}
\examples{
df <- example_data |>
  dplyr::select(DAid, Assay, NPX) |>
  tidyr::pivot_wider(names_from = "Assay", values_from = "NPX")
normality_results <- check_normality(df)
}
