% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{clean_data}
\alias{clean_data}
\title{Clean data}
\usage{
clean_data(
  df_in,
  keep_cols = c("DAid", "Assay", "NPX"),
  cohort = F,
  exclude_plates = F,
  filter_assay = F,
  apply_replacement = F,
  remove_na_cols = c("DAid", "NPX")
)
}
\arguments{
\item{df_in}{(tibble). The input dataframe}

\item{keep_cols}{(string or vector of strings). The columns to keep in the output dataframe}

\item{cohort}{(string or vector of strings). The cohort to keep}

\item{exclude_plates}{(string or vector of strings). The plates to exclude}

\item{filter_assay}{(logical). If T only rows with Assay_Warning == "PASS" are kept, else F}

\item{apply_replacement}{(logical). If T, the specified values are replaced with NA}

\item{remove_na_cols}{(string or vector of strings or F). The columns to check for NAs and remove respective rows}
}
\value{
df_out (tibble). The cleaned dataframe
}
\description{
The function cleans the data by filtering out rows based on the specified criteria.
It keeps only the specified columns. It removes rows with NAs in the DAid and NPX columns.
}
\examples{
df <- clean_data(example_data, exclude_plates = c("Plate1", "Plate2"), filter_assay = TRUE)
}
