% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{clean_data}
\alias{clean_data}
\title{Clean data}
\usage{
clean_data(
  df_in,
  keep_cols = c("DAid", "Assay", "NPX"),
  cohort = NULL,
  exclude_plates = NULL,
  filter_assay = NULL,
  filter_qc = NULL
)
}
\arguments{
\item{df_in}{The input dataframe}

\item{keep_cols}{The columns to keep in the output dataframe}

\item{cohort}{The cohort to keep}

\item{exclude_plates}{The plates to exclude}

\item{filter_assay}{If TRUE only rows with Assay_Warning == "PASS" are kept, else NULL}

\item{filter_qc}{If TRUE only rows with QC_Warning == "PASS" are kept, else NULL}
}
\value{
df_out The cleaned dataframe
}
\description{
The function cleans the data by filtering out rows based on the specified criteria.
It keeps only the specified columns.
}
\examples{
df <- clean_data(example_data, exclude_plates = c("Plate1", "Plate2"), filter_assay = TRUE)
}
