% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{clean_metadata}
\alias{clean_metadata}
\title{Preprocess metadata}
\usage{
clean_metadata(
  df_in,
  keep_cols = c("DAid", "Disease", "Sex", "Age", "BMI"),
  filter_samples = NULL,
  remove_na_cols = c("DAid", "Disease"),
  replace_w_na = c("Unknown", "unknown", "none", NA, "na")
)
}
\arguments{
\item{df_in}{The input metadata.}

\item{keep_cols}{The columns to keep in the output metadata.}

\item{filter_samples}{The samples to filter out.}

\item{remove_na_cols}{The columns to check for NAs and remove respective rows. Defaults is c("DAid", "Disease").}

\item{replace_w_na}{The values to replace with NA. Default is c("Unknown", "unknown", "none", NA, "na").}
}
\value{
The preprocessed metadata.
}
\description{
\code{clean_metadata()} preprocesses the metadata by filtering out rows based on the specified criteria.
\itemize{
\item It keeps only the specified columns.
\item It filters out the specified samples.
\item It removes rows with NAs in the DAid and Disease columns.
\item It replaces the specified values with NA.
}
}
\examples{
# Unprocessed metadata
example_metadata

# Preprocessed metadata
clean_metadata(example_metadata)
}
