% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{clean_metadata}
\alias{clean_metadata}
\title{Clean metadata}
\usage{
clean_metadata(
  df_in,
  keep_cols = c("DAid", "Disease", "Sex", "Age", "BMI"),
  exclude_sample = NULL,
  apply_replacement = NULL,
  remove_na_cols = c("DAid", "Disease")
)
}
\arguments{
\item{df_in}{(tibble). The input metadata}

\item{keep_cols}{(string or vector of strings). The columns to keep in the output metadata}

\item{exclude_sample}{(string). The value in the "Exclude Sample" column to exclude}

\item{apply_replacement}{(TRUE or NULL). If TRUE, the specified values are replaced with NA}

\item{remove_na_cols}{(string or vector of strings or NULL). The columns to check for NAs and remove respective rows}
}
\value{
df_out (tibble). The cleaned metadata
}
\description{
The function cleans the metadata by filtering out rows based on the specified criteria.
It keeps only the specified columns. It removes rows with NAs in the DAid and Disease columns.
}
\examples{
df <- clean_metadata(example_metadata, exclude_sample = "yes")
}
