% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_data}
\alias{cluster_data}
\title{Cluster data}
\usage{
cluster_data(
  df,
  distance_method = "euclidean",
  clustering_method = "ward.D2",
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  wide = TRUE
)
}
\arguments{
\item{df}{The dataset to cluster.}

\item{distance_method}{The distance method to use. Default is "euclidean".}

\item{clustering_method}{The clustering method to use. Default is "ward.D2".}

\item{cluster_rows}{Whether to cluster rows. Default is TRUE.}

\item{cluster_cols}{Whether to cluster columns. Default is TRUE.}

\item{wide}{Whether the data is wide or long. Default is TRUE.}
}
\value{
(list). A list with the following elements:
\itemize{
\item clustered_data: A dataset ordered according to the hierarchical clustering of the rows and columns.
\item hc_rows: The hierarchical clustering object for rows.
\item hc_cols: The hierarchical clustering object for columns.
}
}
\description{
\code{cluster_data()} takes a dataset and returns the same dataset ordered
according to the hierarchical clustering of the rows and columns. This data
can be used to plot a heatmap with ggplot2 that is not having clustering functionality.
}
\examples{
# Original data
clean_df <- example_data |> dplyr::select(DAid, Assay, NPX)
clean_df

# Clustered data
cluster_data(clean_df, wide = FALSE)
}
