% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{create_corr_heatmap}
\alias{create_corr_heatmap}
\title{Plot correlation heatmap}
\usage{
create_corr_heatmap(
  x,
  y = NULL,
  use = "pairwise.complete.obs",
  method = "pearson",
  threshold = 0.8,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  show_heatmap = TRUE
)
}
\arguments{
\item{x}{A numeric vector, matrix or data frame.}

\item{y}{A numeric vector, matrix or data frame with compatible dimensions with \code{x}. Default is NULL.}

\item{use}{A character string. The method to use for computing correlations. Default is "pairwise.complete.obs".}

\item{method}{A character string. The correlation method to use. Default is "pearson".}

\item{threshold}{The reporting protein-protein correlation threshold. Default is 0.8.}

\item{cluster_rows}{Whether to cluster the rows. Default is TRUE.}

\item{cluster_cols}{Whether to cluster the columns. Default is TRUE.}

\item{show_heatmap}{Whether to show the heatmap. Default is TRUE.}
}
\value{
A list containing the following elements:
\itemize{
\item cor_matrix: A matrix of protein-protein correlations.
\item cor_results: A tibble with the filtered protein pairs and their correlation values.
\item cor_plot: A heatmap of protein-protein correlations.
}
}
\description{
\code{create_corr_heatmap()} calculates the correlation matrix of the input dataset.
It creates a heatmap of the correlation matrix. It also filters the protein
pairs with correlation values above the threshold and returns them in a tibble.
}
\examples{
# Prepare data
df <- example_data |>
  dplyr::select(DAid, Assay, NPX) |>
  tidyr::pivot_wider(names_from = "Assay", values_from = "NPX") |>
  dplyr::select(-DAid)

# Correlate proteins
results <- create_corr_heatmap(df, threshold = 0.7)

# Print results
results$cor_plot  # Heatmap of protein-protein correlations

results$cor_matrix[1:5, 1:5]  # Subset of the correlation matrix

results$cor_results  # Filtered protein pairs exceeding correlation threshold
}
