% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{create_corr_heatmap}
\alias{create_corr_heatmap}
\title{Create a correlation heatmap and report the protein-protein correlations above a threshold}
\usage{
create_corr_heatmap(
  x,
  y = NULL,
  use = "pairwise.complete.obs",
  method = "pearson",
  threshold = 0.8,
  cluster_rows = TRUE,
  cluster_cols = TRUE
)
}
\arguments{
\item{x}{(vector or tibble). A numeric vector, matrix or data frame.}

\item{y}{(vector, tibble or NULL). A numeric vector, matrix or data frame with compatible dimensions with \code{x}. Default is NULL}

\item{use}{(character). A character string. The method to use for computing correlations. Default is "pairwise.complete.obs"}

\item{method}{(character). A character string. The correlation method to use. Default is "pearson"}

\item{threshold}{(numeric). The reporting protein-protein correlation threshold. Default is 0.8}

\item{cluster_rows}{(logical). Whether to cluster the rows. Default is TRUE}

\item{cluster_cols}{(logical). Whether to cluster the columns. Default is TRUE}
}
\value{
A list containing the following elements:
\itemize{
\item cor_matrix (matrix). A matrix of protein-protein correlations
\item cor_results (tibble). A tibble with the filtered protein pairs and their correlation values
\item p (plot). A heatmap of protein-protein correlations
}
}
\description{
The function calculates the correlation matrix of the input dataframe.
It filters out the protein pairs with correlation values above or below the threshold.
It then creates a heatmap of the correlation matrix.
}
\examples{
df <- example_data |>
  dplyr::select(DAid, Assay, NPX) |>
  tidyr::pivot_wider(names_from = "Assay", values_from = "NPX") |>
  dplyr::select(-DAid)
cor_results <- create_corr_heatmap(df, threshold = 0.7)
}
