% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dir}
\alias{create_dir}
\title{Create directory w/o system date}
\usage{
create_dir(dir_name, date = FALSE)
}
\arguments{
\item{dir_name}{(string). The name of the directory to create}

\item{date}{(logical). If TRUE, the current date and time will be appended to the directory name}
}
\description{
The function creates a directory with the specified name. If the directory already exists, a message is printed.
}
\examples{
create_dir("my_directory", date = FALSE)
create_dir("my_directory/inner_dir", date = TRUE)
# Clean up the created directory
unlink("my_directory", recursive = TRUE)
}
