% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{do_gsea}
\alias{do_gsea}
\title{Perform gene set enrichment analysis}
\usage{
do_gsea(de_results, database = c("GO", "Reactome"), pval_lim = 0.05)
}
\arguments{
\item{de_results}{A tibble containing the results of a differential expression analysis.}

\item{database}{The database to perform the GSEA. It can be either "GO" or "Reactome".}

\item{pval_lim}{The p-value threshold to consider a term as significant.}
}
\value{
A list containing the results of the GSEA.
}
\description{
This function performs gene set enrichment analysis (GSEA) using the clusterProfiler package.
}
\details{
The ontology option used when database = "GO" is "ALL".
}
\examples{
# Run Differential Expression Analysis and extract results
control = c("BRC", "CLL", "CRC", "CVX", "ENDC", "GLIOM", "LUNGC", "LYMPH", "MYEL", "OVC", "PRC")
de_res <- do_limma(example_data,
                   example_metadata,
                   case = "AML",
                   control = control,
                   wide = FALSE)
de_results <- de_res$de_results

# Run GSEA with Reactome database
do_gsea(de_results,
        database = "GO",
        pval_lim = 0.9)
# Remember that the data is artificial, this is why we use an absurdly high p-value cutoff
}
