% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{do_limma}
\alias{do_limma}
\title{Run differential expression analysis with limma}
\usage{
do_limma(
  olink_data,
  metadata,
  variable = "Disease",
  case,
  control,
  correct = c("Sex", "Age"),
  correct_type = c("factor", "numeric"),
  wide = TRUE,
  only_female = NULL,
  only_male = NULL,
  volcano = TRUE,
  pval_lim = 0.05,
  logfc_lim = 0,
  top_up_prot = 40,
  top_down_prot = 10,
  palette = "diff_exp",
  report_nproteins = TRUE,
  subtitle = NULL,
  save = FALSE
)
}
\arguments{
\item{olink_data}{A tibble with the Olink data in wide format.}

\item{metadata}{A tibble with the metadata.}

\item{variable}{The variable of interest that includes the case and control groups.}

\item{case}{The case group.}

\item{control}{The control groups.}

\item{correct}{The variables to correct the results with. Default c("Sex", "Age").}

\item{correct_type}{The type of the variables to correct the results with. Default c("factor", "numeric", "numeric").}

\item{wide}{If the data is in wide format. Default is TRUE.}

\item{only_female}{The female specific diseases. Default is NULL.}

\item{only_male}{The male specific diseases. Default is NULL.}

\item{volcano}{Generate volcano plots. Default is TRUE.}

\item{pval_lim}{The p-value limit of significance. Default is 0.05.}

\item{logfc_lim}{The logFC limit of significance. Default is 0.}

\item{top_up_prot}{The number of top up regulated proteins to label on the plot. Default is 40.}

\item{top_down_prot}{The number of top down regulated proteins to label on the plot. Default is 10.}

\item{palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is "diff_exp".}

\item{report_nproteins}{If the number of significant proteins should be reported in the subtitle. Default is TRUE.}

\item{subtitle}{The subtitle of the plot or NULL for no subtitle.}

\item{save}{Save the volcano plots. Default is FALSE.}
}
\value{
A list with the differential expression results and volcano plots.
\itemize{
\item de_results: A list with the differential expression results.
\item volcano_plots: A list with the volcano plots.
}
}
\description{
\code{do_limma()} performs differential expression analysis using limma package.
It can correct the results for metadata columns like Sex, Age, or BMI.
The output tibble includes the logFC, p-values, as well as the FDR adjusted p-values.
The function removes the NAs from the columns that are used to correct for.
It can generate and save volcano plots.
}
\details{
For sex-specific diseases, there will be no correction for Sex.
This is performed automatically by the function.
}
\examples{
de_results <- do_limma(example_data,
                       example_metadata,
                       case = "AML",
                       control = c("CLL", "MYEL"),
                       wide = FALSE)

# Results for AML
de_results$de_results

# Volcano plot for AML
de_results$volcano_plot
}
