% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{do_limma_continuous}
\alias{do_limma_continuous}
\title{Run differential expression analysis with limma for continuous variable}
\usage{
do_limma_continuous(
  olink_data,
  metadata,
  variable,
  correct = c("Sex"),
  correct_type = c("factor"),
  wide = TRUE,
  volcano = TRUE,
  pval_lim = 0.05,
  logfc_lim = 0,
  top_up_prot = 40,
  top_down_prot = 10,
  palette = "diff_exp",
  subtitle = TRUE,
  save = FALSE
)
}
\arguments{
\item{olink_data}{(tibble). A tibble with the Olink data in wide format.}

\item{metadata}{(tibble). A tibble with the metadata.}

\item{variable}{(character). The variable of interest.}

\item{correct}{(character). The variables to correct the results with. Default is c("Sex").}

\item{correct_type}{(character). The type of the variables to correct the results with. Default is c("factor").}

\item{wide}{(logical). If the data is in wide format. Default is TRUE.}

\item{volcano}{(logical). Generate volcano plots. Default is TRUE.}

\item{pval_lim}{(numeric). The p-value limit for significance. Default is 0.05.}

\item{logfc_lim}{(numeric). The logFC limit for significance. Default is 0.}

\item{top_up_prot}{(numeric). The number of top up regulated proteins to label on the plot. Default is 40.}

\item{top_down_prot}{(numeric). The number of top down regulated proteins to label on the plot. Default is 10.}

\item{palette}{(character or vector). The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is "diff_exp".}

\item{subtitle}{(logical). If the subtitle should be displayed. Default is TRUE.}

\item{save}{(logical). Save the volcano plots. Default is FALSE.}
}
\value{
A list with the differential expression results and volcano plots.
\itemize{
\item de_results: A list with the differential expression results.
\item volcano_plots: A list with the volcano plots.
}
}
\description{
This function runs differential expression analysis using limma for a continuous variable.
It can generate and save volcano plots.
}
\examples{
do_limma_continuous(example_data, example_metadata, "Age", wide = FALSE)
}
