% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{do_limma_de}
\alias{do_limma_de}
\title{Differential expression analysis with limma}
\usage{
do_limma_de(
  join_data,
  variable = "Disease",
  case,
  correct = c("Sex", "Age"),
  correct_type = c("factor", "numeric", "numeric"),
  only_female = NULL,
  only_male = NULL,
  pval_lim = 0.05,
  logfc_lim = 0
)
}
\arguments{
\item{join_data}{A tibble with the Olink data in wide format joined with metadata.}

\item{variable}{The variable of interest that includes the case and control groups.}

\item{case}{The disease of interest.}

\item{correct}{The variables to correct the results with. Default c("Sex", "Age").}

\item{correct_type}{The type of the variables to correct the results with. Default c("factor", "numeric").}

\item{only_female}{The female specific diseases. Default is NULL.}

\item{only_male}{The male specific diseases. Default is NULL.}

\item{pval_lim}{The p-value limit of significance. Default is 0.05.}

\item{logfc_lim}{The logFC limit of significance. Default is 0.}
}
\value{
A tibble with the differential expression results.
}
\description{
\code{do_limma_de()} performs differential expression analysis using limma package.
It can correct the results for metadata columns like Sex, Age, or BMI.
The output tibble includes the logFC, p-values, as well as the FDR adjusted p-values.
The function removes the NAs from the columns that are used to correct for.
}
\keyword{internal}
