% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{do_ora}
\alias{do_ora}
\title{Perform over-representation analysis}
\usage{
do_ora(
  protein_list,
  database = c("KEGG", "GO", "Reactome"),
  background = NULL,
  pval_lim = 0.05
)
}
\arguments{
\item{protein_list}{A character vector containing the protein names.}

\item{database}{The database to perform the ORA. It can be either "KEGG", "GO", or "Reactome".}

\item{background}{A character vector containing the background genes.}

\item{pval_lim}{The p-value threshold to consider a term as significant.}
}
\value{
A list containing the results of the ORA.
}
\description{
\code{do_ora()} performs over-representation analysis (ORA) using the clusterProfiler package.
}
\details{
The ontology option used when database = "GO" is "BP" (Biological Process).
When Olink data is used, it is recommended to provide a protein list as background.
}
\examples{
# Perform Differential Expression Analysis
de_res <- do_limma(example_data, example_metadata, case = "AML", wide = FALSE)

# Extract the up-regulated proteins for AML
sig_up_proteins_aml <- de_res$de_results |>
  dplyr::filter(sig == "significant up") |>
  dplyr::pull(Assay)

# Perform ORA with GO database
do_ora(sig_up_proteins_aml, database = "GO")
}
