% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{do_ora}
\alias{do_ora}
\title{Perform over-representation analysis (ORA) using clusterProfiler}
\usage{
do_ora(gene_list, database = c("KEGG", "GO"))
}
\arguments{
\item{gene_list}{(vector). A list of gene names.}

\item{database}{(character). The database to perform the ORA. It can be either "KEGG" or "GO".}
}
\value{
enrichment (list). A list containing the results of the ORA.
}
\description{
Perform over-representation analysis (ORA) using clusterProfiler
}
\examples{
enrichment <- do_ora(c("TP53", "BRCA1", "BRCA2"), database = "KEGG")
}
