% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{do_ora}
\alias{do_ora}
\title{Perform over-representation analysis (ORA) using clusterProfiler}
\usage{
do_ora(gene_list, database = c("KEGG", "GO"))
}
\arguments{
\item{gene_list}{A character vector containing the gene names.}

\item{database}{The database to perform the ORA. It can be either "KEGG" or "GO".}
}
\value{
A list containing the results of the ORA.
}
\description{
\code{do_ora()} performs over-representation analysis (ORA) using the clusterProfiler package.
It also produces useful plots to visualize the results.
}
\examples{
# Perform Differential Expression Analysis
de_res <- do_limma(example_data, example_metadata, wide = FALSE)

# Extract the up-regulated proteins for AML
sig_up_proteins_aml <- de_res$de_results$AML |>
  dplyr::filter(sig == "significant up") |>
  dplyr::pull(Assay)

# Perform ORA with GO database
do_ora(sig_up_proteins_aml, database = "GO")
}
