% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{do_ttest}
\alias{do_ttest}
\title{Run differential expression analysis with t-test}
\usage{
do_ttest(
  olink_data,
  metadata,
  case,
  wide = TRUE,
  only_female = NULL,
  only_male = NULL,
  volcano = TRUE,
  pval_lim = 0.05,
  logfc_lim = 0,
  top_up_prot = 40,
  top_down_prot = 10,
  palette = "diff_exp",
  subtitle = TRUE,
  save = FALSE
)
}
\arguments{
\item{olink_data}{A tibble with the Olink data in wide format.}

\item{metadata}{A tibble with the metadata.}

\item{case}{The case group.}

\item{wide}{If the data is in wide format. Default is TRUE.}

\item{only_female}{The female specific diseases. Default is NULL.}

\item{only_male}{The male specific diseases. Default is NULL.}

\item{volcano}{Generate volcano plots. Default is TRUE.}

\item{pval_lim}{The p-value limit of significance. Default is 0.05.}

\item{logfc_lim}{The logFC limit of significance. Default is 0.}

\item{top_up_prot}{The number of top up regulated proteins to label on the plot. Default is 40.}

\item{top_down_prot}{The number of top down regulated proteins to label on the plot. Default is 10.}

\item{palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is "diff_exp".}

\item{subtitle}{If the subtitle should be displayed. Default is TRUE.}

\item{save}{Save the volcano plots. Default is FALSE.}
}
\value{
A list with the differential expression results and volcano plots.
\itemize{
\item de_results: A list with the differential expression results.
\item volcano_plots: A list with the volcano plots.
}
}
\description{
\code{do_ttest()} performs differential expression analysis using t-test.
It separates the data in case-control groups, checks for data normality and
perform a t-test or Wilcoxon test respectively. It also performs p value FDR adjustment.
It can generate and save volcano plots.
}
\examples{
de_results <- do_ttest(example_data, example_metadata, "AML", wide = FALSE)

# Results for AML
de_results$de_results

# Volcano plot for AML
de_results$volcano_plot
}
