% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{do_ttest_de}
\alias{do_ttest_de}
\title{Differential expression analysis with t-test}
\usage{
do_ttest_de(
  de_res,
  long_data,
  disease,
  assay,
  only_female = NULL,
  only_male = NULL
)
}
\arguments{
\item{de_res}{(matrix). A matrix to store the results}

\item{long_data}{(tibble). A tibble with the Olink data in long format.}

\item{disease}{(character). The disease of interest}

\item{assay}{(character). The assay of interest}

\item{only_female}{(character or vector). The female specific diseases. Default is NULL}

\item{only_male}{(character or vector). The male specific diseases. Default is NULL}
}
\value{
de_res(matrix). A matrix with the differential expression results
}
\description{
This function performs differential expression analysis using t-test.
}
\examples{
de_res <- matrix(nrow=0, ncol=4)
colnames(de_res) <- c("Assay", "P.Value", "logFC", "Disease")
test_data <- example_data |>
  dplyr::select(DAid, Assay, NPX) |>
  dplyr::left_join(example_metadata |>
                     dplyr::select(dplyr::any_of(c("DAid", "Disease"))),
                   by = "DAid")

do_ttest_de(de_res, test_data, "AML", "ABL1")
}
