% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{do_ttest_de}
\alias{do_ttest_de}
\title{Differential expression analysis with t-test}
\usage{
do_ttest_de(
  long_data,
  variable = "Disease",
  case,
  assays,
  normality_res,
  only_female = NULL,
  only_male = NULL,
  pval_lim = 0.05,
  logfc_lim = 0
)
}
\arguments{
\item{long_data}{A tibble with the Olink data in long format and the Sex column from metadata.}

\item{variable}{The variable of interest that includes the case and control groups.}

\item{case}{The disease of interest.}

\item{assays}{The assays to run the differential expression analysis on.}

\item{only_female}{The female specific diseases. Default is NULL.}

\item{only_male}{The male specific diseases. Default is NULL.}

\item{pval_lim}{The p-value limit of significance. Default is 0.05.}

\item{logfc_lim}{The logFC limit of significance. Default is 0.}
}
\value{
A tibble with the differential expression results.
}
\description{
\code{do_ttest_de()} performs differential expression analysis using t-test.
It separates the data in case-control groups, checks for data normality and
perform a t-test or Wilcoxon test respectively. It also performs p value FDR adjustment.
}
\keyword{internal}
