% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{elnet_finalfit}
\alias{elnet_finalfit}
\title{Fit the final elastic net model}
\usage{
elnet_finalfit(train_set, tune_res, wf, seed = 123)
}
\arguments{
\item{train_set}{(tibble). Training set.}

\item{tune_res}{(tibble). Hyperparameter optimization results.}

\item{wf}{(workflow). Workflow object.}

\item{seed}{(numeric). Seed for reproducibility. Default is 123.}
}
\value{
A list with three elements:
\itemize{
\item final_elnet (parsnip model). Final elastic net model.
\item best_elnet (tibble). Best hyperparameters from hyperparameter optimization.
\item final_wf (workflow). Final workflow object.
}
}
\description{
This function fits the final elastic net model using the best hyperparameters from hyperparameter optimization.
}
\examples{
wide_data <- example_data |>
  dplyr::select(DAid, Assay, NPX) |>
  tidyr::pivot_wider(names_from = Assay, values_from = NPX)
join_data <- wide_data |>
  dplyr::left_join(example_metadata |> dplyr::select(DAid, Disease, Sex))
diseases <- unique(example_metadata$Disease)
data_split <- split_data(join_data)
train_list <- make_groups(data_split$train_set,
                          diseases,
                          only_female = c("BRC", "CVX", "ENDC", "OVC"),
                          only_male = "PRC")

test_list <- make_groups(data_split$test_set,
                         diseases,
                         only_female = c("BRC", "CVX", "ENDC", "OVC"),
                         only_male = "PRC")

hypopt_res <- elnet_hypopt(train_list,
                           test_list,
                           "AML",
                           type = "elnet",
                           cv_sets = 2,
                           grid_size = 1,
                           ncores = 1,
                           exclude_cols = "Sex")

finalfit_res <- elnet_finalfit(hypopt_res$train_set, hypopt_res$elnet_tune, hypopt_res$wf)
}
