% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{elnet_hypopt}
\alias{elnet_hypopt}
\title{Hyperparameter optimization for elastic net}
\usage{
elnet_hypopt(
  train_data,
  exclude_cols,
  disease,
  type = "lasso",
  metric = roc_auc,
  cv_sets = 5,
  grid_size = 10,
  ncores = 4,
  seed = 123
)
}
\arguments{
\item{train_data}{(list). List of training data sets from make_groups().}

\item{exclude_cols}{(vector). Columns to exclude from the model.}

\item{disease}{(character). Disease to predict.}

\item{type}{(character). Type of regularization. Default is "lasso". Other options are "ridge" and "elnet".}

\item{metric}{(function). Metric to optimize. Default is roc_auc.}

\item{cv_sets}{(numeric). Number of cross-validation sets. Default is 5.}

\item{grid_size}{(numeric). Size of the grid for hyperparameter optimization. Default is 10.}

\item{ncores}{(numeric). Number of cores to use for parallel processing. Default is 4.}

\item{seed}{(numeric). Seed for reproducibility. Default is 123.}
}
\value{
A list with three elements:
\itemize{
\item elnet_tune (tibble). Hyperparameter optimization results.
\item train_set (tibble). Training set.
\item wf (workflow). Workflow object.
}
}
\description{
This function performs hyperparameter optimization for elastic net models.
It uses the glmnet engine for logistic regression and tunes either only penalty or both penalty and mixture.
}
\examples{
wide_data <- example_data |>
  dplyr::select(DAid, Assay, NPX) |>
  tidyr::pivot_wider(names_from = Assay, values_from = NPX)
join_data <- wide_data |>
  dplyr::left_join(example_metadata |> dplyr::select(DAid, Disease, Sex))
diseases <- unique(example_metadata$Disease)
group_list <- make_groups(join_data,
                          diseases,
                          only_female = c("BRC", "CVX", "ENDC", "OVC"),
                          only_male = "PRC")

hypopt_res <- elnet_hypopt(group_list, "Sex", "AML", type = "elnet", grid_size = 5, ncores = 1)
}
