% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{example_metadata}
\alias{example_metadata}
\title{Example cancer metadata}
\format{
\subsection{\code{example_metadata}}{

A data frame with 586 rows and 9 columns:
\describe{
\item{DAid}{The Disease Atlas sample ID}
\item{Sample}{The Sample ID}
\item{GROUP}{The cancer type}
\item{Stage}{The cancer stage}
\item{Grade}{The cancer grade}
\item{Sex}{The patient sex}
\item{Age}{The patient age}
\item{BMI}{The patient BMI}
\item{Cohort}{The cohort the patient belongs to}
}
}
}
\source{
\url{https://github.com/buenoalvezm/Pan-cancer-profiling/blob/main/data/cancer_metadata_synthetic.rds}
}
\usage{
example_metadata
}
\description{
A subset of metadata from a synthetic cancer dataset. With DAid, Age, BMI, and Cohort  as extra
columns. The original dataset was processed with the \code{process_example_metadata} script.
}
\keyword{datasets}
