% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{extract_protein_list}
\alias{extract_protein_list}
\title{Extract protein lists from the upset data}
\usage{
extract_protein_list(upset_data, protein_lists)
}
\arguments{
\item{upset_data}{A tibble with the upset data.}

\item{protein_lists}{A list with the protein lists for each disease.}
}
\value{
A list with the proteins for each combination of diseases.
}
\description{
\code{extract_protein_list()} extracts the protein lists from the upset data.
It creates a list with the proteins for each combination of diseases.
}
\keyword{internal}
