% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{filter_sex_specific_disease}
\alias{filter_sex_specific_disease}
\title{Filter data in case of sex specific diseases}
\usage{
filter_sex_specific_disease(
  control_data,
  disease,
  diseases,
  only_female = NULL,
  only_male = NULL
)
}
\arguments{
\item{control_data}{(tibble). Control data to be filtered.}

\item{disease}{(character). Disease to be filtered.}

\item{diseases}{(vector). Vector of diseases.}

\item{only_female}{(vector). Vector of diseases that are female specific.}

\item{only_male}{(vector). Vector of diseases that are male specific.}
}
\value{
A list with two elements:
\itemize{
\item control_data (tibble). Filtered control data.
\item diseases_subset (vector). Filtered diseases vector.
}
}
\description{
This function filters the control data based on the disease.
It also filters the diseases vector to keep only the diseases that we will pick samples from.
}
\examples{
wide_data <- example_data |>
  dplyr::select(DAid, Assay, NPX) |>
  tidyr::pivot_wider(names_from = Assay, values_from = NPX)
join_data <- wide_data |>
  dplyr::left_join(example_metadata |> dplyr::select(DAid, Disease, Sex))

diseases <- unique(example_metadata$Disease)
control_data <- join_data |> dplyr::filter(Disease != "BRC")

filter_sex_specific_disease(control_data,
                            "BRC",
                            diseases,
                            only_female = c("BRC", "CVX", "ENDC", "OVC"),
                            only_male = "PRC")
}
