% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{filter_sex_specific_disease}
\alias{filter_sex_specific_disease}
\title{Create control groups for sex specific cases}
\usage{
filter_sex_specific_disease(
  control_data,
  case,
  cases,
  only_female = NULL,
  only_male = NULL
)
}
\arguments{
\item{control_data}{Control data to filter.}

\item{case}{Case to create control group.}

\item{cases}{Cases.}

\item{only_female}{Cases that are female specific.}

\item{only_male}{Cases that are male specific.}
}
\value{
A list with two elements:
\itemize{
\item control_data: Control data to filter.
\item cases_subset: Filtered cases vector.
}
}
\description{
\code{filter_sex_specific_disease()} creates control groups for sex-specific cases
by filtering samples to include only those from the relevant sex and by
updating the cases vector to include only the cases that will be sampled.
}
\keyword{internal}
