% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{finalfit}
\alias{finalfit}
\title{Fit the best model}
\usage{
finalfit(train_set, tune_res, wf, seed = 123)
}
\arguments{
\item{train_set}{Training set.}

\item{tune_res}{Hyperparameter optimization results.}

\item{wf}{Workflow object.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A list with three elements:
\itemize{
\item final_elnet: Final model.
\item best_elnet: Best hyperparameters from hyperparameter optimization.
\item final_wf: Final workflow object.
}
}
\description{
\code{finalfit()} fits the model that performed the best in hyperparameter optimization.
}
\keyword{internal}
