% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{generate_df}
\alias{generate_df}
\title{Create and save wide and join dataframes}
\usage{
generate_df(
  long_data,
  metadata = NULL,
  join = TRUE,
  metadata_cols = c("DAid", "Disease", "Sex", "Age", "BMI"),
  save = TRUE
)
}
\arguments{
\item{long_data}{The long data.}

\item{metadata}{The metadata.}

\item{join}{If TRUE, the dataframes are joined with metadata.}

\item{metadata_cols}{The metadata columns to join with data.}

\item{save}{If TRUE, the dataframes are saved.}
}
\value{
A list containing the following elements:
\itemize{
\item wide_data: The wide data.
\item join_data: The joined data with metadata. If join is FALSE, this is not returned.
}
}
\description{
\code{generate_df()} creates wide and join dataframes from the long data and metadata.
It saves them in the data/processed/data_metadata directory in RDA format.
}
\details{
It will filter out rows with missing values in \code{Disease}, as a results of
mismatching between data and metadata
}
\examples{
generate_df(example_data, example_metadata, save = FALSE)
}
