% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{generate_subtitle}
\alias{generate_subtitle}
\title{Create subtitle for variable importance plot}
\usage{
generate_subtitle(
  features,
  accuracy,
  sensitivity,
  specificity,
  auc,
  mixture,
  subtitle = c("accuracy", "sensitivity", "specificity", "auc", "features",
    "top-features", "mixture")
)
}
\arguments{
\item{features}{A tibble with features and their model importance.}

\item{accuracy}{Accuracy of the model.}

\item{sensitivity}{Sensitivity of the model.}

\item{specificity}{Specificity of the model.}

\item{auc}{AUC of the model.}

\item{mixture}{Mixture of lasso and ridge regularization. In random forest models it is NULL.}

\item{subtitle}{Vector of subtitle elements to include in the plot.}
}
\value{
The plot subtitle as character vector.
}
\description{
\code{generate_subtitle()} generates a subtitle for the variable importance plot.
}
\keyword{internal}
