% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{generate_subtitle}
\alias{generate_subtitle}
\title{Generate subtitle for variable importance plot}
\usage{
generate_subtitle(
  features,
  accuracy,
  sensitivity,
  specificity,
  auc,
  mixture,
  subtitle = NULL
)
}
\arguments{
\item{features}{(tibble). Features with importance values.}

\item{accuracy}{(numeric). Accuracy of the model.}

\item{sensitivity}{(numeric). Sensitivity of the model.}

\item{specificity}{(numeric). Specificity of the model.}

\item{auc}{(numeric). AUC of the model.}

\item{mixture}{(numeric). Mixture of lasso and ridge regularization.}

\item{subtitle}{(vector). Vector of subtitles to include in the plot. Default is all.}
}
\value{
subtitle (character). Subtitle for the plot.
}
\description{
This function generates a subtitle for the variable importance plot.
}
\keyword{internal}
