% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{get_hpa_palettes}
\alias{get_hpa_palettes}
\title{HPA color palettes}
\usage{
get_hpa_palettes()
}
\value{
List of HPA color palettes.
}
\description{
\code{get_hpa_palettes()} returns a list of color palettes used by the Human Protein Atlas (HPA) project.
}
\examples{
get_hpa_palettes()
}
