% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hd_bin_columns}
\alias{hd_bin_columns}
\title{Bin columns}
\usage{
hd_bin_columns(dat, column_types, bins = 5, round_digits = 0)
}
\arguments{
\item{dat}{The data to bin.}

\item{column_types}{A vector containing the type of each column in the dataframe.}

\item{bins}{The number of bins to create. Default is 5.}

\item{round_digits}{The number of digits to round the bin ranges to. Default is 0.}
}
\value{
The data with binned continuous columns.
}
\description{
\code{hd_bin_columns()} bins continuous columns and labels them with ranges.
}
\examples{
# Example dataframe
test_data <- data.frame(
  age = c(25, 35, 45, 55, 65),
  BMI = c(25, 35, 30, 32, 28),
  sex = c("M", "F", "M", "F", "M")
)

column_types <- c("continuous", "continuous", "categorical")
hd_bin_columns(test_data, column_types, bins = 3)
}
