% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{hd_correlate}
\alias{hd_correlate}
\title{Correlate data}
\usage{
hd_correlate(x, y = NULL, use = "pairwise.complete.obs", method = "pearson")
}
\arguments{
\item{x}{A numeric vector, matrix or tibble.}

\item{y}{A numeric vector, matrix or tibble with compatible dimensions with \code{x}. Default is NULL.}

\item{use}{A character string. The method to use for computing correlations. Default is "pairwise.complete.obs". Other options are "everything", "all.obs", "complete.obs", or "na.or.complete".}

\item{method}{A character string. The correlation method to use. Default is "pearson". Other options are "kendall" or "spearman".}
}
\value{
A matrix of protein-protein correlations.
}
\description{
\code{hd_correlate()} calculates the correlation matrix of the input dataset.
}
\examples{
# Correlate features in a dataset (column wise)
dat <- example_data |>
 dplyr::select(DAid, Assay, NPX) |>
 tidyr::pivot_wider(names_from = "Assay", values_from = "NPX") |>
 dplyr::select(-DAid)

hd_correlate(dat)[1:5, 1:5]  # Subset of the correlation matrix

# Correlate 2 vectors
vec1 <- c(1, 2, 3, 4, 5)
vec2 <- c(5, 4, 3, 2, 1)
hd_correlate(vec1, vec2)
}
