% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{hd_de_limma}
\alias{hd_de_limma}
\title{Differential expression analysis with limma}
\usage{
hd_de_limma(
  dat,
  metadata = NULL,
  variable = "Disease",
  case,
  control = NULL,
  correct = NULL
)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format and sample_id as its first column.}

\item{metadata}{A dataset containing the metadata information with the sample ID as the first column. If a HDAnalyzeR object is provided, this parameter is not needed.}

\item{variable}{The name of the column containing the case and control groups or a continuous variable.}

\item{case}{The case group. In case of a continuous variable, it must be NULL.}

\item{control}{The control groups. If NULL, it will be set to all other unique values of the variable that are not the case. In case of a continuous variable, it must be NULL.}

\item{correct}{The variables to correct the results with. Default is NULL.}
}
\value{
An object with the DE results.
}
\description{
\code{hd_de_limma()} performs differential expression analysis using limma package.
It can correct the results for metadata columns like Sex, Age, or BMI.
The output tibble includes the logFC, p-values, as well as the FDR adjusted p-values.
The function removes the NAs from the columns that are used to correct for.
}
\details{
In case of a continuous variable or if you are correcting based on a continuous variable,
the variable should be numeric and contain at least 6 unique variables.
}
\examples{
# Initialize an HDAnalyzeR object
hd_object <- hd_initialize(example_data, example_metadata)

# Run differential expression analysis for AML vs all others
hd_de_limma(hd_object, case = "AML")

# Run differential expression analysis for AML vs CLL and MYEL and correct for metadata variables
hd_de_limma(hd_object,
                case = "AML",
                control = c("CLL", "MYEL"),
                correct = c("Sex", "Age", "BMI"))

# Run differential expression analysis for continuous variable
hd_de_limma(hd_object, variable = "Age", case = NULL, correct = c("Sex"))
}
