% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{hd_de_ttest}
\alias{hd_de_ttest}
\title{Differential expression analysis with t-test}
\usage{
hd_de_ttest(
  dat,
  metadata = NULL,
  variable = "Disease",
  case,
  control = NULL,
  log_transform = FALSE
)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format and sample ID as its first column.}

\item{metadata}{A dataset containing the metadata information with the sample ID as the first column. If a HDAnalyzeR object is provided, this parameter is not needed.}

\item{variable}{The name of the metadata variable containing the case and control groups.}

\item{case}{The case group.}

\item{control}{The control groups. If NULL, it will be set to all other unique values of the variable that are not the case.}

\item{log_transform}{If the data should be log transformed. Default is FALSE.}
}
\value{
An object with the DE results.
}
\description{
\code{hd_de_ttest()} performs differential expression analysis using t-test.
The output tibble includes the logFC, p-values, as well as the FDR adjusted p-values.
The function removes the rows with NAs in the variables that are used to correct for as
well as the variable of interest.
}
\details{
The variable of interest should be categorical and present in the metadata.
In case your data are not already log transformed, you can set \code{log_transform = TRUE} to log
transform the data with base 2 before the analysis start.

When using this function you cannot correct for other variables or run it against a continuous variable.
If you need to correct for other variables or run the analysis against a continuous variable, use \code{hd_de_limma()}.
}
\examples{
# Initialize an HDAnalyzeR object
hd_object <- hd_initialize(example_data, example_metadata)

# Run differential expression analysis for AML vs all others
hd_de_ttest(hd_object, case = "AML")

# Run differential expression analysis for AML vs CLL
hd_de_ttest(hd_object, case = "AML", control = "CLL")
}
