% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hd_filter_by_sex}
\alias{hd_filter_by_sex}
\title{Filter data and metadata by sex}
\usage{
hd_filter_by_sex(dat, metadata = NULL, variable = "Sex", sex)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format and sample_id as its first column.}

\item{metadata}{A dataset containing the metadata information with the sample ID as the first column. If a HDAnalyzeR object is provided, this parameter is not needed.}

\item{variable}{The name of the variable in the metadata contain information for sex. Default is "Sex".}

\item{sex}{The value of the sex variable to filter by.}
}
\value{
A list containing the filtered data and metadata.
}
\description{
Filter data and metadata by sex
}
\examples{
# Create the HDAnalyzeR object providing the data and metadata
hd_object <- hd_initialize(example_data, example_metadata)

hd_filter_by_sex(hd_object, variable = "Sex", sex = "F")
}
