% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{hd_gsea}
\alias{hd_gsea}
\title{Gene set enrichment analysis}
\usage{
hd_gsea(
  de_results,
  database = c("GO", "Reactome", "KEGG"),
  ontology = c("BP", "CC", "MF", "ALL"),
  ranked_by = "logFC",
  pval_lim = 0.05
)
}
\arguments{
\item{de_results}{An \code{hd_de} object from \code{hd_de_limma()} or \code{hd_de_ttest()} or a tibble containing the results of a differential expression analysis.}

\item{database}{The database to perform the ORA. It can be either "GO", "KEGG", or "Reactome".}

\item{ontology}{The ontology to use when database = "GO". It can be "BP" (Biological Process), "CC" (Cellular Component), "MF" (Molecular Function), or "ALL". In the case of KEGG and Reactome, this parameter is ignored.}

\item{ranked_by}{The variable to rank the proteins. It can be "logFC", "both" which is the product of logFC and -log(adj.P.Val) or a custom sorting variable. It should be however a column in the DE results tibble (\code{de_results} argument).}

\item{pval_lim}{The p-value threshold to consider a term as significant in the enrichment analysis. Default is 0.05.}
}
\value{
A list containing the results of the GSEA.
}
\description{
\code{hd_gsea()} performs gene set enrichment analysis (GSEA) using the clusterProfiler package.
}
\details{
To perform the GSEA, \code{clusterProfiler} package is used. For more information, please
refer to the \code{clusterProfiler} documentation.

If you want to learn more about GSEA, please refer to the following publications:
\itemize{
\item Chicco D, Agapito G. Nine quick tips for pathway enrichment analysis. PLoS Comput Biol. 2022 Aug 11;18(8):e1010348. doi: 10.1371/journal.pcbi.1010348. PMID: 35951505; PMCID: PMC9371296. https://pmc.ncbi.nlm.nih.gov/articles/PMC9371296/
\item https://yulab-smu.top/biomedical-knowledge-mining-book/enrichment-overview.html#gsea-algorithm
}
}
\examples{
# Initialize an HDAnalyzeR object
hd_object <- hd_initialize(example_data, example_metadata)

# Run differential expression analysis for AML vs all others
de_results <- hd_de_limma(hd_object, case = "AML")

# Run GSEA with GO database
hd_gsea(de_results,
        database = "GO",
        ontology = "BP",
        ranked_by = "logFC",
        pval_lim = 0.9)
# Remember that the data is artificial, this is why we use an absurdly high p-value cutoff

# Run GSEA with different ranking variable
hd_gsea(de_results,
        database = "GO",
        ontology = "BP",
        ranked_by = "both",
        pval_lim = 0.9)
}
