% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{hd_impute_missForest}
\alias{hd_impute_missForest}
\title{Impute via missForest}
\usage{
hd_impute_missForest(dat, maxiter = 10, ntree = 100, verbose = TRUE)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format and sample_id as its first column.}

\item{maxiter}{The maximum number of iterations.}

\item{ntree}{The number of trees to grow.}

\item{verbose}{If TRUE, the percentage of missing values in each column is displayed.}
}
\value{
The imputed dataset.
}
\description{
\code{impute_missForest()} imputes missing values in a dataset using the \code{missForest} method.
It allows the user to exclude certain columns from imputation and can also display the
percentage of missing values in each column before imputation.
}
\details{
This is the slowest and more complex imputation method. If KNN works fine, it is
recommended to use it instead of \code{missForest}.
}
\examples{
# Create the HDAnalyzeR object providing the data and metadata
hd_object <- hd_initialize(example_data, example_metadata)

# Data after imputation
res <- hd_impute_missForest(hd_object, maxiter = 1, ntree = 50)
res$data
}
