% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hd_initialize}
\alias{hd_initialize}
\title{Initialize HDAnalyzeR object}
\usage{
hd_initialize(
  dat,
  metadata = NULL,
  is_wide = FALSE,
  sample_id = "DAid",
  var_name = "Assay",
  value_name = "NPX"
)
}
\arguments{
\item{dat}{A tibble containing the omics data.}

\item{metadata}{A tibble containing the metadata. Default is NULL.}

\item{is_wide}{A logical indicating if the data is in wide format. Default is FALSE.}

\item{sample_id}{The name of the column containing the sample IDs. Default is "DAid".}

\item{var_name}{The name of the column containing the variable names. Default is "Assay".}

\item{value_name}{The name of the column containing the values. Default is "NPX".}
}
\value{
An HDAnalyzeR object.
}
\description{
\code{hd_initialize()} initializes an HDAnalyzeR object with the data, metadata, and other parameters.
}
\examples{
# Initialize an HDAnalyzeR object with long format data
hd_initialize(example_data, example_metadata)
}
