% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{hd_normalize}
\alias{hd_normalize}
\title{Normalize data and remove batch effects}
\usage{
hd_normalize(
  dat,
  metadata = NULL,
  center = TRUE,
  scale = TRUE,
  batch = NULL,
  batch2 = NULL
)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format and sample_id as its first column.}

\item{metadata}{A dataset containing the metadata information. If a HDAnalyzeR object is provided, this parameter is not needed.}

\item{center}{A logical value indicating whether to center the data. Default is TRUE.}

\item{scale}{A logical value indicating whether to scale the data. Default is TRUE.}

\item{batch}{The metadata column containing the batch information. In order to correct for batch effects, this parameter should be provided. Default is NULL.}

\item{batch2}{The metadata column containing the second batch information if available. Default is NULL.}
}
\value{
The normalized dataset.
}
\description{
\code{hd_normalize()} normalizes the data by scaling them and removing their batch effects.
It first converts the data to wide format if they are not already. It then removes
the batch effects and scales or centers the data. To remove batch effects, it uses the
\code{remove_batch_effects()}, that utilizes limma package. For scaling, it uses the \code{scale()}
from base R.
}
\details{
You can read more about the scaling and centering methods in the documentation of the \code{scale()} function in the \code{base} package,
as well as about the method for removing batch effects in the documentation of the \code{removeBatchEffect()} function in the \code{limma} package.
}
\examples{
# Create the HDAnalyzeR object providing the data and metadata
hd_object <- hd_initialize(example_data, example_metadata)

# Center data
hd_normalize(hd_object, center = TRUE, scale = FALSE)

# Center and scale data (z-score scaling)
hd_normalize(hd_object, center = TRUE, scale = TRUE)

# Center, scale and remove batch effects
hd_normalize(hd_object, batch = "Cohort")
}
