% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{hd_palettes}
\alias{hd_palettes}
\title{HDAnalyzeR color palettes}
\usage{
hd_palettes()
}
\value{
List of HPA color palettes.
}
\description{
\code{hd_palettes()} returns a list of color palettes used by the Human Protein Atlas (HPA) project.
}
\keyword{internal}
