% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{hd_plot_cor_heatmap}
\alias{hd_plot_cor_heatmap}
\title{Plot correlation heatmap}
\usage{
hd_plot_cor_heatmap(
  x,
  y = NULL,
  use = "pairwise.complete.obs",
  method = "pearson",
  threshold = 0.8,
  cluster_rows = TRUE,
  cluster_cols = TRUE
)
}
\arguments{
\item{x}{A numeric vector, matrix or data frame.}

\item{y}{A numeric vector, matrix or data frame with compatible dimensions with \code{x}. Default is NULL.}

\item{use}{A character string. The method to use for computing correlations.
Default is "pairwise.complete.obs". Other options are "everything", "all.obs", "
complete.obs", or "na.or.complete".}

\item{method}{A character string. The correlation method to use.
Default is "pearson". Other options are "kendall" or "spearman".}

\item{threshold}{The reporting correlation threshold. Default is 0.8.}

\item{cluster_rows}{Whether to cluster the rows. Default is TRUE.}

\item{cluster_cols}{Whether to cluster the columns. Default is TRUE.}
}
\value{
A list with the correlation matrix, the filtered pairs and their correlation values, and the heatmap.
}
\description{
\code{hd_plot_cor_heatmap()} calculates the correlation matrix of the input dataset.
It creates a heatmap of the correlation matrix. This matrix is created via \code{hd_correlate()}.
It also filters the feature pairs with correlation values above the threshold and
returns them in a tibble.
}
\examples{
# Prepare data
dat <- example_data |>
  dplyr::select(DAid, Assay, NPX) |>
  tidyr::pivot_wider(names_from = "Assay", values_from = "NPX") |>
  dplyr::select(-DAid)

# Correlate proteins
results <- hd_plot_cor_heatmap(dat, threshold = 0.7)

# Print results
results$cor_matrix[seq_len(5), seq_len(5)]  # Subset of the correlation matrix

results$cor_results  # Filtered protein pairs exceeding correlation threshold

results$cor_heatmap  # Heatmap of protein-protein correlations
}
