% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{hd_plot_de_summary}
\alias{hd_plot_de_summary}
\title{Plot summary visualizations for the differential expression results}
\usage{
hd_plot_de_summary(
  de_results,
  variable = "Disease",
  class_palette = NULL,
  diff_exp_palette = "diff_exp",
  pval_lim = 0.05,
  logfc_lim = 0
)
}
\arguments{
\item{de_results}{A list of differential expression results. It should be a list of objects created by \code{hd_run_de_limma()} with the classes as names. See the examples for more details.}

\item{variable}{The name of the column containing the case and control groups.}

\item{class_palette}{The color palette for the classes. If it is a character, it should be one of the palettes from \code{hd_palettes()}. Default is NULL.}

\item{diff_exp_palette}{The color palette for the differential expression. If it is a character, it should be one of the palettes from \code{hd_palettes()}. Default is "diff_exp".}

\item{pval_lim}{The p-value limit for significance. Default is 0.05.}

\item{logfc_lim}{The logFC limit for significance. Default is 0.}
}
\value{
A list with the DE summary plots and results
}
\description{
\code{hd_plot_de_summary()} creates summary visualizations for the differential expression results.
It plots a barplot with the number of significant proteins for each disease.
It also creates upset plots both for the significant up and down regulated proteins for each disease.
}
\examples{
# Initialize an HDAnalyzeR object
hd_object <- hd_initialize(example_data, example_metadata)

# Run differential expression analysis for AML vs all others
de_results_aml <- hd_run_de_limma(hd_object, case = "AML", control)
de_results_lungc <- hd_run_de_limma(hd_object, case = "LUNGC")
de_results_cll <- hd_run_de_limma(hd_object, case = "CLL")
de_results_myel <- hd_run_de_limma(hd_object, case = "MYEL")
de_results_gliom <- hd_run_de_limma(hd_object, case = "GLIOM")

res <- list("AML" = de_results_aml,
            "LUNGC" = de_results_lungc,
            "CLL" = de_results_cll,
            "MYEL" = de_results_myel,
            "GLIOM" = de_results_gliom)

# Plot summary visualizations
hd_plot_de_summary(res, class_palette = "cancers12")
}
