% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{hd_plot_dim}
\alias{hd_plot_dim}
\title{Plot PCA or UMAP results on a 2D plane}
\usage{
hd_plot_dim(
  dim_object,
  hd_object,
  x,
  y,
  color = NULL,
  palette = NULL,
  plot_loadings = NULL,
  nloadings = 5,
  axis_variance = TRUE
)
}
\arguments{
\item{dim_object}{A PCA or UMAP object containing the results of the dimensionality reduction analysis. Created by \code{hd_run_pca()} or \code{hd_run_umap()}.}

\item{hd_object}{A HDAnalyzeR object containing the data and metadata. Created by \code{hd_initialize()}.}

\item{x}{The name of the column in \code{dim_res} that contains the x-axis values.}

\item{y}{The name of the column in \code{dim_res} that contains the y-axis values.}

\item{color}{The name of the column in \code{dim_res} that contains the variable to be used to plot the points color.}

\item{palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}.}

\item{plot_loadings}{The component to be plotted. Default is NULL.}

\item{nloadings}{The number of loadings to be plotted. Default is 5.}

\item{axis_variance}{If TRUE, the explained variance of the components is added to the axis labels. Default is TRUE.}
}
\value{
A ggplot object
}
\description{
\code{hd_plot_dim()} plots the sample data points on a 2D plane.
The points can be plotted in the PCx/PCy or UMAP1/UMAP2 space.
}
\examples{
# Create the HDAnalyzeR object providing the data and metadata
hd_object <- hd_initialize(example_data, example_metadata)

# Run the PCA analysis
pca_object <- hd_run_pca(hd_object, components = 5, by_sample = TRUE, seed = 123) |>
  hd_plot_dim(hd_object, x = "PC1", y = "PC2", color = "Disease", palette = "cancers12")
}
