% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{hd_plot_model_summary}
\alias{hd_plot_model_summary}
\title{Plot features summary visualizations}
\usage{
hd_plot_model_summary(
  model_results,
  importance = 0.5,
  class_palette = NULL,
  upset_top_features = FALSE
)
}
\arguments{
\item{model_results}{A list of binary classification model results. It should be a list of objects created by \code{hd_run_rreg()}, \code{hd_run_rf()} or \code{hd_run_lr()} with the classes as names. See the examples for more details.}

\item{importance}{The importance threshold to consider a feature as top. Default is 0.5.}

\item{class_palette}{The color palette for the classes. If it is a character, it should be one of the palettes from \code{hd_palettes()}. Default is NULL.}

\item{upset_top_features}{Whether to plot the upset plot for the top features or all features. Default is FALSE (all features).}
}
\value{
A list with the binary classification model summary plots and results
}
\description{
\code{plot_features_summary()} plots the number of proteins and the number of top
proteins for each disease in a barplot. It also plots the upset plot of the
top or all protein features, as well as a summary line plot of the model
performance metrics.
}
\examples{
# Initialize an HDAnalyzeR object with only a subset of the predictors
hd_object <- hd_initialize(example_data, example_metadata)

# Split the data into training and test sets
hd_split <- hd_run_data_split(hd_object, variable = "Disease")

# Run the regularized regression model pipeline
model_results_aml <- hd_run_rreg(hd_split,
                                 variable = "Disease",
                                 case = "AML",
                                 grid_size = 2,
                                 cv_sets = 2)

model_results_cll <- hd_run_rreg(hd_split,
                                 variable = "Disease",
                                 case = "CLL",
                                 grid_size = 2,
                                 cv_sets = 2)

model_results_myel <- hd_run_rreg(hd_split,
                                  variable = "Disease",
                                  case = "MYEL",
                                  grid_size = 2,
                                  cv_sets = 2)

model_results_lungc <- hd_run_rreg(hd_split,
                                   variable = "Disease",
                                   case = "LUNGC",
                                   grid_size = 2,
                                   cv_sets = 2)

model_results_gliom <- hd_run_rreg(hd_split,
                                   variable = "Disease",
                                   case = "GLIOM",
                                   grid_size = 2,
                                   cv_sets = 2)

res <- list("AML" = model_results_aml,
            "LUNGC" = model_results_lungc,
            "CLL" = model_results_cll,
            "MYEL" = model_results_myel,
            "GLIOM" = model_results_gliom)

# Plot summary visualizations
hd_plot_model_summary(res, class_palette = "cancers12")
}
