% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{hd_plot_ora}
\alias{hd_plot_ora}
\title{Plot over-representation analysis results}
\usage{
hd_plot_ora(enrichment, seed = 123)
}
\arguments{
\item{enrichment}{The enrichment results obtained from \code{hd_ora()}.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
The input object enriched with the plots.
}
\description{
\code{hd_plot_ora()} generates useful visualizations for the results of the
over-representation analysis.
}
\details{
When KEGG database is used, a cnetplot is generated with ENTREZIDs instead of gene names.
For GO and Reactome databases the ENTREZIDs are converted to gene names.
If you get the "grid.Call(C_convert, x, as.integer(whatfrom), as.integer(whatto),  :
Viewport has zero dimension(s)" warning or error, try to increase the RStudio's viewer window size.
}
\examples{
# Initialize an HDAnalyzeR object
hd_object <- hd_initialize(example_data, example_metadata)

# Run differential expression analysis for AML vs all others
de_results <- hd_de_limma(hd_object, case = "AML")

# Extract the up-regulated proteins for AML
sig_up_proteins_aml <- de_results$de_res |>
  dplyr::filter(adj.P.Val < 0.05 & logFC > 0.5) |>
  dplyr::pull(Feature)

# Perform ORA with `GO` database and `BP` ontology
enrichment <- hd_ora(sig_up_proteins_aml, database = "GO", ontology = "BP")

# Plot the results
enrichment <- hd_plot_ora(enrichment)

# Access the plots
enrichment$dotplot
enrichment$treeplot
enrichment$cnetplot
}
