% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{hd_umap}
\alias{hd_umap}
\title{Run UMAP analysis}
\usage{
hd_umap(dat, by_sample = TRUE, components = 2, seed = 123)
}
\arguments{
\item{dat}{An HDAnalyzeR object or a dataset in wide format and sample ID as its first column.}

\item{by_sample}{If TRUE, points represent samples. If FALSE, points represent features. Default is TRUE.}

\item{components}{The number of components to be calculated. Default is 10.}

\item{seed}{The seed to be used in the UMAP analysis. Default is 123.}
}
\value{
A list with the UMAP results.
}
\description{
\code{hd_umap()} runs a UMAP analysis on the provided data. If data contain missing values,
the function imputes them using the k-nearest neighbors algorithm (k = 5). The number of components
to be calculated is defined by the user. The function returns a tibble with the UMAP results.
}
\examples{
# Create the HDAnalyzeR object providing the data and metadata
hd_object <- hd_initialize(example_data, example_metadata)

# Run the UMAP analysis
hd_umap(hd_object, components = 2, by_sample = TRUE, seed = 123)

# Run the UMAP analysis by feature
hd_umap(hd_object, components = 2, by_sample = FALSE, seed = 123)
}
