% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hd_widen_data}
\alias{hd_widen_data}
\title{Convert omics data to wide format}
\usage{
hd_widen_data(dat, exclude = "DAid", names_from = "Assay", values_from = "NPX")
}
\arguments{
\item{dat}{A tibble containing data in long format.}

\item{exclude}{The name of the columns to exclude from the transformation. Default is "DAid".}

\item{names_from}{The name of the column containing the variable names. Default is "Assay".}

\item{values_from}{The name of the column containing the values. Default is "NPX".}
}
\value{
A tibble containing the data in wide format.
}
\description{
\code{hd_widen_data()} transforms omics data from long to wide format with variables
like different "Assays" as column names and expression values like "NPX" as values.
}
\examples{
# Olink data in long format
example_data

# Transform Olink data in wide format
hd_widen_data(example_data)

# Use Sample name instead of Sample ID and Olink IDs instead of Assay names
hd_widen_data(example_data, exclude = "Sample", names_from = "OlinkID")
}
