% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_df}
\alias{import_df}
\title{Import dataframe from file}
\usage{
import_df(file_path)
}
\arguments{
\item{file_path}{The path to the file to import.}
}
\value{
The imported dataframe as a tibble.
}
\description{
\code{import_df()} imports a dataframe from a file. The file format can be CSV,
TSV, TXT, RDA, RDS, XLSX, or Parquet format. It recognizes the file format,
reads it and returns it as a tibble.
}
\examples{
# Save a dataframe as an RDA file
save_df(example_metadata, "metadata", "my_data", file_type = "rda")

# Import the saved RDA file again as a tibble
import_df("my_data/metadata.rda")

unlink("my_data", recursive = TRUE)  # Clean up the created directory
}
