% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_knn}
\alias{impute_knn}
\title{Impute via k-nearest neighbors}
\usage{
impute_knn(
  olink_data,
  wide = T,
  k = 5,
  exclude_cols = c("DAid", "Disease"),
  show_na_percentage = TRUE
)
}
\arguments{
\item{olink_data}{(tibble). The input dataframe.}

\item{wide}{(logical). If T, the data is in wide format.}

\item{k}{(integer). The number of neighbors to consider for imputation.}

\item{exclude_cols}{(string or vector of strings). The columns to exclude from imputation.}

\item{show_na_percentage}{(logical). If T, the percentage of missing values in each column is displayed.}
}
\value{
imputed_data (tibble). The dataframe with missing values imputed.
}
\description{
This function imputes missing values in a dataset using the k-nearest neighbors algorithm.
It allows the user to exclude certain columns from imputation and can also display the
percentage of missing values in each column before imputation.
}
\examples{
test_data <- tibble::tibble(
  A = c(80, 44, NA, 50, 29),
  B = c(30, NA, 85, 70, 54),
  C = c(7, 10, 25, 74, 49)
)
imputed_data <- impute_knn(test_data, k = 3)
}
