% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_mice}
\alias{impute_mice}
\title{Impute via MICE}
\usage{
impute_mice(
  olink_data,
  wide = T,
  m = 5,
  maxit = 5,
  method = "pmm",
  exclude_cols = c("DAid", "Disease"),
  show_na_percentage = T
)
}
\arguments{
\item{olink_data}{(tibble). The input dataframe.}

\item{wide}{(logical). If T, the data is in wide format.}

\item{m}{(integer). The number of imputed datasets to create. Default is 5.}

\item{maxit}{(integer). The maximum number of iterations. Default is 5.}

\item{method}{(string). The imputation method to use. Type \code{methods(mice::mice)} after \code{library(mice)} for all options. Default is "pmm".}

\item{exclude_cols}{(string or vector of strings). The columns to exclude from imputation.}

\item{show_na_percentage}{(logical). If T, the percentage of missing values in each column is displayed.}
}
\value{
imputed_data (tibble). The dataframe with missing values imputed.
}
\description{
This function imputes missing values in a dataset using the MICE algorithm.
It allows the user to exclude certain columns from imputation and can also display the
percentage of missing values in each column before imputation.
}
\examples{
test_data <- example_data |>
  dplyr::select(DAid, Assay, NPX) |>
  tidyr::pivot_wider(names_from = "Assay", values_from = "NPX") |>
  dplyr::slice_head(n = 100)
imputed_data <- impute_mice(test_data)
}
