% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_missForest}
\alias{impute_missForest}
\title{Impute via missForest}
\usage{
impute_missForest(
  olink_data,
  wide = T,
  maxiter = 10,
  ntree = 100,
  parallelize = "variables",
  ncores = 4,
  exclude_cols = c("DAid", "Disease"),
  show_na_percentage = T
)
}
\arguments{
\item{olink_data}{(tibble). The input dataframe}

\item{wide}{(logical). If T, the data is in wide format}

\item{maxiter}{(integer). The maximum number of iterations}

\item{ntree}{(integer). The number of trees to grow}

\item{parallelize}{(string). The type of parallelization to use. Options are "no", "variables", or "forests"}

\item{ncores}{(integer). The number of cores to use for parallelization}

\item{exclude_cols}{(string or vector of strings). The columns to exclude from imputation}

\item{show_na_percentage}{(logical). If T, the percentage of missing values in each column is displayed}
}
\value{
A data frame with imputed values.
}
\description{
This function imputes missing values in a dataset using the \code{missForest} method.
It allows the user to exclude certain columns from imputation and can also display the
percentage of missing values in each column before imputation.
}
\examples{
test_data <- example_data |>
  dplyr::select(DAid, Assay, NPX) |>
  tidyr::pivot_wider(names_from = "Assay", values_from = "NPX") |>
  dplyr::slice_head(n = 100)
imputed_data <- impute_missForest(test_data, maxiter = 1, ntree = 50, parallelize = "no")
}
