% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{make_groups}
\alias{make_groups}
\title{Create class-balanced case-control groups for classification models}
\usage{
make_groups(
  join_data,
  diseases,
  only_female = NULL,
  only_male = NULL,
  seed = 123
)
}
\arguments{
\item{join_data}{Olink data in wide format joined with metadata.}

\item{diseases}{Diseases.}

\item{only_female}{Diseases that are female specific.}

\item{only_male}{Diseases that are male specific.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A list with combined, class-balanced control-case groups for each disease.
}
\description{
\code{make_groups()} creates class-balanced case-control groups for classification models.
It separates the data into control and case groups. It then calculates the amount of data
from each disease in the control group and randomly selects the number of each control class samples.
It also filters the control data of sex specific diseases.
}
\keyword{internal}
