% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize data}
\usage{
normalize_data(
  olink_data,
  metadata = NULL,
  wide = T,
  center = T,
  scale = T,
  batch = NULL,
  batch2 = NULL,
  return_long = F,
  save = F,
  file_name = "normalized_data"
)
}
\arguments{
\item{olink_data}{(tibble). A dataframe containing Olink data to be normalized.}

\item{metadata}{(tibble). A dataframe containing the metadata information.}

\item{wide}{(logical). A logical value indicating whether the data is in wide format. Default is TRUE.}

\item{center}{(logical). A logical value indicating whether to center the data. Default is TRUE.}

\item{scale}{(logical). A logical value indicating whether to scale the data. Default is TRUE.}

\item{batch}{(character). The metadata column containing the batch information. Default is NULL.}

\item{batch2}{(character). The metadata column containing the second batch information. Default is NULL.}

\item{return_long}{(logical). A logical value indicating whether to return the data in long format. Default is FALSE.}

\item{save}{(logical). A logical value indicating whether to save the data. Default is FALSE.}

\item{file_name}{(character). The name of the file to be saved. Default is "normalized_data".}
}
\value{
scaled_data (tibble). A dataframe containing the normalized data.
}
\description{
This function normalizes the data by removing batch effects and scaling the data.
It first converts the data to wide format if it is not already in wide format.
It then removes the batch effects and scales or centers the data.
To remove batch effects, it uses the remove_batch_effects function, that utilizes limma package.
For scaling, it uses the scale function from base R.
}
\examples{
scaled_data <- normalize_data(example_data, example_metadata, wide = FALSE, batch = "Cohort")
}
