% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{plot_de_summary}
\alias{plot_de_summary}
\title{Plot summary visualizations for the differential expression results}
\usage{
plot_de_summary(
  de_results,
  disease_palette = NULL,
  diff_exp_palette = "diff_exp"
)
}
\arguments{
\item{de_results}{A list with the differential expression results.}

\item{disease_palette}{The color palette for the disease. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is NULL.}

\item{diff_exp_palette}{The color palette for the differential expression. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is "diff_exp".}
}
\value{
A list containing the following plots:
\itemize{
\item de_barplot: A barplot with the number of significant proteins for each disease.
\item upset_plot_up: An upset plot with the significant up regulated proteins for each disease.
\item upset_plot_down: An upset plot with the significant down regulated proteins for each disease.
\item proteins_df_up: A tibble with the significant up regulated proteins for each combination of diseases.
\item proteins_df_down: A tibble with the significant down regulated proteins for each combination of diseases.
\item proteins_list_up: A list with the significant up regulated proteins for each combination of diseases.
\item proteins_list_down: A list with the significant down regulated proteins for each combination of diseases.
}
}
\description{
\code{plot_de_summary()} creates summary visualizations for the differential expression results.
It plots a barplot with the number of significant proteins for each disease.
It also creates upset plots both for the significant up and down regulated proteins for each disease.
}
\examples{
# Run differential expression analysis for 3 different cases
de_results_aml <- do_limma(example_data,
                           example_metadata,
                           case = "AML",
                           control = c("BRC", "PRC"),
                           wide = FALSE,
                           only_female = "BRC",
                           only_male = "PRC")

de_results_brc <- do_limma(example_data,
                           example_metadata,
                           case = "BRC",
                           control = c("AML", "PRC"),
                           wide = FALSE,
                           only_female = "BRC",
                           only_male = "PRC")

de_results_prc <- do_limma(example_data,
                           example_metadata,
                           case = "PRC",
                           control = c("AML", "BRC"),
                           wide = FALSE,
                           only_female = "BRC",
                           only_male = "PRC")

# Combine the results
res <- list("AML" = de_results_aml,
            "BRC" = de_results_brc,
            "PRC" = de_results_prc)

# Plot summary visualizations
plot_de_summary(res)
}
