% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{plot_dim_reduction}
\alias{plot_dim_reduction}
\title{Plot sample data points in a two-dimensional plane.}
\usage{
plot_dim_reduction(
  res,
  x,
  y,
  metadata,
  color,
  palette,
  loadings = FALSE,
  variance_explained = NULL,
  loadings_data = NULL
)
}
\arguments{
\item{res}{A tibble with the results of the dimensionality reduction analysis.}

\item{x}{The name of the column in \code{res} that contains the x-axis values.}

\item{y}{The name of the column in \code{res} that contains the y-axis values.}

\item{metadata}{A tibble with metadata information to be added to the plot. It will be used for coloring the points.}

\item{color}{The name of the column in \code{metadata} that contains the variable to be used to plot the points color.}

\item{palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}.}
}
\value{
A ggplot object
}
\description{
\code{plot_dim_reduction()} plots the sample data points in a two-dimensional plane.
The points can be plotted in the PC1/PC2 or UMAP1/UMAP2 space.
}
\keyword{internal}
