% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{plot_features_summary}
\alias{plot_features_summary}
\title{Plot features summary visualizations}
\usage{
plot_features_summary(
  ml_results,
  importance = 50,
  upset_top_features = FALSE,
  disease_palette = NULL,
  feature_type_palette = c(`all-features` = "pink", `top-features` = "darkblue")
)
}
\arguments{
\item{ml_results}{Results from \code{do_elnet()} or \code{do_rf()}.}

\item{importance}{Importance threshold for top features. Default is 50.}

\item{upset_top_features}{Whether to plot the upset plot for the top features. Default is FALSE.}

\item{disease_palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is NULL.}

\item{feature_type_palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is "all-features" = "pink" and "top-features" = "darkblue".}
}
\value{
A list with two elements:
\itemize{
\item features_barplot: Barplot of the number of proteins and top proteins for each disease.
\item upset_plot_features: Upset plot of the top or all proteins.
}
}
\description{
\code{plot_features_summary()} plots the number of proteins and the number of top
proteins for each disease in a barplot. It also plots the upset plot of the
top or all protein features.
}
\examples{
# Create subset of example_data
unique_samples <- unique(example_data$Sample)
filtered_data <- example_data |>
 dplyr::filter(Sample \%in\% unique_samples[1:100])

# Run the elastic net model pipeline
res <- do_elnet(filtered_data,
                example_metadata,
                wide = FALSE,
                cv_sets = 2,
                grid_size = 1,
                ncores = 1)

# Plot features summary
plot_features_summary(res)
}
