% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Create PCA loadings plot}
\usage{
plot_loadings(tidied_res)
}
\arguments{
\item{tidied_res}{(tibble). A tibble with the results of the PCA analysis.}
}
\value{
(plot). A ggplot object.
}
\description{
This function creates a plot of the PCA loadings for the top-8 features and first 4 PCs.
}
\keyword{internal}
