% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Prepare PCA loadings to be plotted on the 2D plane}
\usage{
plot_loadings(dim_object, plot_loadings, nloadings)
}
\arguments{
\item{dim_object}{A PCA object containing the PCA loadings. Created by \code{hd_pca()}.}

\item{plot_loadings}{The component to be plotted. Default is NULL.}

\item{nloadings}{The number of loadings to be plotted. Default is 5.}
}
\value{
A tibble with the PCA loadings to be plotted.
}
\description{
\code{plot_loadings()} prepares the PCA loadings to be plotted on the 2D plane.
}
\keyword{internal}
