% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{plot_metadata_summary}
\alias{plot_metadata_summary}
\title{Plot summary visualization for Sex, Age and BMI metadata}
\usage{
plot_metadata_summary(
  metadata,
  categorical = "Sex",
  numeric = "Age",
  disease_palette = NULL,
  categ_palette = "sex_hpa"
)
}
\arguments{
\item{metadata}{The metadata dataframe.}

\item{disease_palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}.}

\item{categ_palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is "sex_hpa".}
}
\value{
A list containing plots and sample counts.
}
\description{
\code{plot_metadata_summary()} creates three plots:
\itemize{
\item Two ridge plots for the Age and BMI distributions.
\item A bar plot for the number of samples per Sex.
}
}
\keyword{internal}
