% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{plot_metadata_summary}
\alias{plot_metadata_summary}
\title{Create summary plots for age and sex}
\usage{
plot_metadata_summary(
  metadata,
  disease_palette = NULL,
  sex_palette = "sex_hpa"
)
}
\arguments{
\item{metadata}{(tibble). The metadata dataframe.}

\item{disease_palette}{(character). The name of the palette to use for the disease variable. Default is NULL.}

\item{sex_palette}{(character). The name of the palette. Default is "sex_hpa".}
}
\value{
A list containing the following elements:
\itemize{
\item sex_barplot (plot). A bar plot for the number of samples per sex.
\item age_dist_plot (plot). A ridge plot for the age distributions.
\item bmi_dist_plot (plot). A ridge plot for the BMI distributions.
}
}
\description{
The function creates three plots: two ridge plots for the age and BMI distributions and a bar plot for the number of samples per sex.
}
\keyword{internal}
