% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{plot_metadata_summary}
\alias{plot_metadata_summary}
\title{Plot summary visualization for Sex, Age and BMI metadata}
\usage{
plot_metadata_summary(
  metadata,
  sample_id,
  variable,
  palette = NULL,
  unique_threshold = 5
)
}
\arguments{
\item{metadata}{A dataset containing the metadata information with the sample ID as the first column.}

\item{sample_id}{The name of the column containing the sample IDs.}

\item{variable}{The name of the column containing the different classes.}

\item{palette}{A list of color palettes for the plots. The names of the list should match the column names in the metadata. Default is NULL.}

\item{unique_threshold}{The threshold to consider a numeric variable as categorical. Default is 5.}
}
\value{
A list containing plots and sample counts.
}
\description{
\code{plot_metadata_summary()} creates three plots:
\itemize{
\item Two ridge plots for the Age and BMI distributions.
\item A bar plot for the number of samples per Sex.
}
}
\keyword{internal}
