% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_protein_boxplot}
\alias{plot_protein_boxplot}
\title{Plot protein boxplots}
\usage{
plot_protein_boxplot(join_data, proteins, disease, points = T, palette = NULL)
}
\arguments{
\item{join_data}{The dataset with the wide Olink data joined with the metadata.}

\item{proteins}{The proteins to include in the boxplot.}

\item{disease}{The disease to annotate.}

\item{points}{Whether to add points to the boxplot.}

\item{palette}{The color palette to use. Default is "red3" for the annotated disease.}
}
\value{
The boxplot panel with the selected proteins.
}
\description{
\code{plot_protein_boxplot()} plots boxplots for the specified proteins in the dataset.
It annotates the boxplot with color for the selected disease.
It is also possible to add points to the boxplot.
}
\examples{
# Prepare the data
wide_data <- widen_data(example_data)
join_data <- wide_data |>
  dplyr::left_join(example_metadata |> dplyr::select(DAid, Disease, Sex))

# Boxplots for AARSD1 and ABL1 in AML
plot_protein_boxplot(join_data, c("AARSD1", "ABL1"), "AML", palette = "cancers12")
}
