% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{plot_var_imp}
\alias{plot_var_imp}
\title{Plot feature variable importance}
\usage{
plot_var_imp(
  finalfit_res,
  disease,
  accuracy,
  sensitivity,
  specificity,
  auc,
  mixture,
  palette = NULL,
  vline = T,
  subtitle = c("accuracy", "sensitivity", "specificity", "auc", "features",
    "top-features", "mixture"),
  yaxis_names = FALSE
)
}
\arguments{
\item{finalfit_res}{Results from \code{finalfit()}.}

\item{disease}{Disease to predict.}

\item{accuracy}{Accuracy of the model.}

\item{sensitivity}{Sensitivity of the model.}

\item{specificity}{Specificity of the model.}

\item{auc}{AUC of the model.}

\item{mixture}{Mixture of lasso and ridge regularization.}

\item{palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is NULL.}

\item{vline}{Whether to add a vertical line at 50\% importance. Default is TRUE.}

\item{subtitle}{Vector of subtitle elements to include in the plot.}

\item{yaxis_names}{Whether to add y-axis names to the plot. Default is FALSE.}
}
\value{
A list with two elements:
\itemize{
\item features: A tibble with features and their model importance.
\item var_imp_plot: Variable importance plot.
}
}
\description{
\code{plot_var_imp()} collects the features and their model importance.
It scales their importance and plots it against them.
}
\keyword{internal}
