% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{plot_var_imp}
\alias{plot_var_imp}
\title{Plot variable importance}
\usage{
plot_var_imp(
  finalfit_res,
  disease,
  accuracy,
  sensitivity,
  specificity,
  auc,
  mixture,
  palette = NULL,
  vline = T,
  subtitle = c("accuracy", "sensitivity", "specificity", "auc", "features",
    "top-features", "mixture")
)
}
\arguments{
\item{finalfit_res}{(list). Results from elnet_finalfit().}

\item{disease}{(character). Disease to predict.}

\item{accuracy}{(numeric). Accuracy of the model.}

\item{sensitivity}{(numeric). Sensitivity of the model.}

\item{specificity}{(numeric). Specificity of the model.}

\item{auc}{(numeric). AUC of the model.}

\item{mixture}{(numeric). Mixture of lasso and ridge regularization.}

\item{palette}{(character or vector). The color palette for the plot. If it is a character, it should be one of the palettes from get_hpa_palettes(). Default is NULL.}

\item{vline}{(logical). Whether to add a vertical line at 50\% importance. Default is TRUE.}

\item{subtitle}{(vector). Vector of subtitles to include in the plot. Default is a list with all.}
}
\value{
A list with two elements:
\itemize{
\item features (tibble). Features with importance values.
\item var_imp_plot (plot). Variable importance plot.
}
}
\description{
This function collects the features and their importance.
It scales their importance and plots it against them.
}
\keyword{internal}
