% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Create volcano plots}
\usage{
plot_volcano(
  de_result,
  pval_lim = 0.05,
  logfc_lim = 0,
  top_up_prot = 40,
  top_down_prot = 10,
  palette = "diff_exp",
  title = NULL,
  report_nproteins = TRUE,
  user_defined_proteins = NULL,
  subtitle = NULL
)
}
\arguments{
\item{de_result}{The differential expression results.}

\item{pval_lim}{The p-value limit for significance. Default is 0.05.}

\item{logfc_lim}{The logFC limit for significance. Default is 0.}

\item{top_up_prot}{The number of top up regulated proteins to label on the plot. Default is 40.}

\item{top_down_prot}{The number of top down regulated proteins to label on the plot. Default is 10.}

\item{palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is "diff_exp".}

\item{title}{The title of the plot or NULL for no title.}

\item{report_nproteins}{If the number of significant proteins should be reported in the subtitle. Default is TRUE.}

\item{subtitle}{The subtitle of the plot or NULL for no subtitle.}
}
\value{
A ggplot object with the volcano plot.
}
\description{
\code{plot_volcano()} creates volcano plots for the differential expression results.
It colors and labels the top up and down regulated proteins.
}
\keyword{internal}
