% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{prepare_plot_data}
\alias{prepare_plot_data}
\title{Extract information from the HDAnalyzeR and PCA or UMAP object}
\usage{
prepare_plot_data(dim_object, metadata, color, x, y)
}
\arguments{
\item{dim_object}{A PCA or UMAP object containing the results of the dimensionality reduction analysis. Created by \code{hd_pca()} or \code{hd_umap()}.}

\item{metadata}{An HDAnalyzeR object or a dataset in wide format and sample ID as its first column.}

\item{color}{The name of the column that contains the variable to be used to plot the points color.}

\item{x}{The name of the column in \code{dim_object} that contains the x-axis values.}

\item{y}{The name of the column in \code{dim_object} that contains the y-axis values.}
}
\value{
A tibble with the PCA or UMAP results and metadata information if available.
}
\description{
\code{prepare_plot_data()} extracts the PCA or UMAP results and metadata information from the respective objects.
}
\keyword{internal}
