% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{print_summary}
\alias{print_summary}
\title{Print the summary of the quality control results}
\usage{
print_summary(
  sample_n,
  var_n,
  class_summary,
  na_percentage_col,
  na_percentage_row,
  normality_results = FALSE,
  cor_results = FALSE,
  heatmap = FALSE,
  threshold = FALSE
)
}
\arguments{
\item{sample_n}{The number of samples.}

\item{var_n}{The number of variables.}

\item{class_summary}{A table with the counts of each class in the dataframe.}

\item{na_percentage_col}{A tibble with the column names and the percentage of NAs in each column.}

\item{na_percentage_row}{A tibble with the DAids and the percentage of NAs in each row.}

\item{normality_results}{A tibble with the protein names, p-values, adjusted p-values, and normality status.}

\item{cor_results}{A tibble with the filtered protein pairs and their correlation values.}

\item{heatmap}{A heatmap of protein-protein correlations.}

\item{threshold}{The reporting protein-protein correlation threshold.}
}
\description{
\code{print_summary()} prints the summary of the quality control results of the
input dataset. It includes the number of samples and variables, the counts of
each class, the percentage of NAs in each column and row, the normality test
results, the protein-protein correlations above a certain threshold, and the
correlation heatmap.
}
\keyword{internal}
