% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{print_summary}
\alias{print_summary}
\title{Print the summary of the quality control results}
\usage{
print_summary(
  sample_n,
  var_n,
  class_summary,
  na_percentage_col,
  na_percentage_row,
  normality_results = F,
  cor_results = F,
  heatmap = F,
  threshold = F
)
}
\arguments{
\item{sample_n}{(numeric). The number of samples.}

\item{var_n}{(numeric). The number of variables.}

\item{class_summary}{(table). A table with the counts of each class in the dataframe.}

\item{na_percentage_col}{(tibble). A tibble with the column names and the percentage of NAs in each column.}

\item{na_percentage_row}{(tibble). A tibble with the DAids and the percentage of NAs in each row.}

\item{normality_results}{(tibble). A tibble with the protein names, p-values, adjusted p-values, and normality status.}

\item{cor_results}{(tibble). A tibble with the filtered protein pairs and their correlation values.}

\item{heatmap}{(plot). A heatmap of protein-protein correlations.}

\item{threshold}{(numeric). The reporting protein-protein correlation threshold.}
}
\value{
NULL.
}
\description{
The function prints the summary of the quality control results of the input dataframe.
}
\keyword{internal}
