% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{qc_summary_data}
\alias{qc_summary_data}
\title{Summarize the quality control results of Olink data}
\usage{
qc_summary_data(df, wide = T, threshold = 0.8, report = T)
}
\arguments{
\item{df}{The input dataset.}

\item{wide}{Whether the input dataset is in wide format. Default is TRUE.}

\item{threshold}{The reporting protein-protein correlation threshold. Default is 0.8.}

\item{report}{Whether to print the summary. Default is TRUE.}
}
\value{
A list containing the following elements:
\itemize{
\item na_percentage_col: A tibble with the column names and the percentage of NAs in each column.
\item na_percentage_row: A tibble with the DAids and the percentage of NAs in each row.
\item normality_results: A tibble with the protein names, p-values, adjusted p-values, and normality status.
\item cor_matrix: A matrix of protein-protein correlations.
\item cor_results: A tibble with the filtered protein pairs and their correlation values.
\item heatmap: A heatmap of protein-protein correlations.
}
}
\description{
\code{qc_summary_data()} summarizes the quality control results of the input dataset.
It can handles both long and wide dataframes. The function checks the column types,
calculates the percentage of NAs in each column and row, performs a normality test,
calculates the protein-protein correlations, and creates a heatmap of the correlations.
The user can specify the reporting protein-protein correlation threshold.
}
\examples{
qc_res <- qc_summary_data(example_data, wide = FALSE, threshold = 0.7)
}
