% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{qc_summary_metadata}
\alias{qc_summary_metadata}
\title{Summarize the quality control results of metadata}
\usage{
qc_summary_metadata(
  metadata,
  categorical = "Sex",
  numeric = "Age",
  disease_palette = NULL,
  categ_palette = "sex_hpa",
  report = TRUE
)
}
\arguments{
\item{metadata}{The metadata dataframe.}

\item{categorical}{The categorical variables to summarize. Default is "Sex".}

\item{numeric}{The numeric variables to summarize. Default is "Age".}

\item{disease_palette}{The color palette for the different diseases. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}.}

\item{categ_palette}{The categorical color palette. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is "sex_hpa".}

\item{report}{Whether to print the summary. Default is TRUE.}
}
\value{
A list containing the following elements:
\itemize{
\item na_percentage_col: A tibble with the column names and the percentage of NAs in each column.
\item na_percentage_row: A tibble with the DAids and the percentage of NAs in each row.
\item Several distribution and barplots, as well as the counts of samples.
}
}
\description{
\code{qc_summary_metadata()} summarizes the quality control results of the metadata dataframe.
It checks the column types, calculates the percentage of NAs in each column and row,
and creates summary visualizations for user selected categorical and numeric variables.
}
\examples{
qc_res <- qc_summary_metadata(example_metadata, disease_palette = "cancers12")

# Metadata distributions
qc_res$sex_barplot
qc_res$age_dist_plot
qc_res$bmi_dist_plot
}
