% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{qc_summary_metadata}
\alias{qc_summary_metadata}
\title{Summarize the quality control results of metadata}
\usage{
qc_summary_metadata(
  metadata,
  disease_palette = NULL,
  sex_palette = "sex_hpa",
  report = T
)
}
\arguments{
\item{metadata}{The metadata dataframe.}

\item{disease_palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}.}

\item{sex_palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is "sex_hpa".}

\item{report}{Whether to print the summary. Default is TRUE.}
}
\value{
A list containing the following elements:
\itemize{
\item na_percentage_col: A tibble with the column names and the percentage of NAs in each column.
\item na_percentage_row: A tibble with the DAids and the percentage of NAs in each row.
\item sex_barplot: A bar plot for the number of samples per sex.
\item age_dist_plot: A ridge plot for the age distributions.
\item bmi_dist_plot: A ridge plot for the BMI distributions.
}
}
\description{
\code{qc_summary_metadata()} summarizes the quality control results of the metadata dataframe.
It checks the column types, calculates the percentage of NAs in each column and row,
and creates summary visualizations for Sex, Age and BMI.
}
\examples{
qc_res <- qc_summary_metadata(example_metadata, disease_palette = "cancers12")

# Metadata distributions
qc_res$sex_barplot
qc_res$age_dist_plot
qc_res$bmi_dist_plot
}
