% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summary.R
\name{qc_summary_metadata}
\alias{qc_summary_metadata}
\title{Summarize the quality control results of metadata}
\usage{
qc_summary_metadata(
  metadata,
  sample_id,
  variable,
  palette = NULL,
  unique_threshold = 5,
  verbose = TRUE
)
}
\arguments{
\item{metadata}{A dataset containing the metadata information with the sample ID as the first column.}

\item{sample_id}{The name of the column containing the sample IDs.}

\item{variable}{The name of the column containing the different classes.}

\item{palette}{A list of color palettes for the plots. The names of the list should match the column names in the metadata. Default is NULL.}

\item{unique_threshold}{The threshold to consider a numeric variable as categorical. Default is 5.}

\item{verbose}{Whether to print the summary. Default is TRUE.}
}
\value{
A list of the qc summary of data
}
\description{
\code{qc_summary_metadata()} summarizes the quality control results of the metadata dataframe.
It checks the column types, calculates the percentage of NAs in each column and row,
and creates summary visualizations for categorical and numeric variables.
}
\keyword{internal}
