% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{remove_batch_effects}
\alias{remove_batch_effects}
\title{Remove batch effects}
\usage{
remove_batch_effects(wide_data, metadata, batch, batch2 = NULL)
}
\arguments{
\item{wide_data}{(tibble). A dataframe containing the data to be normalized. The data should be in wide format.}

\item{metadata}{(tibble). A dataframe containing the metadata information.}

\item{batch}{(character). The metadata column containing the batch information.}

\item{batch2}{(character). The metadata column containing the second batch information. Default is NULL.}
}
\value{
no_batch_effects (tibble). A dataframe containing the data without batch effects.
}
\description{
This function removes batch effects from the data using the limma package.
It converts the dataframe into matrix and transposes it to get it ready for limma.
It removes the batch effects and then converts the data back to normal.
}
\keyword{internal}
