% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{replace_with_na}
\alias{replace_with_na}
\title{Replace values with NA}
\usage{
replace_with_na(df_in)
}
\arguments{
\item{df_in}{(tibble). The input dataframe}
}
\value{
df_out (tibble). The dataframe with specified values replaced with NA
}
\description{
The function replaces the specified values with NA in the input vector.
}
\examples{
df <- replace_with_na(example_metadata)
}
