% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_df}
\alias{save_df}
\title{Save dataframe}
\usage{
save_df(df, dir_name, file_name, file_type = c("csv", "tsv", "rda"))
}
\arguments{
\item{df}{(tibble). The dataframe to save}

\item{dir_name}{(string). The directory where the file will be saved}

\item{file_name}{(string). The name of the file to save}

\item{file_type}{(string). The type of file to save the dataframe as. Options are "csv", "tsv", or "rda"}
}
\description{
The function saves a dataframe in the specified format (CSV, TSV, or RDA) in the specified directory.
}
\examples{
df <- data.frame(x = 1:10, y = rnorm(10))
save_df(df, "my_data", "sample_data", "csv")
# Clean up the created directory
unlink("my_data", recursive = TRUE)
}
