% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_fill_hpa}
\alias{scale_fill_hpa}
\title{HPA fill scales}
\usage{
scale_fill_hpa(palette = "palette1")
}
\arguments{
\item{palette}{(character). The name of the palette to use. Available palettes are: "sex", "diff_exp", "cancers12", "cancers15".}
}
\value{
A ggplot2 scale for fill aesthetics.
}
\description{
HPA fill scales
}
\keyword{internal}
