% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_fill_hpa}
\alias{scale_fill_hpa}
\title{HPA fill scales}
\usage{
scale_fill_hpa(palette)
}
\arguments{
\item{palette}{The name of the palette to use. It should be one of the palettes from \code{get_hpa_palettes()}.}
}
\value{
A ggplot2 scale for fill aesthetics.
}
\description{
\code{scale_fill_hpa()} creates a ggplot2 scale for fill aesthetics using the color
palettes from the Human Protein Atlas (HPA) project.
}
\examples{
# Create an example dataframe
data <- data.frame(
  Sex = c("M", "F"),
  Count = c(60, 40)
)

# Create a plot
plot <- ggplot2::ggplot(data, ggplot2::aes(x = Sex, y = Count, fill = Sex)) +
  ggplot2::geom_bar(stat = "identity", position = "dodge")
plot

# Add a custom palette
plot + scale_fill_hpa("sex_hpa")
}
