% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{split_data}
\alias{split_data}
\title{Split dataset into training and test sets}
\usage{
split_data(
  join_data,
  variable = "Disease",
  strata = TRUE,
  ratio = 0.75,
  seed = 123
)
}
\arguments{
\item{join_data}{Olink data in wide format joined with metadata.}

\item{variable}{Variable to stratify the data. Default is "Disease".}

\item{strata}{Whether to stratify the data. Default is TRUE.}

\item{ratio}{Ratio of training data to test data. Default is 0.75.}

\item{seed}{Seed for reproducibility. Default is 123.}
}
\value{
A list with three elements:
\itemize{
\item train_set: The training set.
\item test_set: The test set.
\item data_split: The data split object.
}
}
\description{
\code{split_data()} splits the dataset into training and test sets based on user defined ratio.
It also stratifies the data based on the variable of interest.
}
\keyword{internal}
