% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{split_data}
\alias{split_data}
\title{Split data into training and test sets}
\usage{
split_data(join_data, ratio = 0.75, seed = 123)
}
\arguments{
\item{join_data}{(tibble). Wide data joined with metadata.}

\item{ratio}{(numeric). Ratio of training data to test data. Default is 0.75.}

\item{seed}{(numeric). Seed for reproducibility. Default is 123.}
}
\value{
A list with two elements:
\itemize{
\item train_set (tibble). The training set.
\item test_set (tibble). The test set.
}
}
\description{
This function splits the data into training and test sets based on user defined ratio.
}
\examples{
wide_data <- example_data |>
  dplyr::select(DAid, Assay, NPX) |>
  tidyr::pivot_wider(names_from = Assay, values_from = NPX)
join_data <- wide_data |>
  dplyr::left_join(example_metadata |> dplyr::select(DAid, Disease, Sex))

data_splits <- split_data(join_data)
}
