% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{testfit}
\alias{testfit}
\title{Test the best model}
\usage{
testfit(
  train_set,
  test_set,
  variable = "Disease",
  case,
  finalfit_res,
  exclude_cols = NULL,
  type = "lasso",
  seed = 123,
  palette = NULL
)
}
\arguments{
\item{train_set}{Training set.}

\item{test_set}{Testing set.}

\item{variable}{The variable to predict. Default is "Disease".}

\item{case}{Case to predict.}

\item{finalfit_res}{Results from \code{elnet_finalfit()}.}

\item{exclude_cols}{Columns to exclude from the data before the model is tuned. Default is NULL.}

\item{type}{Type of regularization. Default is "lasso". Other options are "ridge" and "elnet".}

\item{seed}{Seed for reproducibility. Default is 123.}

\item{palette}{The color palette for the plot. If it is a character, it should be one of the palettes from \code{get_hpa_palettes()}. Default is NULL.}
}
\value{
A list with two elements:
\itemize{
\item metrics: A list with 5 metrics:
\item accuracy: Accuracy of the model.
\item sensitivity: Sensitivity of the model.
\item specificity: Specificity of the model.
\item auc: AUC of the model.
\item conf_matrix: Confusion matrix of the model.
\item roc_curve: ROC curve of the model.
\item mixture: Mixture of lasso and ridge regularization.
}
}
\description{
\code{testfit()} tests the best model on the test set and calculate metrics.
It calculates the accuracy, sensitivity, specificity, AUC, confusion matrix,
and ROC curve.
}
\details{
In random forest models, mixture is returned as NULL.
}
\keyword{internal}
