% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{vis_hypopt}
\alias{vis_hypopt}
\title{Visualize hyperparameter optimization results}
\usage{
vis_hypopt(tune_res, x, color, case)
}
\arguments{
\item{tune_res}{Hyperparameter optimization results.}

\item{x}{X-axis variable of the plot.}

\item{color}{Color variable of the plot.}

\item{case}{Case to predict.}
}
\value{
Hyperparameter optimization plot.
}
\description{
\code{vis_hypopt()} plots the hyperparameter optimization results.
}
\keyword{internal}
