% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_models.R
\name{vis_hypopt}
\alias{vis_hypopt}
\title{Visualize hyperparameter optimization results}
\usage{
vis_hypopt(tune_res, x, color, disease)
}
\arguments{
\item{tune_res}{(tibble). Hyperparameter optimization results.}

\item{x}{(character). X-axis variable of the plot.}

\item{color}{(character). Color variable of the plot. Default is NULL.}

\item{disease}{(character). Disease to predict.}
}
\value{
hypopt_plot (plot). Hyperparameter optimization plot.
}
\description{
This function visualizes the hyperparameter optimization results.
}
\examples{
wide_data <- example_data |>
  dplyr::select(DAid, Assay, NPX) |>
  tidyr::pivot_wider(names_from = Assay, values_from = NPX)
join_data <- wide_data |>
  dplyr::left_join(example_metadata |> dplyr::select(DAid, Disease, Sex))
diseases <- unique(example_metadata$Disease)
data_split <- split_data(join_data)
train_list <- make_groups(data_split$train_set,
                          diseases,
                          only_female = c("BRC", "CVX", "ENDC", "OVC"),
                          only_male = "PRC")

test_list <- make_groups(data_split$test_set,
                         diseases,
                         only_female = c("BRC", "CVX", "ENDC", "OVC"),
                         only_male = "PRC")

hypopt_res <- elnet_hypopt(train_list,
                           test_list,
                           "AML",
                           type = "elnet",
                           cv_sets = 2,
                           grid_size = 1,
                           ncores = 1,
                           exclude_cols = "Sex")

vis_hypopt(hypopt_res$elnet_tune, "penalty", "mixture", "AML")
}
