% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{widen_data}
\alias{widen_data}
\title{Widen data}
\usage{
widen_data(olink_data, wide)
}
\arguments{
\item{olink_data}{(tibble). A dataframe containing Olink data to be normalized}

\item{wide}{(logical). A logical value indicating whether the data is in wide format}
}
\value{
wide_data (tibble). A dataframe containing the data in wide format
}
\description{
The function widens the data from long to wide format.
}
\examples{
wide_data <- widen_data(example_data, wide = FALSE)
}
